Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 1  41 70  6  1
#> 2  35 30  8  1
#> 3   2 43 43  1
#> 4  13 19 53  1
#> 5   3 37 49  1
#> 6   4 59 36  1
#> 7  13 18 57  1
#> 8  43 13 19  1
#> 9  51 62  2  1
#> 10 53  5 71  1
#> 12 58 70  6  0
#> 13 41 75  6  0
#> 14 58 75  6  0
#> 15 41 70 76  0
#> 16 58 70 76  0
#> 17 41 75 76  0
#> 18 58 75 76  0
#> 20 58 30  8  0
#> 21 35 75  8  0
#> 22 58 75  8  0
#> 23 35 30 76  0
#> 24 58 30 76  0
#> 25 35 75 76  0
#> 28 58 43 43  0
#> 29  2 75 43  0
#> 30 58 75 43  0
#> 31  2 43 76  0
#> 32 58 43 76  0
#> 33  2 75 76  0
#> 36 58 19 53  0
#> 37 13 75 53  0
#> 38 58 75 53  0
#> 39 13 19 76  0
#> 40 58 19 76  0
#> 41 13 75 76  0
#> 44 58 37 49  0
#> 45  3 75 49  0
#> 46 58 75 49  0
#> 47  3 37 76  0
#> 48 58 37 76  0
#> 49  3 75 76  0
#> 52 58 59 36  0
#> 53  4 75 36  0
#> 54 58 75 36  0
#> 55  4 59 76  0
#> 56 58 59 76  0
#> 57  4 75 76  0
#> 60 58 18 57  0
#> 61 13 75 57  0
#> 62 58 75 57  0
#> 63 13 18 76  0
#> 64 58 18 76  0
#> 68 58 13 19  0
#> 69 43 75 19  0
#> 70 58 75 19  0
#> 71 43 13 76  0
#> 72 58 13 76  0
#> 73 43 75 76  0
#> 76 58 62  2  0
#> 77 51 75  2  0
#> 78 58 75  2  0
#> 79 51 62 76  0
#> 80 58 62 76  0
#> 81 51 75 76  0
#> 84 58  5 71  0
#> 85 53 75 71  0
#> 86 58 75 71  0
#> 87 53  5 76  0
#> 88 58  5 76  0
#> 89 53 75 76  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 1  41 70  6  1
#> 2  35 30  8  1
#> 3   2 43 43  1
#> 4  13 19 53  1
#> 5   3 37 49  1
#> 6   4 59 36  1
#> 7  13 18 57  1
#> 8  43 13 19  1
#> 9  51 62  2  1
#> 10 53  5 71  1
#> 12 58 70  6  0
#> 13 41  0  6  0
#> 14 58  0  6  0
#> 15 41 70 76  0
#> 16 58 70 76  0
#> 17 41  0 76  0
#> 18 58  0 76  0
#> 20 58 30  8  0
#> 21 35  0  8  0
#> 22 58  0  8  0
#> 23 35 30 76  0
#> 24 58 30 76  0
#> 25 35  0 76  0
#> 28 58 43 43  0
#> 29  2  0 43  0
#> 30 58  0 43  0
#> 31  2 43 76  0
#> 32 58 43 76  0
#> 33  2  0 76  0
#> 36 58 19 53  0
#> 37 13  0 53  0
#> 38 58  0 53  0
#> 39 13 19 76  0
#> 40 58 19 76  0
#> 41 13  0 76  0
#> 44 58 37 49  0
#> 45  3  0 49  0
#> 46 58  0 49  0
#> 47  3 37 76  0
#> 48 58 37 76  0
#> 49  3  0 76  0
#> 52 58 59 36  0
#> 53  4  0 36  0
#> 54 58  0 36  0
#> 55  4 59 76  0
#> 56 58 59 76  0
#> 57  4  0 76  0
#> 60 58 18 57  0
#> 61 13  0 57  0
#> 62 58  0 57  0
#> 63 13 18 76  0
#> 64 58 18 76  0
#> 68 58 13 19  0
#> 69 43  0 19  0
#> 70 58  0 19  0
#> 71 43 13 76  0
#> 72 58 13 76  0
#> 73 43  0 76  0
#> 76 58 62  2  0
#> 77 51  0  2  0
#> 78 58  0  2  0
#> 79 51 62 76  0
#> 80 58 62 76  0
#> 81 51  0 76  0
#> 84 58  5 71  0
#> 85 53  0 71  0
#> 86 58  0 71  0
#> 87 53  5 76  0
#> 88 58  5 76  0
#> 89 53  0 76  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 1  41 70   6  1
#> 2  35 30   8  1
#> 3   2 43  43  1
#> 4  13 19  53  1
#> 5   3 37  49  1
#> 6   4 59  36  1
#> 7  13 18  57  1
#> 8  43 13  19  1
#> 9  51 62   2  1
#> 10 53  5  71  1
#> 12  0 70   6  0
#> 13 41  0   6  0
#> 14  0  0   6  0
#> 15 41 70 100  0
#> 16  0 70 100  0
#> 17 41  0 100  0
#> 18  0  0 100  0
#> 20  0 30   8  0
#> 21 35  0   8  0
#> 22  0  0   8  0
#> 23 35 30 100  0
#> 24  0 30 100  0
#> 25 35  0 100  0
#> 28  0 43  43  0
#> 29  2  0  43  0
#> 30  0  0  43  0
#> 31  2 43 100  0
#> 32  0 43 100  0
#> 33  2  0 100  0
#> 36  0 19  53  0
#> 37 13  0  53  0
#> 38  0  0  53  0
#> 39 13 19 100  0
#> 40  0 19 100  0
#> 41 13  0 100  0
#> 44  0 37  49  0
#> 45  3  0  49  0
#> 46  0  0  49  0
#> 47  3 37 100  0
#> 48  0 37 100  0
#> 49  3  0 100  0
#> 52  0 59  36  0
#> 53  4  0  36  0
#> 54  0  0  36  0
#> 55  4 59 100  0
#> 56  0 59 100  0
#> 57  4  0 100  0
#> 60  0 18  57  0
#> 61 13  0  57  0
#> 62  0  0  57  0
#> 63 13 18 100  0
#> 64  0 18 100  0
#> 68  0 13  19  0
#> 69 43  0  19  0
#> 70  0  0  19  0
#> 71 43 13 100  0
#> 72  0 13 100  0
#> 73 43  0 100  0
#> 76  0 62   2  0
#> 77 51  0   2  0
#> 78  0  0   2  0
#> 79 51 62 100  0
#> 80  0 62 100  0
#> 81 51  0 100  0
#> 84  0  5  71  0
#> 85 53  0  71  0
#> 86  0  0  71  0
#> 87 53  5 100  0
#> 88  0  5 100  0
#> 89 53  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     z1 z2 z3 z4 z5 pt
#> 1   19 45 24 32 31  1
#> 2   69 38  8 46 62  1
#> 3   81 35 21 35 63  1
#> 4   23 69 14 56 53  1
#> 5   14 66 62 24 43  1
#> 6   25 41 26 40 82  1
#> 7   44 50 23 83 26  1
#> 8    7 46 63 32 58  1
#> 9   38 42  6 38 40  1
#> 10  34  9 45 58 69  1
#> 11  57 79 44 34 15  1
#> 12  48 26 58 13 71  1
#> 13  29 34 46 13 67  1
#> 14  53  6 59 69 50  1
#> 15  30 68 58 11 60  1
#> 16   7 56 35 42 66  1
#> 17  69 45 47  5 54  1
#> 18  23 60 32 20 69  1
#> 19  34 25 30 76 72  1
#> 20  35 19 37 77 68  1
#> 22  86 45 24 32 31  0
#> 23  19 84 24 32 31  0
#> 24  86 84 24 32 31  0
#> 25  19 45 68 32 31  0
#> 26  86 45 68 32 31  0
#> 27  19 84 68 32 31  0
#> 28  86 84 68 32 31  0
#> 29  19 45 24 88 31  0
#> 30  86 45 24 88 31  0
#> 31  19 84 24 88 31  0
#> 32  86 84 24 88 31  0
#> 33  19 45 68 88 31  0
#> 34  86 45 68 88 31  0
#> 35  19 84 68 88 31  0
#> 36  86 84 68 88 31  0
#> 37  19 45 24 32 87  0
#> 38  86 45 24 32 87  0
#> 39  19 84 24 32 87  0
#> 40  86 84 24 32 87  0
#> 41  19 45 68 32 87  0
#> 42  86 45 68 32 87  0
#> 43  19 84 68 32 87  0
#> 44  86 84 68 32 87  0
#> 45  19 45 24 88 87  0
#> 46  86 45 24 88 87  0
#> 47  19 84 24 88 87  0
#> 48  86 84 24 88 87  0
#> 49  19 45 68 88 87  0
#> 50  86 45 68 88 87  0
#> 51  19 84 68 88 87  0
#> 52  86 84 68 88 87  0
#> 54  86 38  8 46 62  0
#> 55  69 84  8 46 62  0
#> 56  86 84  8 46 62  0
#> 57  69 38 68 46 62  0
#> 58  86 38 68 46 62  0
#> 59  69 84 68 46 62  0
#> 60  86 84 68 46 62  0
#> 61  69 38  8 88 62  0
#> 62  86 38  8 88 62  0
#> 63  69 84  8 88 62  0
#> 64  86 84  8 88 62  0
#> 65  69 38 68 88 62  0
#> 66  86 38 68 88 62  0
#> 67  69 84 68 88 62  0
#> 68  86 84 68 88 62  0
#> 69  69 38  8 46 87  0
#> 70  86 38  8 46 87  0
#> 71  69 84  8 46 87  0
#> 72  86 84  8 46 87  0
#> 73  69 38 68 46 87  0
#> 74  86 38 68 46 87  0
#> 75  69 84 68 46 87  0
#> 76  86 84 68 46 87  0
#> 77  69 38  8 88 87  0
#> 78  86 38  8 88 87  0
#> 79  69 84  8 88 87  0
#> 80  86 84  8 88 87  0
#> 81  69 38 68 88 87  0
#> 82  86 38 68 88 87  0
#> 83  69 84 68 88 87  0
#> 86  86 35 21 35 63  0
#> 87  81 84 21 35 63  0
#> 88  86 84 21 35 63  0
#> 89  81 35 68 35 63  0
#> 90  86 35 68 35 63  0
#> 91  81 84 68 35 63  0
#> 92  86 84 68 35 63  0
#> 93  81 35 21 88 63  0
#> 94  86 35 21 88 63  0
#> 95  81 84 21 88 63  0
#> 96  86 84 21 88 63  0
#> 97  81 35 68 88 63  0
#> 98  86 35 68 88 63  0
#> 99  81 84 68 88 63  0
#> 100 86 84 68 88 63  0
#> 101 81 35 21 35 87  0
#> 102 86 35 21 35 87  0
#> 103 81 84 21 35 87  0
#> 104 86 84 21 35 87  0
#> 105 81 35 68 35 87  0
#> 106 86 35 68 35 87  0
#> 107 81 84 68 35 87  0
#> 108 86 84 68 35 87  0
#> 109 81 35 21 88 87  0
#> 110 86 35 21 88 87  0
#> 111 81 84 21 88 87  0
#> 112 86 84 21 88 87  0
#> 113 81 35 68 88 87  0
#> 114 86 35 68 88 87  0
#> 115 81 84 68 88 87  0
#> 118 86 69 14 56 53  0
#> 119 23 84 14 56 53  0
#> 120 86 84 14 56 53  0
#> 121 23 69 68 56 53  0
#> 122 86 69 68 56 53  0
#> 123 23 84 68 56 53  0
#> 124 86 84 68 56 53  0
#> 125 23 69 14 88 53  0
#> 126 86 69 14 88 53  0
#> 127 23 84 14 88 53  0
#> 128 86 84 14 88 53  0
#> 129 23 69 68 88 53  0
#> 130 86 69 68 88 53  0
#> 131 23 84 68 88 53  0
#> 132 86 84 68 88 53  0
#> 133 23 69 14 56 87  0
#> 134 86 69 14 56 87  0
#> 135 23 84 14 56 87  0
#> 136 86 84 14 56 87  0
#> 137 23 69 68 56 87  0
#> 138 86 69 68 56 87  0
#> 139 23 84 68 56 87  0
#> 140 86 84 68 56 87  0
#> 141 23 69 14 88 87  0
#> 142 86 69 14 88 87  0
#> 143 23 84 14 88 87  0
#> 144 86 84 14 88 87  0
#> 145 23 69 68 88 87  0
#> 146 86 69 68 88 87  0
#> 147 23 84 68 88 87  0
#> 150 86 66 62 24 43  0
#> 151 14 84 62 24 43  0
#> 152 86 84 62 24 43  0
#> 153 14 66 68 24 43  0
#> 154 86 66 68 24 43  0
#> 155 14 84 68 24 43  0
#> 156 86 84 68 24 43  0
#> 157 14 66 62 88 43  0
#> 158 86 66 62 88 43  0
#> 159 14 84 62 88 43  0
#> 160 86 84 62 88 43  0
#> 161 14 66 68 88 43  0
#> 162 86 66 68 88 43  0
#> 163 14 84 68 88 43  0
#> 164 86 84 68 88 43  0
#> 165 14 66 62 24 87  0
#> 166 86 66 62 24 87  0
#> 167 14 84 62 24 87  0
#> 168 86 84 62 24 87  0
#> 169 14 66 68 24 87  0
#> 170 86 66 68 24 87  0
#> 171 14 84 68 24 87  0
#> 172 86 84 68 24 87  0
#> 173 14 66 62 88 87  0
#> 174 86 66 62 88 87  0
#> 175 14 84 62 88 87  0
#> 176 86 84 62 88 87  0
#> 177 14 66 68 88 87  0
#> 178 86 66 68 88 87  0
#> 179 14 84 68 88 87  0
#> 182 86 41 26 40 82  0
#> 183 25 84 26 40 82  0
#> 184 86 84 26 40 82  0
#> 185 25 41 68 40 82  0
#> 186 86 41 68 40 82  0
#> 187 25 84 68 40 82  0
#> 188 86 84 68 40 82  0
#> 189 25 41 26 88 82  0
#> 190 86 41 26 88 82  0
#> 191 25 84 26 88 82  0
#> 192 86 84 26 88 82  0
#> 193 25 41 68 88 82  0
#> 194 86 41 68 88 82  0
#> 195 25 84 68 88 82  0
#> 196 86 84 68 88 82  0
#> 197 25 41 26 40 87  0
#> 198 86 41 26 40 87  0
#> 199 25 84 26 40 87  0
#> 200 86 84 26 40 87  0
#> 201 25 41 68 40 87  0
#> 202 86 41 68 40 87  0
#> 203 25 84 68 40 87  0
#> 204 86 84 68 40 87  0
#> 205 25 41 26 88 87  0
#> 206 86 41 26 88 87  0
#> 207 25 84 26 88 87  0
#> 208 86 84 26 88 87  0
#> 209 25 41 68 88 87  0
#> 210 86 41 68 88 87  0
#> 211 25 84 68 88 87  0
#> 214 86 50 23 83 26  0
#> 215 44 84 23 83 26  0
#> 216 86 84 23 83 26  0
#> 217 44 50 68 83 26  0
#> 218 86 50 68 83 26  0
#> 219 44 84 68 83 26  0
#> 220 86 84 68 83 26  0
#> 221 44 50 23 88 26  0
#> 222 86 50 23 88 26  0
#> 223 44 84 23 88 26  0
#> 224 86 84 23 88 26  0
#> 225 44 50 68 88 26  0
#> 226 86 50 68 88 26  0
#> 227 44 84 68 88 26  0
#> 228 86 84 68 88 26  0
#> 229 44 50 23 83 87  0
#> 230 86 50 23 83 87  0
#> 231 44 84 23 83 87  0
#> 232 86 84 23 83 87  0
#> 233 44 50 68 83 87  0
#> 234 86 50 68 83 87  0
#> 235 44 84 68 83 87  0
#> 236 86 84 68 83 87  0
#> 237 44 50 23 88 87  0
#> 238 86 50 23 88 87  0
#> 239 44 84 23 88 87  0
#> 240 86 84 23 88 87  0
#> 241 44 50 68 88 87  0
#> 242 86 50 68 88 87  0
#> 243 44 84 68 88 87  0
#> 246 86 46 63 32 58  0
#> 247  7 84 63 32 58  0
#> 248 86 84 63 32 58  0
#> 249  7 46 68 32 58  0
#> 250 86 46 68 32 58  0
#> 251  7 84 68 32 58  0
#> 252 86 84 68 32 58  0
#> 253  7 46 63 88 58  0
#> 254 86 46 63 88 58  0
#> 255  7 84 63 88 58  0
#> 256 86 84 63 88 58  0
#> 257  7 46 68 88 58  0
#> 258 86 46 68 88 58  0
#> 259  7 84 68 88 58  0
#> 260 86 84 68 88 58  0
#> 261  7 46 63 32 87  0
#> 262 86 46 63 32 87  0
#> 263  7 84 63 32 87  0
#> 264 86 84 63 32 87  0
#> 265  7 46 68 32 87  0
#> 266 86 46 68 32 87  0
#> 267  7 84 68 32 87  0
#> 269  7 46 63 88 87  0
#> 270 86 46 63 88 87  0
#> 271  7 84 63 88 87  0
#> 272 86 84 63 88 87  0
#> 273  7 46 68 88 87  0
#> 274 86 46 68 88 87  0
#> 275  7 84 68 88 87  0
#> 278 86 42  6 38 40  0
#> 279 38 84  6 38 40  0
#> 280 86 84  6 38 40  0
#> 281 38 42 68 38 40  0
#> 282 86 42 68 38 40  0
#> 283 38 84 68 38 40  0
#> 284 86 84 68 38 40  0
#> 285 38 42  6 88 40  0
#> 286 86 42  6 88 40  0
#> 287 38 84  6 88 40  0
#> 288 86 84  6 88 40  0
#> 289 38 42 68 88 40  0
#> 290 86 42 68 88 40  0
#> 291 38 84 68 88 40  0
#> 292 86 84 68 88 40  0
#> 293 38 42  6 38 87  0
#> 294 86 42  6 38 87  0
#> 295 38 84  6 38 87  0
#> 296 86 84  6 38 87  0
#> 297 38 42 68 38 87  0
#> 298 86 42 68 38 87  0
#> 299 38 84 68 38 87  0
#> 300 86 84 68 38 87  0
#> 301 38 42  6 88 87  0
#> 302 86 42  6 88 87  0
#> 303 38 84  6 88 87  0
#> 304 86 84  6 88 87  0
#> 305 38 42 68 88 87  0
#> 306 86 42 68 88 87  0
#> 307 38 84 68 88 87  0
#> 310 86  9 45 58 69  0
#> 311 34 84 45 58 69  0
#> 312 86 84 45 58 69  0
#> 313 34  9 68 58 69  0
#> 314 86  9 68 58 69  0
#> 315 34 84 68 58 69  0
#> 316 86 84 68 58 69  0
#> 317 34  9 45 88 69  0
#> 318 86  9 45 88 69  0
#> 319 34 84 45 88 69  0
#> 320 86 84 45 88 69  0
#> 321 34  9 68 88 69  0
#> 322 86  9 68 88 69  0
#> 323 34 84 68 88 69  0
#> 324 86 84 68 88 69  0
#> 325 34  9 45 58 87  0
#> 326 86  9 45 58 87  0
#> 327 34 84 45 58 87  0
#> 328 86 84 45 58 87  0
#> 329 34  9 68 58 87  0
#> 330 86  9 68 58 87  0
#> 331 34 84 68 58 87  0
#> 332 86 84 68 58 87  0
#> 333 34  9 45 88 87  0
#> 334 86  9 45 88 87  0
#> 335 34 84 45 88 87  0
#> 336 86 84 45 88 87  0
#> 337 34  9 68 88 87  0
#> 338 86  9 68 88 87  0
#> 339 34 84 68 88 87  0
#> 342 86 79 44 34 15  0
#> 343 57 84 44 34 15  0
#> 344 86 84 44 34 15  0
#> 345 57 79 68 34 15  0
#> 346 86 79 68 34 15  0
#> 347 57 84 68 34 15  0
#> 348 86 84 68 34 15  0
#> 349 57 79 44 88 15  0
#> 350 86 79 44 88 15  0
#> 351 57 84 44 88 15  0
#> 352 86 84 44 88 15  0
#> 353 57 79 68 88 15  0
#> 354 86 79 68 88 15  0
#> 355 57 84 68 88 15  0
#> 356 86 84 68 88 15  0
#> 357 57 79 44 34 87  0
#> 358 86 79 44 34 87  0
#> 359 57 84 44 34 87  0
#> 360 86 84 44 34 87  0
#> 361 57 79 68 34 87  0
#> 362 86 79 68 34 87  0
#> 363 57 84 68 34 87  0
#> 364 86 84 68 34 87  0
#> 365 57 79 44 88 87  0
#> 366 86 79 44 88 87  0
#> 367 57 84 44 88 87  0
#> 368 86 84 44 88 87  0
#> 369 57 79 68 88 87  0
#> 370 86 79 68 88 87  0
#> 371 57 84 68 88 87  0
#> 374 86 26 58 13 71  0
#> 375 48 84 58 13 71  0
#> 376 86 84 58 13 71  0
#> 377 48 26 68 13 71  0
#> 378 86 26 68 13 71  0
#> 379 48 84 68 13 71  0
#> 380 86 84 68 13 71  0
#> 381 48 26 58 88 71  0
#> 382 86 26 58 88 71  0
#> 383 48 84 58 88 71  0
#> 384 86 84 58 88 71  0
#> 385 48 26 68 88 71  0
#> 386 86 26 68 88 71  0
#> 387 48 84 68 88 71  0
#> 388 86 84 68 88 71  0
#> 389 48 26 58 13 87  0
#> 390 86 26 58 13 87  0
#> 391 48 84 58 13 87  0
#> 392 86 84 58 13 87  0
#> 393 48 26 68 13 87  0
#> 394 86 26 68 13 87  0
#> 395 48 84 68 13 87  0
#> 396 86 84 68 13 87  0
#> 397 48 26 58 88 87  0
#> 398 86 26 58 88 87  0
#> 399 48 84 58 88 87  0
#> 400 86 84 58 88 87  0
#> 401 48 26 68 88 87  0
#> 402 86 26 68 88 87  0
#> 403 48 84 68 88 87  0
#> 406 86 34 46 13 67  0
#> 407 29 84 46 13 67  0
#> 408 86 84 46 13 67  0
#> 409 29 34 68 13 67  0
#> 410 86 34 68 13 67  0
#> 411 29 84 68 13 67  0
#> 412 86 84 68 13 67  0
#> 413 29 34 46 88 67  0
#> 414 86 34 46 88 67  0
#> 415 29 84 46 88 67  0
#> 416 86 84 46 88 67  0
#> 417 29 34 68 88 67  0
#> 418 86 34 68 88 67  0
#> 419 29 84 68 88 67  0
#> 420 86 84 68 88 67  0
#> 421 29 34 46 13 87  0
#> 422 86 34 46 13 87  0
#> 423 29 84 46 13 87  0
#> 424 86 84 46 13 87  0
#> 425 29 34 68 13 87  0
#> 426 86 34 68 13 87  0
#> 427 29 84 68 13 87  0
#> 429 29 34 46 88 87  0
#> 430 86 34 46 88 87  0
#> 431 29 84 46 88 87  0
#> 432 86 84 46 88 87  0
#> 433 29 34 68 88 87  0
#> 434 86 34 68 88 87  0
#> 435 29 84 68 88 87  0
#> 438 86  6 59 69 50  0
#> 439 53 84 59 69 50  0
#> 440 86 84 59 69 50  0
#> 441 53  6 68 69 50  0
#> 442 86  6 68 69 50  0
#> 443 53 84 68 69 50  0
#> 444 86 84 68 69 50  0
#> 445 53  6 59 88 50  0
#> 446 86  6 59 88 50  0
#> 447 53 84 59 88 50  0
#> 448 86 84 59 88 50  0
#> 449 53  6 68 88 50  0
#> 450 86  6 68 88 50  0
#> 451 53 84 68 88 50  0
#> 452 86 84 68 88 50  0
#> 453 53  6 59 69 87  0
#> 454 86  6 59 69 87  0
#> 455 53 84 59 69 87  0
#> 456 86 84 59 69 87  0
#> 457 53  6 68 69 87  0
#> 458 86  6 68 69 87  0
#> 459 53 84 68 69 87  0
#> 460 86 84 68 69 87  0
#> 461 53  6 59 88 87  0
#> 462 86  6 59 88 87  0
#> 463 53 84 59 88 87  0
#> 464 86 84 59 88 87  0
#> 465 53  6 68 88 87  0
#> 466 86  6 68 88 87  0
#> 467 53 84 68 88 87  0
#> 470 86 68 58 11 60  0
#> 471 30 84 58 11 60  0
#> 472 86 84 58 11 60  0
#> 473 30 68 68 11 60  0
#> 474 86 68 68 11 60  0
#> 475 30 84 68 11 60  0
#> 476 86 84 68 11 60  0
#> 477 30 68 58 88 60  0
#> 478 86 68 58 88 60  0
#> 479 30 84 58 88 60  0
#> 480 86 84 58 88 60  0
#> 481 30 68 68 88 60  0
#> 482 86 68 68 88 60  0
#> 483 30 84 68 88 60  0
#> 484 86 84 68 88 60  0
#> 485 30 68 58 11 87  0
#> 486 86 68 58 11 87  0
#> 487 30 84 58 11 87  0
#> 488 86 84 58 11 87  0
#> 489 30 68 68 11 87  0
#> 490 86 68 68 11 87  0
#> 491 30 84 68 11 87  0
#> 492 86 84 68 11 87  0
#> 493 30 68 58 88 87  0
#> 494 86 68 58 88 87  0
#> 495 30 84 58 88 87  0
#> 497 30 68 68 88 87  0
#> 498 86 68 68 88 87  0
#> 499 30 84 68 88 87  0
#> 502 86 56 35 42 66  0
#> 503  7 84 35 42 66  0
#> 504 86 84 35 42 66  0
#> 505  7 56 68 42 66  0
#> 506 86 56 68 42 66  0
#> 507  7 84 68 42 66  0
#> 508 86 84 68 42 66  0
#> 509  7 56 35 88 66  0
#> 510 86 56 35 88 66  0
#> 511  7 84 35 88 66  0
#> 512 86 84 35 88 66  0
#> 513  7 56 68 88 66  0
#> 514 86 56 68 88 66  0
#> 515  7 84 68 88 66  0
#> 516 86 84 68 88 66  0
#> 517  7 56 35 42 87  0
#> 518 86 56 35 42 87  0
#> 519  7 84 35 42 87  0
#> 520 86 84 35 42 87  0
#> 521  7 56 68 42 87  0
#> 522 86 56 68 42 87  0
#> 523  7 84 68 42 87  0
#> 524 86 84 68 42 87  0
#> 525  7 56 35 88 87  0
#> 526 86 56 35 88 87  0
#> 527  7 84 35 88 87  0
#> 528 86 84 35 88 87  0
#> 529  7 56 68 88 87  0
#> 530 86 56 68 88 87  0
#> 534 86 45 47  5 54  0
#> 535 69 84 47  5 54  0
#> 536 86 84 47  5 54  0
#> 537 69 45 68  5 54  0
#> 538 86 45 68  5 54  0
#> 539 69 84 68  5 54  0
#> 540 86 84 68  5 54  0
#> 541 69 45 47 88 54  0
#> 542 86 45 47 88 54  0
#> 543 69 84 47 88 54  0
#> 544 86 84 47 88 54  0
#> 545 69 45 68 88 54  0
#> 546 86 45 68 88 54  0
#> 547 69 84 68 88 54  0
#> 548 86 84 68 88 54  0
#> 549 69 45 47  5 87  0
#> 550 86 45 47  5 87  0
#> 551 69 84 47  5 87  0
#> 552 86 84 47  5 87  0
#> 553 69 45 68  5 87  0
#> 554 86 45 68  5 87  0
#> 555 69 84 68  5 87  0
#> 556 86 84 68  5 87  0
#> 557 69 45 47 88 87  0
#> 558 86 45 47 88 87  0
#> 559 69 84 47 88 87  0
#> 560 86 84 47 88 87  0
#> 561 69 45 68 88 87  0
#> 566 86 60 32 20 69  0
#> 567 23 84 32 20 69  0
#> 568 86 84 32 20 69  0
#> 569 23 60 68 20 69  0
#> 570 86 60 68 20 69  0
#> 571 23 84 68 20 69  0
#> 572 86 84 68 20 69  0
#> 573 23 60 32 88 69  0
#> 574 86 60 32 88 69  0
#> 575 23 84 32 88 69  0
#> 576 86 84 32 88 69  0
#> 577 23 60 68 88 69  0
#> 578 86 60 68 88 69  0
#> 579 23 84 68 88 69  0
#> 581 23 60 32 20 87  0
#> 582 86 60 32 20 87  0
#> 583 23 84 32 20 87  0
#> 584 86 84 32 20 87  0
#> 585 23 60 68 20 87  0
#> 586 86 60 68 20 87  0
#> 587 23 84 68 20 87  0
#> 588 86 84 68 20 87  0
#> 589 23 60 32 88 87  0
#> 590 86 60 32 88 87  0
#> 591 23 84 32 88 87  0
#> 592 86 84 32 88 87  0
#> 593 23 60 68 88 87  0
#> 594 86 60 68 88 87  0
#> 598 86 25 30 76 72  0
#> 599 34 84 30 76 72  0
#> 600 86 84 30 76 72  0
#> 601 34 25 68 76 72  0
#> 602 86 25 68 76 72  0
#> 603 34 84 68 76 72  0
#> 604 86 84 68 76 72  0
#> 605 34 25 30 88 72  0
#> 606 86 25 30 88 72  0
#> 607 34 84 30 88 72  0
#> 608 86 84 30 88 72  0
#> 609 34 25 68 88 72  0
#> 610 86 25 68 88 72  0
#> 611 34 84 68 88 72  0
#> 612 86 84 68 88 72  0
#> 613 34 25 30 76 87  0
#> 614 86 25 30 76 87  0
#> 615 34 84 30 76 87  0
#> 616 86 84 30 76 87  0
#> 617 34 25 68 76 87  0
#> 618 86 25 68 76 87  0
#> 619 34 84 68 76 87  0
#> 620 86 84 68 76 87  0
#> 621 34 25 30 88 87  0
#> 622 86 25 30 88 87  0
#> 623 34 84 30 88 87  0
#> 624 86 84 30 88 87  0
#> 625 34 25 68 88 87  0
#> 626 86 25 68 88 87  0
#> 630 86 19 37 77 68  0
#> 631 35 84 37 77 68  0
#> 632 86 84 37 77 68  0
#> 633 35 19 68 77 68  0
#> 634 86 19 68 77 68  0
#> 635 35 84 68 77 68  0
#> 636 86 84 68 77 68  0
#> 637 35 19 37 88 68  0
#> 638 86 19 37 88 68  0
#> 639 35 84 37 88 68  0
#> 640 86 84 37 88 68  0
#> 641 35 19 68 88 68  0
#> 642 86 19 68 88 68  0
#> 643 35 84 68 88 68  0
#> 644 86 84 68 88 68  0
#> 645 35 19 37 77 87  0
#> 646 86 19 37 77 87  0
#> 647 35 84 37 77 87  0
#> 648 86 84 37 77 87  0
#> 649 35 19 68 77 87  0
#> 650 86 19 68 77 87  0
#> 651 35 84 68 77 87  0
#> 652 86 84 68 77 87  0
#> 653 35 19 37 88 87  0
#> 654 86 19 37 88 87  0
#> 655 35 84 37 88 87  0
#> 656 86 84 37 88 87  0
#> 657 35 19 68 88 87  0
#> 658 86 19 68 88 87  0
#> 659 35 84 68 88 87  0