Plot TeX in the margin

Plot TeX in the margin

Usage

plotMTeX3D(tex, edge, line = 0, at = NULL, pos = NA, ...)

Arguments

tex

TeX string

edge

The position at which to draw the axis or text.

line

The “line” of the plot margin to draw the label on.

at

The value of a coordinate at which to draw the axis.

pos

The position at which to draw the axis or text.

...

Further arguments passed to plotTeX3D().

Value

The object IDs of objects added to the scene.