Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 2   5 66 39  1
#> 7  17 18 34  1
#> 1  17 44 15  1
#> 6  32 66  7  1
#> 3  36 53  3  1
#> 9  66 23 12  1
#> 5  74 41  8  1
#> 21 79 66 39  0
#> 31  5 71 39  0
#> 4  79 71 39  0
#> 51  5 66 44  0
#> 61 79 66 44  0
#> 71  5 71 44  0
#> 8  79 71 44  0
#> 22 79 18 34  0
#> 32 17 71 34  0
#> 41 79 71 34  0
#> 52 17 18 44  0
#> 62 79 18 44  0
#> 72 17 71 44  0
#> 23 79 44 15  0
#> 33 17 71 15  0
#> 42 79 71 15  0
#> 53 17 44 44  0
#> 63 79 44 44  0
#> 24 79 66  7  0
#> 34 32 71  7  0
#> 43 79 71  7  0
#> 54 32 66 44  0
#> 74 32 71 44  0
#> 25 79 53  3  0
#> 35 36 71  3  0
#> 44 79 71  3  0
#> 55 36 53 44  0
#> 65 79 53 44  0
#> 75 36 71 44  0
#> 26 79 23 12  0
#> 36 66 71 12  0
#> 45 79 71 12  0
#> 56 66 23 44  0
#> 66 79 23 44  0
#> 76 66 71 44  0
#> 27 79 41  8  0
#> 37 74 71  8  0
#> 46 79 71  8  0
#> 57 74 41 44  0
#> 67 79 41 44  0
#> 77 74 71 44  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 2   5 66 39  1
#> 7  17 18 34  1
#> 1  17 44 15  1
#> 6  32 66  7  1
#> 3  36 53  3  1
#> 9  66 23 12  1
#> 5  74 41  8  1
#> 21 79 66 39  0
#> 31  5 13 39  0
#> 4  79 13 39  0
#> 51  5 66 44  0
#> 61 79 66 44  0
#> 71  5 13 44  0
#> 8  79 13 44  0
#> 22 79 18 34  0
#> 32 17 13 34  0
#> 41 79 13 34  0
#> 52 17 18 44  0
#> 62 79 18 44  0
#> 72 17 13 44  0
#> 23 79 44 15  0
#> 33 17 13 15  0
#> 42 79 13 15  0
#> 53 17 44 44  0
#> 63 79 44 44  0
#> 24 79 66  7  0
#> 34 32 13  7  0
#> 43 79 13  7  0
#> 54 32 66 44  0
#> 74 32 13 44  0
#> 25 79 53  3  0
#> 35 36 13  3  0
#> 44 79 13  3  0
#> 55 36 53 44  0
#> 65 79 53 44  0
#> 75 36 13 44  0
#> 26 79 23 12  0
#> 36 66 13 12  0
#> 45 79 13 12  0
#> 56 66 23 44  0
#> 66 79 23 44  0
#> 76 66 13 44  0
#> 27 79 41  8  0
#> 37 74 13  8  0
#> 46 79 13  8  0
#> 57 74 41 44  0
#> 67 79 41 44  0
#> 77 74 13 44  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 2   5 66  39  1
#> 7  17 18  34  1
#> 1  17 44  15  1
#> 6  32 66   7  1
#> 3  36 53   3  1
#> 9  66 23  12  1
#> 5  74 41   8  1
#> 21  0 66  39  0
#> 31  5  0  39  0
#> 4   0  0  39  0
#> 51  5 66 100  0
#> 61  0 66 100  0
#> 71  5  0 100  0
#> 8   0  0 100  0
#> 22  0 18  34  0
#> 32 17  0  34  0
#> 41  0  0  34  0
#> 52 17 18 100  0
#> 62  0 18 100  0
#> 72 17  0 100  0
#> 23  0 44  15  0
#> 33 17  0  15  0
#> 42  0  0  15  0
#> 53 17 44 100  0
#> 63  0 44 100  0
#> 24  0 66   7  0
#> 34 32  0   7  0
#> 43  0  0   7  0
#> 54 32 66 100  0
#> 74 32  0 100  0
#> 25  0 53   3  0
#> 35 36  0   3  0
#> 44  0  0   3  0
#> 55 36 53 100  0
#> 65  0 53 100  0
#> 75 36  0 100  0
#> 26  0 23  12  0
#> 36 66  0  12  0
#> 45  0  0  12  0
#> 56 66 23 100  0
#> 66  0 23 100  0
#> 76 66  0 100  0
#> 27  0 41   8  0
#> 37 74  0   8  0
#> 46  0  0   8  0
#> 57 74 41 100  0
#> 67  0 41 100  0
#> 77 74  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   15 75 34 52 68  1
#> 2   65 66 58 13 28  1
#> 3   64 55  6 65 59  1
#> 4   26 69 75 49 20  1
#> 5   26 55 78 67 23  1
#> 6   65  5 46 38 56  1
#> 7   48 88 27 42 30  1
#> 8   15 55 19 64 50  1
#> 9   24 17 51 25 56  1
#> 10  67 52 24 14 37  1
#> 11  61 42 30 64 10  1
#> 12  38 25 22 68 73  1
#> 13  76 46  9 45 46  1
#> 14  36 74 16 54 29  1
#> 15  31 28 83 29 52  1
#> 16  14 44 52 27 27  1
#> 17  56  9 43 36 30  1
#> 18  66 71  9 43 47  1
#> 19  51 61  5 50 68  1
#> 20  41  3 59 51 52  1
#> 22  81 75 34 52 68  0
#> 23  15 93 34 52 68  0
#> 24  81 93 34 52 68  0
#> 25  15 75 88 52 68  0
#> 26  81 75 88 52 68  0
#> 27  15 93 88 52 68  0
#> 28  81 93 88 52 68  0
#> 29  15 75 34 73 68  0
#> 30  81 75 34 73 68  0
#> 31  15 93 34 73 68  0
#> 32  81 93 34 73 68  0
#> 33  15 75 88 73 68  0
#> 34  81 75 88 73 68  0
#> 35  15 93 88 73 68  0
#> 36  81 93 88 73 68  0
#> 37  15 75 34 52 78  0
#> 38  81 75 34 52 78  0
#> 39  15 93 34 52 78  0
#> 40  81 93 34 52 78  0
#> 41  15 75 88 52 78  0
#> 42  81 75 88 52 78  0
#> 43  15 93 88 52 78  0
#> 44  81 93 88 52 78  0
#> 45  15 75 34 73 78  0
#> 46  81 75 34 73 78  0
#> 47  15 93 34 73 78  0
#> 48  81 93 34 73 78  0
#> 49  15 75 88 73 78  0
#> 50  81 75 88 73 78  0
#> 51  15 93 88 73 78  0
#> 52  81 93 88 73 78  0
#> 54  81 66 58 13 28  0
#> 55  65 93 58 13 28  0
#> 56  81 93 58 13 28  0
#> 57  65 66 88 13 28  0
#> 58  81 66 88 13 28  0
#> 59  65 93 88 13 28  0
#> 60  81 93 88 13 28  0
#> 61  65 66 58 73 28  0
#> 62  81 66 58 73 28  0
#> 63  65 93 58 73 28  0
#> 64  81 93 58 73 28  0
#> 65  65 66 88 73 28  0
#> 66  81 66 88 73 28  0
#> 67  65 93 88 73 28  0
#> 68  81 93 88 73 28  0
#> 69  65 66 58 13 78  0
#> 70  81 66 58 13 78  0
#> 71  65 93 58 13 78  0
#> 72  81 93 58 13 78  0
#> 73  65 66 88 13 78  0
#> 74  81 66 88 13 78  0
#> 75  65 93 88 13 78  0
#> 76  81 93 88 13 78  0
#> 77  65 66 58 73 78  0
#> 78  81 66 58 73 78  0
#> 79  65 93 58 73 78  0
#> 80  81 93 58 73 78  0
#> 81  65 66 88 73 78  0
#> 82  81 66 88 73 78  0
#> 83  65 93 88 73 78  0
#> 86  81 55  6 65 59  0
#> 87  64 93  6 65 59  0
#> 88  81 93  6 65 59  0
#> 89  64 55 88 65 59  0
#> 90  81 55 88 65 59  0
#> 91  64 93 88 65 59  0
#> 92  81 93 88 65 59  0
#> 93  64 55  6 73 59  0
#> 94  81 55  6 73 59  0
#> 95  64 93  6 73 59  0
#> 96  81 93  6 73 59  0
#> 97  64 55 88 73 59  0
#> 98  81 55 88 73 59  0
#> 99  64 93 88 73 59  0
#> 100 81 93 88 73 59  0
#> 101 64 55  6 65 78  0
#> 102 81 55  6 65 78  0
#> 103 64 93  6 65 78  0
#> 104 81 93  6 65 78  0
#> 105 64 55 88 65 78  0
#> 106 81 55 88 65 78  0
#> 107 64 93 88 65 78  0
#> 108 81 93 88 65 78  0
#> 109 64 55  6 73 78  0
#> 110 81 55  6 73 78  0
#> 111 64 93  6 73 78  0
#> 112 81 93  6 73 78  0
#> 113 64 55 88 73 78  0
#> 114 81 55 88 73 78  0
#> 115 64 93 88 73 78  0
#> 118 81 69 75 49 20  0
#> 119 26 93 75 49 20  0
#> 120 81 93 75 49 20  0
#> 121 26 69 88 49 20  0
#> 122 81 69 88 49 20  0
#> 123 26 93 88 49 20  0
#> 124 81 93 88 49 20  0
#> 125 26 69 75 73 20  0
#> 126 81 69 75 73 20  0
#> 127 26 93 75 73 20  0
#> 128 81 93 75 73 20  0
#> 129 26 69 88 73 20  0
#> 130 81 69 88 73 20  0
#> 131 26 93 88 73 20  0
#> 132 81 93 88 73 20  0
#> 133 26 69 75 49 78  0
#> 134 81 69 75 49 78  0
#> 135 26 93 75 49 78  0
#> 136 81 93 75 49 78  0
#> 137 26 69 88 49 78  0
#> 138 81 69 88 49 78  0
#> 139 26 93 88 49 78  0
#> 140 81 93 88 49 78  0
#> 141 26 69 75 73 78  0
#> 142 81 69 75 73 78  0
#> 143 26 93 75 73 78  0
#> 144 81 93 75 73 78  0
#> 145 26 69 88 73 78  0
#> 146 81 69 88 73 78  0
#> 147 26 93 88 73 78  0
#> 150 81 55 78 67 23  0
#> 151 26 93 78 67 23  0
#> 152 81 93 78 67 23  0
#> 153 26 55 88 67 23  0
#> 154 81 55 88 67 23  0
#> 155 26 93 88 67 23  0
#> 156 81 93 88 67 23  0
#> 157 26 55 78 73 23  0
#> 158 81 55 78 73 23  0
#> 159 26 93 78 73 23  0
#> 160 81 93 78 73 23  0
#> 161 26 55 88 73 23  0
#> 162 81 55 88 73 23  0
#> 163 26 93 88 73 23  0
#> 164 81 93 88 73 23  0
#> 165 26 55 78 67 78  0
#> 166 81 55 78 67 78  0
#> 167 26 93 78 67 78  0
#> 168 81 93 78 67 78  0
#> 169 26 55 88 67 78  0
#> 170 81 55 88 67 78  0
#> 171 26 93 88 67 78  0
#> 172 81 93 88 67 78  0
#> 173 26 55 78 73 78  0
#> 174 81 55 78 73 78  0
#> 175 26 93 78 73 78  0
#> 176 81 93 78 73 78  0
#> 177 26 55 88 73 78  0
#> 182 81  5 46 38 56  0
#> 183 65 93 46 38 56  0
#> 184 81 93 46 38 56  0
#> 185 65  5 88 38 56  0
#> 186 81  5 88 38 56  0
#> 187 65 93 88 38 56  0
#> 188 81 93 88 38 56  0
#> 189 65  5 46 73 56  0
#> 190 81  5 46 73 56  0
#> 191 65 93 46 73 56  0
#> 192 81 93 46 73 56  0
#> 193 65  5 88 73 56  0
#> 194 81  5 88 73 56  0
#> 195 65 93 88 73 56  0
#> 196 81 93 88 73 56  0
#> 197 65  5 46 38 78  0
#> 198 81  5 46 38 78  0
#> 199 65 93 46 38 78  0
#> 200 81 93 46 38 78  0
#> 201 65  5 88 38 78  0
#> 202 81  5 88 38 78  0
#> 203 65 93 88 38 78  0
#> 204 81 93 88 38 78  0
#> 205 65  5 46 73 78  0
#> 206 81  5 46 73 78  0
#> 207 65 93 46 73 78  0
#> 208 81 93 46 73 78  0
#> 209 65  5 88 73 78  0
#> 210 81  5 88 73 78  0
#> 214 81 88 27 42 30  0
#> 215 48 93 27 42 30  0
#> 216 81 93 27 42 30  0
#> 217 48 88 88 42 30  0
#> 218 81 88 88 42 30  0
#> 219 48 93 88 42 30  0
#> 220 81 93 88 42 30  0
#> 221 48 88 27 73 30  0
#> 222 81 88 27 73 30  0
#> 223 48 93 27 73 30  0
#> 224 81 93 27 73 30  0
#> 225 48 88 88 73 30  0
#> 226 81 88 88 73 30  0
#> 227 48 93 88 73 30  0
#> 228 81 93 88 73 30  0
#> 229 48 88 27 42 78  0
#> 230 81 88 27 42 78  0
#> 231 48 93 27 42 78  0
#> 232 81 93 27 42 78  0
#> 233 48 88 88 42 78  0
#> 234 81 88 88 42 78  0
#> 235 48 93 88 42 78  0
#> 236 81 93 88 42 78  0
#> 237 48 88 27 73 78  0
#> 238 81 88 27 73 78  0
#> 239 48 93 27 73 78  0
#> 240 81 93 27 73 78  0
#> 241 48 88 88 73 78  0
#> 242 81 88 88 73 78  0
#> 243 48 93 88 73 78  0
#> 246 81 55 19 64 50  0
#> 247 15 93 19 64 50  0
#> 248 81 93 19 64 50  0
#> 249 15 55 88 64 50  0
#> 250 81 55 88 64 50  0
#> 251 15 93 88 64 50  0
#> 252 81 93 88 64 50  0
#> 253 15 55 19 73 50  0
#> 254 81 55 19 73 50  0
#> 255 15 93 19 73 50  0
#> 256 81 93 19 73 50  0
#> 257 15 55 88 73 50  0
#> 258 81 55 88 73 50  0
#> 259 15 93 88 73 50  0
#> 260 81 93 88 73 50  0
#> 261 15 55 19 64 78  0
#> 262 81 55 19 64 78  0
#> 263 15 93 19 64 78  0
#> 264 81 93 19 64 78  0
#> 265 15 55 88 64 78  0
#> 266 81 55 88 64 78  0
#> 267 15 93 88 64 78  0
#> 268 81 93 88 64 78  0
#> 269 15 55 19 73 78  0
#> 270 81 55 19 73 78  0
#> 271 15 93 19 73 78  0
#> 272 81 93 19 73 78  0
#> 273 15 55 88 73 78  0
#> 278 81 17 51 25 56  0
#> 279 24 93 51 25 56  0
#> 280 81 93 51 25 56  0
#> 281 24 17 88 25 56  0
#> 282 81 17 88 25 56  0
#> 283 24 93 88 25 56  0
#> 284 81 93 88 25 56  0
#> 285 24 17 51 73 56  0
#> 286 81 17 51 73 56  0
#> 287 24 93 51 73 56  0
#> 288 81 93 51 73 56  0
#> 289 24 17 88 73 56  0
#> 290 81 17 88 73 56  0
#> 291 24 93 88 73 56  0
#> 293 24 17 51 25 78  0
#> 294 81 17 51 25 78  0
#> 295 24 93 51 25 78  0
#> 296 81 93 51 25 78  0
#> 297 24 17 88 25 78  0
#> 298 81 17 88 25 78  0
#> 299 24 93 88 25 78  0
#> 300 81 93 88 25 78  0
#> 301 24 17 51 73 78  0
#> 302 81 17 51 73 78  0
#> 303 24 93 51 73 78  0
#> 304 81 93 51 73 78  0
#> 305 24 17 88 73 78  0
#> 306 81 17 88 73 78  0
#> 307 24 93 88 73 78  0
#> 310 81 52 24 14 37  0
#> 311 67 93 24 14 37  0
#> 312 81 93 24 14 37  0
#> 313 67 52 88 14 37  0
#> 314 81 52 88 14 37  0
#> 315 67 93 88 14 37  0
#> 316 81 93 88 14 37  0
#> 317 67 52 24 73 37  0
#> 318 81 52 24 73 37  0
#> 319 67 93 24 73 37  0
#> 320 81 93 24 73 37  0
#> 321 67 52 88 73 37  0
#> 322 81 52 88 73 37  0
#> 323 67 93 88 73 37  0
#> 324 81 93 88 73 37  0
#> 325 67 52 24 14 78  0
#> 326 81 52 24 14 78  0
#> 327 67 93 24 14 78  0
#> 328 81 93 24 14 78  0
#> 329 67 52 88 14 78  0
#> 330 81 52 88 14 78  0
#> 331 67 93 88 14 78  0
#> 332 81 93 88 14 78  0
#> 333 67 52 24 73 78  0
#> 334 81 52 24 73 78  0
#> 335 67 93 24 73 78  0
#> 336 81 93 24 73 78  0
#> 337 67 52 88 73 78  0
#> 338 81 52 88 73 78  0
#> 339 67 93 88 73 78  0
#> 342 81 42 30 64 10  0
#> 343 61 93 30 64 10  0
#> 344 81 93 30 64 10  0
#> 345 61 42 88 64 10  0
#> 346 81 42 88 64 10  0
#> 347 61 93 88 64 10  0
#> 348 81 93 88 64 10  0
#> 349 61 42 30 73 10  0
#> 350 81 42 30 73 10  0
#> 351 61 93 30 73 10  0
#> 352 81 93 30 73 10  0
#> 353 61 42 88 73 10  0
#> 354 81 42 88 73 10  0
#> 355 61 93 88 73 10  0
#> 356 81 93 88 73 10  0
#> 357 61 42 30 64 78  0
#> 358 81 42 30 64 78  0
#> 359 61 93 30 64 78  0
#> 360 81 93 30 64 78  0
#> 361 61 42 88 64 78  0
#> 362 81 42 88 64 78  0
#> 363 61 93 88 64 78  0
#> 365 61 42 30 73 78  0
#> 366 81 42 30 73 78  0
#> 367 61 93 30 73 78  0
#> 368 81 93 30 73 78  0
#> 369 61 42 88 73 78  0
#> 370 81 42 88 73 78  0
#> 371 61 93 88 73 78  0
#> 374 81 25 22 68 73  0
#> 375 38 93 22 68 73  0
#> 376 81 93 22 68 73  0
#> 377 38 25 88 68 73  0
#> 378 81 25 88 68 73  0
#> 379 38 93 88 68 73  0
#> 380 81 93 88 68 73  0
#> 381 38 25 22 73 73  0
#> 382 81 25 22 73 73  0
#> 383 38 93 22 73 73  0
#> 384 81 93 22 73 73  0
#> 385 38 25 88 73 73  0
#> 386 81 25 88 73 73  0
#> 387 38 93 88 73 73  0
#> 388 81 93 88 73 73  0
#> 389 38 25 22 68 78  0
#> 390 81 25 22 68 78  0
#> 391 38 93 22 68 78  0
#> 392 81 93 22 68 78  0
#> 393 38 25 88 68 78  0
#> 394 81 25 88 68 78  0
#> 395 38 93 88 68 78  0
#> 396 81 93 88 68 78  0
#> 397 38 25 22 73 78  0
#> 398 81 25 22 73 78  0
#> 399 38 93 22 73 78  0
#> 400 81 93 22 73 78  0
#> 401 38 25 88 73 78  0
#> 402 81 25 88 73 78  0
#> 403 38 93 88 73 78  0
#> 406 81 46  9 45 46  0
#> 407 76 93  9 45 46  0
#> 408 81 93  9 45 46  0
#> 409 76 46 88 45 46  0
#> 410 81 46 88 45 46  0
#> 411 76 93 88 45 46  0
#> 412 81 93 88 45 46  0
#> 413 76 46  9 73 46  0
#> 414 81 46  9 73 46  0
#> 415 76 93  9 73 46  0
#> 416 81 93  9 73 46  0
#> 417 76 46 88 73 46  0
#> 418 81 46 88 73 46  0
#> 419 76 93 88 73 46  0
#> 420 81 93 88 73 46  0
#> 421 76 46  9 45 78  0
#> 422 81 46  9 45 78  0
#> 423 76 93  9 45 78  0
#> 424 81 93  9 45 78  0
#> 425 76 46 88 45 78  0
#> 426 81 46 88 45 78  0
#> 427 76 93 88 45 78  0
#> 428 81 93 88 45 78  0
#> 429 76 46  9 73 78  0
#> 430 81 46  9 73 78  0
#> 431 76 93  9 73 78  0
#> 432 81 93  9 73 78  0
#> 433 76 46 88 73 78  0
#> 434 81 46 88 73 78  0
#> 435 76 93 88 73 78  0
#> 438 81 74 16 54 29  0
#> 439 36 93 16 54 29  0
#> 440 81 93 16 54 29  0
#> 441 36 74 88 54 29  0
#> 442 81 74 88 54 29  0
#> 443 36 93 88 54 29  0
#> 444 81 93 88 54 29  0
#> 445 36 74 16 73 29  0
#> 446 81 74 16 73 29  0
#> 447 36 93 16 73 29  0
#> 448 81 93 16 73 29  0
#> 449 36 74 88 73 29  0
#> 450 81 74 88 73 29  0
#> 451 36 93 88 73 29  0
#> 452 81 93 88 73 29  0
#> 453 36 74 16 54 78  0
#> 454 81 74 16 54 78  0
#> 455 36 93 16 54 78  0
#> 456 81 93 16 54 78  0
#> 457 36 74 88 54 78  0
#> 458 81 74 88 54 78  0
#> 459 36 93 88 54 78  0
#> 460 81 93 88 54 78  0
#> 461 36 74 16 73 78  0
#> 462 81 74 16 73 78  0
#> 463 36 93 16 73 78  0
#> 464 81 93 16 73 78  0
#> 465 36 74 88 73 78  0
#> 466 81 74 88 73 78  0
#> 467 36 93 88 73 78  0
#> 470 81 28 83 29 52  0
#> 471 31 93 83 29 52  0
#> 472 81 93 83 29 52  0
#> 473 31 28 88 29 52  0
#> 474 81 28 88 29 52  0
#> 475 31 93 88 29 52  0
#> 476 81 93 88 29 52  0
#> 477 31 28 83 73 52  0
#> 478 81 28 83 73 52  0
#> 479 31 93 83 73 52  0
#> 480 81 93 83 73 52  0
#> 481 31 28 88 73 52  0
#> 482 81 28 88 73 52  0
#> 483 31 93 88 73 52  0
#> 484 81 93 88 73 52  0
#> 485 31 28 83 29 78  0
#> 486 81 28 83 29 78  0
#> 487 31 93 83 29 78  0
#> 488 81 93 83 29 78  0
#> 489 31 28 88 29 78  0
#> 490 81 28 88 29 78  0
#> 491 31 93 88 29 78  0
#> 492 81 93 88 29 78  0
#> 493 31 28 83 73 78  0
#> 494 81 28 83 73 78  0
#> 495 31 93 83 73 78  0
#> 496 81 93 83 73 78  0
#> 497 31 28 88 73 78  0
#> 498 81 28 88 73 78  0
#> 499 31 93 88 73 78  0
#> 502 81 44 52 27 27  0
#> 503 14 93 52 27 27  0
#> 504 81 93 52 27 27  0
#> 505 14 44 88 27 27  0
#> 506 81 44 88 27 27  0
#> 507 14 93 88 27 27  0
#> 508 81 93 88 27 27  0
#> 509 14 44 52 73 27  0
#> 510 81 44 52 73 27  0
#> 511 14 93 52 73 27  0
#> 512 81 93 52 73 27  0
#> 513 14 44 88 73 27  0
#> 514 81 44 88 73 27  0
#> 515 14 93 88 73 27  0
#> 516 81 93 88 73 27  0
#> 517 14 44 52 27 78  0
#> 518 81 44 52 27 78  0
#> 519 14 93 52 27 78  0
#> 520 81 93 52 27 78  0
#> 521 14 44 88 27 78  0
#> 522 81 44 88 27 78  0
#> 523 14 93 88 27 78  0
#> 524 81 93 88 27 78  0
#> 525 14 44 52 73 78  0
#> 526 81 44 52 73 78  0
#> 527 14 93 52 73 78  0
#> 528 81 93 52 73 78  0
#> 529 14 44 88 73 78  0
#> 530 81 44 88 73 78  0
#> 531 14 93 88 73 78  0
#> 534 81  9 43 36 30  0
#> 535 56 93 43 36 30  0
#> 536 81 93 43 36 30  0
#> 537 56  9 88 36 30  0
#> 538 81  9 88 36 30  0
#> 539 56 93 88 36 30  0
#> 540 81 93 88 36 30  0
#> 541 56  9 43 73 30  0
#> 542 81  9 43 73 30  0
#> 543 56 93 43 73 30  0
#> 544 81 93 43 73 30  0
#> 545 56  9 88 73 30  0
#> 546 81  9 88 73 30  0
#> 547 56 93 88 73 30  0
#> 549 56  9 43 36 78  0
#> 550 81  9 43 36 78  0
#> 551 56 93 43 36 78  0
#> 552 81 93 43 36 78  0
#> 553 56  9 88 36 78  0
#> 554 81  9 88 36 78  0
#> 555 56 93 88 36 78  0
#> 556 81 93 88 36 78  0
#> 557 56  9 43 73 78  0
#> 558 81  9 43 73 78  0
#> 559 56 93 43 73 78  0
#> 560 81 93 43 73 78  0
#> 561 56  9 88 73 78  0
#> 562 81  9 88 73 78  0
#> 563 56 93 88 73 78  0
#> 566 81 71  9 43 47  0
#> 567 66 93  9 43 47  0
#> 568 81 93  9 43 47  0
#> 569 66 71 88 43 47  0
#> 570 81 71 88 43 47  0
#> 571 66 93 88 43 47  0
#> 572 81 93 88 43 47  0
#> 573 66 71  9 73 47  0
#> 574 81 71  9 73 47  0
#> 575 66 93  9 73 47  0
#> 576 81 93  9 73 47  0
#> 577 66 71 88 73 47  0
#> 578 81 71 88 73 47  0
#> 579 66 93 88 73 47  0
#> 580 81 93 88 73 47  0
#> 581 66 71  9 43 78  0
#> 582 81 71  9 43 78  0
#> 583 66 93  9 43 78  0
#> 584 81 93  9 43 78  0
#> 585 66 71 88 43 78  0
#> 586 81 71 88 43 78  0
#> 587 66 93 88 43 78  0
#> 588 81 93 88 43 78  0
#> 589 66 71  9 73 78  0
#> 590 81 71  9 73 78  0
#> 591 66 93  9 73 78  0
#> 593 66 71 88 73 78  0
#> 594 81 71 88 73 78  0
#> 595 66 93 88 73 78  0
#> 598 81 61  5 50 68  0
#> 599 51 93  5 50 68  0
#> 600 81 93  5 50 68  0
#> 601 51 61 88 50 68  0
#> 602 81 61 88 50 68  0
#> 603 51 93 88 50 68  0
#> 604 81 93 88 50 68  0
#> 605 51 61  5 73 68  0
#> 606 81 61  5 73 68  0
#> 607 51 93  5 73 68  0
#> 608 81 93  5 73 68  0
#> 609 51 61 88 73 68  0
#> 610 81 61 88 73 68  0
#> 611 51 93 88 73 68  0
#> 613 51 61  5 50 78  0
#> 614 81 61  5 50 78  0
#> 615 51 93  5 50 78  0
#> 616 81 93  5 50 78  0
#> 617 51 61 88 50 78  0
#> 618 81 61 88 50 78  0
#> 619 51 93 88 50 78  0
#> 620 81 93 88 50 78  0
#> 621 51 61  5 73 78  0
#> 622 81 61  5 73 78  0
#> 623 51 93  5 73 78  0
#> 624 81 93  5 73 78  0
#> 625 51 61 88 73 78  0
#> 626 81 61 88 73 78  0
#> 627 51 93 88 73 78  0
#> 630 81  3 59 51 52  0
#> 631 41 93 59 51 52  0
#> 632 81 93 59 51 52  0
#> 633 41  3 88 51 52  0
#> 634 81  3 88 51 52  0
#> 635 41 93 88 51 52  0
#> 636 81 93 88 51 52  0
#> 637 41  3 59 73 52  0
#> 638 81  3 59 73 52  0
#> 639 41 93 59 73 52  0
#> 640 81 93 59 73 52  0
#> 641 41  3 88 73 52  0
#> 642 81  3 88 73 52  0
#> 643 41 93 88 73 52  0
#> 645 41  3 59 51 78  0
#> 646 81  3 59 51 78  0
#> 647 41 93 59 51 78  0
#> 648 81 93 59 51 78  0
#> 649 41  3 88 51 78  0
#> 650 81  3 88 51 78  0
#> 651 41 93 88 51 78  0
#> 652 81 93 88 51 78  0
#> 653 41  3 59 73 78  0
#> 654 81  3 59 73 78  0
#> 655 41 93 59 73 78  0
#> 656 81 93 59 73 78  0
#> 657 41  3 88 73 78  0
#> 658 81  3 88 73 78  0
#> 659 41 93 88 73 78  0