Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 2  36  5 38  1
#> 5  36 80 14  1
#> 10 40 82 13  1
#> 8  46 73  7  1
#> 7  50  7 26  1
#> 1  50 31  5  1
#> 6  60 62  3  1
#> 21 65  5 38  0
#> 3  36 87 38  0
#> 4  65 87 38  0
#> 51 36  5 43  0
#> 61 65  5 43  0
#> 71 36 87 43  0
#> 81 65 87 43  0
#> 22 65 80 14  0
#> 31 36 87 14  0
#> 41 65 87 14  0
#> 52 36 80 43  0
#> 62 65 80 43  0
#> 23 65 82 13  0
#> 32 40 87 13  0
#> 42 65 87 13  0
#> 53 40 82 43  0
#> 63 65 82 43  0
#> 73 40 87 43  0
#> 24 65 73  7  0
#> 33 46 87  7  0
#> 43 65 87  7  0
#> 54 46 73 43  0
#> 64 65 73 43  0
#> 74 46 87 43  0
#> 25 65  7 26  0
#> 34 50 87 26  0
#> 44 65 87 26  0
#> 55 50  7 43  0
#> 65 65  7 43  0
#> 75 50 87 43  0
#> 26 65 31  5  0
#> 35 50 87  5  0
#> 45 65 87  5  0
#> 56 50 31 43  0
#> 66 65 31 43  0
#> 27 65 62  3  0
#> 36 60 87  3  0
#> 46 65 87  3  0
#> 57 60 62 43  0
#> 67 65 62 43  0
#> 77 60 87 43  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 2  36  5 38  1
#> 5  36 80 14  1
#> 10 40 82 13  1
#> 8  46 73  7  1
#> 7  50  7 26  1
#> 1  50 31  5  1
#> 6  60 62  3  1
#> 21 65  5 38  0
#> 3  36  0 38  0
#> 4  65  0 38  0
#> 51 36  5 43  0
#> 61 65  5 43  0
#> 71 36  0 43  0
#> 81 65  0 43  0
#> 22 65 80 14  0
#> 31 36  0 14  0
#> 41 65  0 14  0
#> 52 36 80 43  0
#> 62 65 80 43  0
#> 23 65 82 13  0
#> 32 40  0 13  0
#> 42 65  0 13  0
#> 53 40 82 43  0
#> 63 65 82 43  0
#> 73 40  0 43  0
#> 24 65 73  7  0
#> 33 46  0  7  0
#> 43 65  0  7  0
#> 54 46 73 43  0
#> 64 65 73 43  0
#> 74 46  0 43  0
#> 25 65  7 26  0
#> 34 50  0 26  0
#> 44 65  0 26  0
#> 55 50  7 43  0
#> 65 65  7 43  0
#> 75 50  0 43  0
#> 26 65 31  5  0
#> 35 50  0  5  0
#> 45 65  0  5  0
#> 56 50 31 43  0
#> 66 65 31 43  0
#> 27 65 62  3  0
#> 36 60  0  3  0
#> 46 65  0  3  0
#> 57 60 62 43  0
#> 67 65 62 43  0
#> 77 60  0 43  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 2  36  5  38  1
#> 5  36 80  14  1
#> 10 40 82  13  1
#> 8  46 73   7  1
#> 7  50  7  26  1
#> 1  50 31   5  1
#> 6  60 62   3  1
#> 21  0  5  38  0
#> 3  36  0  38  0
#> 4   0  0  38  0
#> 51 36  5 100  0
#> 61  0  5 100  0
#> 71 36  0 100  0
#> 81  0  0 100  0
#> 22  0 80  14  0
#> 31 36  0  14  0
#> 41  0  0  14  0
#> 52 36 80 100  0
#> 62  0 80 100  0
#> 23  0 82  13  0
#> 32 40  0  13  0
#> 42  0  0  13  0
#> 53 40 82 100  0
#> 63  0 82 100  0
#> 73 40  0 100  0
#> 24  0 73   7  0
#> 33 46  0   7  0
#> 43  0  0   7  0
#> 54 46 73 100  0
#> 64  0 73 100  0
#> 74 46  0 100  0
#> 25  0  7  26  0
#> 34 50  0  26  0
#> 44  0  0  26  0
#> 55 50  7 100  0
#> 65  0  7 100  0
#> 75 50  0 100  0
#> 26  0 31   5  0
#> 35 50  0   5  0
#> 45  0  0   5  0
#> 56 50 31 100  0
#> 66  0 31 100  0
#> 27  0 62   3  0
#> 36 60  0   3  0
#> 46  0  0   3  0
#> 57 60 62 100  0
#> 67  0 62 100  0
#> 77 60  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   71 17 38 42 76  1
#> 2    8 37 44 56 70  1
#> 3   21 28 72 27 59  1
#> 4   32 13 31 54 32  1
#> 5   61 68 24 14 55  1
#> 6   43 71 66 30 14  1
#> 7   14 52 38 22 39  1
#> 8    8 53 55 66 30  1
#> 9   36 52 28 10 44  1
#> 10  12 78 50 36 55  1
#> 11  72 51 35  9 51  1
#> 12  15 60 70 38 27  1
#> 13  17 50 47 32 81  1
#> 14  58 14 25 68 61  1
#> 15  43 30 55 11 71  1
#> 16  19 83 32 57 52  1
#> 17  28 67 51 10 51  1
#> 18  56 62 54  6 35  1
#> 19  77 12 39 58 39  1
#> 20  70 10 45 64 37  1
#> 22  82 17 38 42 76  0
#> 23  71 88 38 42 76  0
#> 24  82 88 38 42 76  0
#> 25  71 17 77 42 76  0
#> 26  82 17 77 42 76  0
#> 27  71 88 77 42 76  0
#> 28  82 88 77 42 76  0
#> 29  71 17 38 73 76  0
#> 30  82 17 38 73 76  0
#> 31  71 88 38 73 76  0
#> 32  82 88 38 73 76  0
#> 33  71 17 77 73 76  0
#> 34  82 17 77 73 76  0
#> 35  71 88 77 73 76  0
#> 36  82 88 77 73 76  0
#> 37  71 17 38 42 86  0
#> 38  82 17 38 42 86  0
#> 39  71 88 38 42 86  0
#> 40  82 88 38 42 86  0
#> 41  71 17 77 42 86  0
#> 42  82 17 77 42 86  0
#> 43  71 88 77 42 86  0
#> 44  82 88 77 42 86  0
#> 45  71 17 38 73 86  0
#> 46  82 17 38 73 86  0
#> 47  71 88 38 73 86  0
#> 48  82 88 38 73 86  0
#> 49  71 17 77 73 86  0
#> 50  82 17 77 73 86  0
#> 51  71 88 77 73 86  0
#> 52  82 88 77 73 86  0
#> 54  82 37 44 56 70  0
#> 55   8 88 44 56 70  0
#> 56  82 88 44 56 70  0
#> 57   8 37 77 56 70  0
#> 58  82 37 77 56 70  0
#> 59   8 88 77 56 70  0
#> 60  82 88 77 56 70  0
#> 61   8 37 44 73 70  0
#> 62  82 37 44 73 70  0
#> 63   8 88 44 73 70  0
#> 64  82 88 44 73 70  0
#> 65   8 37 77 73 70  0
#> 66  82 37 77 73 70  0
#> 67   8 88 77 73 70  0
#> 68  82 88 77 73 70  0
#> 69   8 37 44 56 86  0
#> 70  82 37 44 56 86  0
#> 71   8 88 44 56 86  0
#> 72  82 88 44 56 86  0
#> 73   8 37 77 56 86  0
#> 74  82 37 77 56 86  0
#> 75   8 88 77 56 86  0
#> 76  82 88 77 56 86  0
#> 77   8 37 44 73 86  0
#> 78  82 37 44 73 86  0
#> 79   8 88 44 73 86  0
#> 80  82 88 44 73 86  0
#> 81   8 37 77 73 86  0
#> 82  82 37 77 73 86  0
#> 83   8 88 77 73 86  0
#> 86  82 28 72 27 59  0
#> 87  21 88 72 27 59  0
#> 88  82 88 72 27 59  0
#> 89  21 28 77 27 59  0
#> 90  82 28 77 27 59  0
#> 91  21 88 77 27 59  0
#> 92  82 88 77 27 59  0
#> 93  21 28 72 73 59  0
#> 94  82 28 72 73 59  0
#> 95  21 88 72 73 59  0
#> 96  82 88 72 73 59  0
#> 97  21 28 77 73 59  0
#> 98  82 28 77 73 59  0
#> 99  21 88 77 73 59  0
#> 100 82 88 77 73 59  0
#> 101 21 28 72 27 86  0
#> 102 82 28 72 27 86  0
#> 103 21 88 72 27 86  0
#> 104 82 88 72 27 86  0
#> 105 21 28 77 27 86  0
#> 106 82 28 77 27 86  0
#> 107 21 88 77 27 86  0
#> 108 82 88 77 27 86  0
#> 109 21 28 72 73 86  0
#> 110 82 28 72 73 86  0
#> 111 21 88 72 73 86  0
#> 112 82 88 72 73 86  0
#> 113 21 28 77 73 86  0
#> 114 82 28 77 73 86  0
#> 115 21 88 77 73 86  0
#> 118 82 13 31 54 32  0
#> 119 32 88 31 54 32  0
#> 120 82 88 31 54 32  0
#> 121 32 13 77 54 32  0
#> 122 82 13 77 54 32  0
#> 123 32 88 77 54 32  0
#> 124 82 88 77 54 32  0
#> 125 32 13 31 73 32  0
#> 126 82 13 31 73 32  0
#> 127 32 88 31 73 32  0
#> 128 82 88 31 73 32  0
#> 129 32 13 77 73 32  0
#> 130 82 13 77 73 32  0
#> 131 32 88 77 73 32  0
#> 132 82 88 77 73 32  0
#> 133 32 13 31 54 86  0
#> 134 82 13 31 54 86  0
#> 135 32 88 31 54 86  0
#> 136 82 88 31 54 86  0
#> 137 32 13 77 54 86  0
#> 138 82 13 77 54 86  0
#> 139 32 88 77 54 86  0
#> 140 82 88 77 54 86  0
#> 141 32 13 31 73 86  0
#> 142 82 13 31 73 86  0
#> 143 32 88 31 73 86  0
#> 144 82 88 31 73 86  0
#> 145 32 13 77 73 86  0
#> 146 82 13 77 73 86  0
#> 147 32 88 77 73 86  0
#> 150 82 68 24 14 55  0
#> 151 61 88 24 14 55  0
#> 152 82 88 24 14 55  0
#> 153 61 68 77 14 55  0
#> 154 82 68 77 14 55  0
#> 155 61 88 77 14 55  0
#> 156 82 88 77 14 55  0
#> 157 61 68 24 73 55  0
#> 158 82 68 24 73 55  0
#> 159 61 88 24 73 55  0
#> 160 82 88 24 73 55  0
#> 161 61 68 77 73 55  0
#> 162 82 68 77 73 55  0
#> 163 61 88 77 73 55  0
#> 164 82 88 77 73 55  0
#> 165 61 68 24 14 86  0
#> 166 82 68 24 14 86  0
#> 167 61 88 24 14 86  0
#> 168 82 88 24 14 86  0
#> 169 61 68 77 14 86  0
#> 170 82 68 77 14 86  0
#> 171 61 88 77 14 86  0
#> 172 82 88 77 14 86  0
#> 173 61 68 24 73 86  0
#> 174 82 68 24 73 86  0
#> 175 61 88 24 73 86  0
#> 176 82 88 24 73 86  0
#> 177 61 68 77 73 86  0
#> 178 82 68 77 73 86  0
#> 179 61 88 77 73 86  0
#> 182 82 71 66 30 14  0
#> 183 43 88 66 30 14  0
#> 184 82 88 66 30 14  0
#> 185 43 71 77 30 14  0
#> 186 82 71 77 30 14  0
#> 187 43 88 77 30 14  0
#> 188 82 88 77 30 14  0
#> 189 43 71 66 73 14  0
#> 190 82 71 66 73 14  0
#> 191 43 88 66 73 14  0
#> 192 82 88 66 73 14  0
#> 193 43 71 77 73 14  0
#> 194 82 71 77 73 14  0
#> 195 43 88 77 73 14  0
#> 196 82 88 77 73 14  0
#> 197 43 71 66 30 86  0
#> 198 82 71 66 30 86  0
#> 199 43 88 66 30 86  0
#> 200 82 88 66 30 86  0
#> 201 43 71 77 30 86  0
#> 202 82 71 77 30 86  0
#> 203 43 88 77 30 86  0
#> 204 82 88 77 30 86  0
#> 205 43 71 66 73 86  0
#> 206 82 71 66 73 86  0
#> 207 43 88 66 73 86  0
#> 208 82 88 66 73 86  0
#> 209 43 71 77 73 86  0
#> 210 82 71 77 73 86  0
#> 211 43 88 77 73 86  0
#> 214 82 52 38 22 39  0
#> 215 14 88 38 22 39  0
#> 216 82 88 38 22 39  0
#> 217 14 52 77 22 39  0
#> 218 82 52 77 22 39  0
#> 219 14 88 77 22 39  0
#> 220 82 88 77 22 39  0
#> 221 14 52 38 73 39  0
#> 222 82 52 38 73 39  0
#> 223 14 88 38 73 39  0
#> 224 82 88 38 73 39  0
#> 225 14 52 77 73 39  0
#> 226 82 52 77 73 39  0
#> 227 14 88 77 73 39  0
#> 228 82 88 77 73 39  0
#> 229 14 52 38 22 86  0
#> 230 82 52 38 22 86  0
#> 231 14 88 38 22 86  0
#> 232 82 88 38 22 86  0
#> 233 14 52 77 22 86  0
#> 234 82 52 77 22 86  0
#> 235 14 88 77 22 86  0
#> 236 82 88 77 22 86  0
#> 237 14 52 38 73 86  0
#> 238 82 52 38 73 86  0
#> 239 14 88 38 73 86  0
#> 241 14 52 77 73 86  0
#> 242 82 52 77 73 86  0
#> 243 14 88 77 73 86  0
#> 246 82 53 55 66 30  0
#> 247  8 88 55 66 30  0
#> 248 82 88 55 66 30  0
#> 249  8 53 77 66 30  0
#> 250 82 53 77 66 30  0
#> 251  8 88 77 66 30  0
#> 252 82 88 77 66 30  0
#> 253  8 53 55 73 30  0
#> 254 82 53 55 73 30  0
#> 255  8 88 55 73 30  0
#> 256 82 88 55 73 30  0
#> 257  8 53 77 73 30  0
#> 258 82 53 77 73 30  0
#> 259  8 88 77 73 30  0
#> 260 82 88 77 73 30  0
#> 261  8 53 55 66 86  0
#> 262 82 53 55 66 86  0
#> 263  8 88 55 66 86  0
#> 264 82 88 55 66 86  0
#> 265  8 53 77 66 86  0
#> 266 82 53 77 66 86  0
#> 267  8 88 77 66 86  0
#> 268 82 88 77 66 86  0
#> 269  8 53 55 73 86  0
#> 270 82 53 55 73 86  0
#> 271  8 88 55 73 86  0
#> 272 82 88 55 73 86  0
#> 273  8 53 77 73 86  0
#> 274 82 53 77 73 86  0
#> 278 82 52 28 10 44  0
#> 279 36 88 28 10 44  0
#> 280 82 88 28 10 44  0
#> 281 36 52 77 10 44  0
#> 282 82 52 77 10 44  0
#> 283 36 88 77 10 44  0
#> 284 82 88 77 10 44  0
#> 285 36 52 28 73 44  0
#> 286 82 52 28 73 44  0
#> 287 36 88 28 73 44  0
#> 288 82 88 28 73 44  0
#> 289 36 52 77 73 44  0
#> 290 82 52 77 73 44  0
#> 291 36 88 77 73 44  0
#> 292 82 88 77 73 44  0
#> 293 36 52 28 10 86  0
#> 294 82 52 28 10 86  0
#> 295 36 88 28 10 86  0
#> 296 82 88 28 10 86  0
#> 297 36 52 77 10 86  0
#> 298 82 52 77 10 86  0
#> 299 36 88 77 10 86  0
#> 300 82 88 77 10 86  0
#> 301 36 52 28 73 86  0
#> 302 82 52 28 73 86  0
#> 303 36 88 28 73 86  0
#> 304 82 88 28 73 86  0
#> 305 36 52 77 73 86  0
#> 307 36 88 77 73 86  0
#> 310 82 78 50 36 55  0
#> 311 12 88 50 36 55  0
#> 312 82 88 50 36 55  0
#> 313 12 78 77 36 55  0
#> 314 82 78 77 36 55  0
#> 315 12 88 77 36 55  0
#> 316 82 88 77 36 55  0
#> 317 12 78 50 73 55  0
#> 318 82 78 50 73 55  0
#> 319 12 88 50 73 55  0
#> 320 82 88 50 73 55  0
#> 321 12 78 77 73 55  0
#> 322 82 78 77 73 55  0
#> 323 12 88 77 73 55  0
#> 325 12 78 50 36 86  0
#> 326 82 78 50 36 86  0
#> 327 12 88 50 36 86  0
#> 328 82 88 50 36 86  0
#> 329 12 78 77 36 86  0
#> 330 82 78 77 36 86  0
#> 331 12 88 77 36 86  0
#> 332 82 88 77 36 86  0
#> 333 12 78 50 73 86  0
#> 334 82 78 50 73 86  0
#> 335 12 88 50 73 86  0
#> 336 82 88 50 73 86  0
#> 337 12 78 77 73 86  0
#> 338 82 78 77 73 86  0
#> 339 12 88 77 73 86  0
#> 342 82 51 35  9 51  0
#> 343 72 88 35  9 51  0
#> 344 82 88 35  9 51  0
#> 345 72 51 77  9 51  0
#> 346 82 51 77  9 51  0
#> 347 72 88 77  9 51  0
#> 348 82 88 77  9 51  0
#> 349 72 51 35 73 51  0
#> 350 82 51 35 73 51  0
#> 351 72 88 35 73 51  0
#> 352 82 88 35 73 51  0
#> 353 72 51 77 73 51  0
#> 354 82 51 77 73 51  0
#> 355 72 88 77 73 51  0
#> 356 82 88 77 73 51  0
#> 357 72 51 35  9 86  0
#> 358 82 51 35  9 86  0
#> 359 72 88 35  9 86  0
#> 360 82 88 35  9 86  0
#> 361 72 51 77  9 86  0
#> 362 82 51 77  9 86  0
#> 363 72 88 77  9 86  0
#> 364 82 88 77  9 86  0
#> 365 72 51 35 73 86  0
#> 366 82 51 35 73 86  0
#> 367 72 88 35 73 86  0
#> 368 82 88 35 73 86  0
#> 369 72 51 77 73 86  0
#> 370 82 51 77 73 86  0
#> 371 72 88 77 73 86  0
#> 374 82 60 70 38 27  0
#> 375 15 88 70 38 27  0
#> 376 82 88 70 38 27  0
#> 377 15 60 77 38 27  0
#> 378 82 60 77 38 27  0
#> 379 15 88 77 38 27  0
#> 380 82 88 77 38 27  0
#> 381 15 60 70 73 27  0
#> 382 82 60 70 73 27  0
#> 383 15 88 70 73 27  0
#> 384 82 88 70 73 27  0
#> 385 15 60 77 73 27  0
#> 386 82 60 77 73 27  0
#> 387 15 88 77 73 27  0
#> 388 82 88 77 73 27  0
#> 389 15 60 70 38 86  0
#> 390 82 60 70 38 86  0
#> 391 15 88 70 38 86  0
#> 392 82 88 70 38 86  0
#> 393 15 60 77 38 86  0
#> 394 82 60 77 38 86  0
#> 395 15 88 77 38 86  0
#> 396 82 88 77 38 86  0
#> 397 15 60 70 73 86  0
#> 398 82 60 70 73 86  0
#> 399 15 88 70 73 86  0
#> 400 82 88 70 73 86  0
#> 401 15 60 77 73 86  0
#> 402 82 60 77 73 86  0
#> 403 15 88 77 73 86  0
#> 406 82 50 47 32 81  0
#> 407 17 88 47 32 81  0
#> 408 82 88 47 32 81  0
#> 409 17 50 77 32 81  0
#> 410 82 50 77 32 81  0
#> 411 17 88 77 32 81  0
#> 412 82 88 77 32 81  0
#> 413 17 50 47 73 81  0
#> 414 82 50 47 73 81  0
#> 415 17 88 47 73 81  0
#> 416 82 88 47 73 81  0
#> 417 17 50 77 73 81  0
#> 418 82 50 77 73 81  0
#> 419 17 88 77 73 81  0
#> 420 82 88 77 73 81  0
#> 421 17 50 47 32 86  0
#> 422 82 50 47 32 86  0
#> 423 17 88 47 32 86  0
#> 424 82 88 47 32 86  0
#> 425 17 50 77 32 86  0
#> 426 82 50 77 32 86  0
#> 427 17 88 77 32 86  0
#> 428 82 88 77 32 86  0
#> 429 17 50 47 73 86  0
#> 430 82 50 47 73 86  0
#> 431 17 88 47 73 86  0
#> 432 82 88 47 73 86  0
#> 433 17 50 77 73 86  0
#> 434 82 50 77 73 86  0
#> 435 17 88 77 73 86  0
#> 438 82 14 25 68 61  0
#> 439 58 88 25 68 61  0
#> 440 82 88 25 68 61  0
#> 441 58 14 77 68 61  0
#> 442 82 14 77 68 61  0
#> 443 58 88 77 68 61  0
#> 444 82 88 77 68 61  0
#> 445 58 14 25 73 61  0
#> 446 82 14 25 73 61  0
#> 447 58 88 25 73 61  0
#> 448 82 88 25 73 61  0
#> 449 58 14 77 73 61  0
#> 450 82 14 77 73 61  0
#> 451 58 88 77 73 61  0
#> 452 82 88 77 73 61  0
#> 453 58 14 25 68 86  0
#> 454 82 14 25 68 86  0
#> 455 58 88 25 68 86  0
#> 456 82 88 25 68 86  0
#> 457 58 14 77 68 86  0
#> 458 82 14 77 68 86  0
#> 459 58 88 77 68 86  0
#> 460 82 88 77 68 86  0
#> 461 58 14 25 73 86  0
#> 462 82 14 25 73 86  0
#> 463 58 88 25 73 86  0
#> 464 82 88 25 73 86  0
#> 465 58 14 77 73 86  0
#> 466 82 14 77 73 86  0
#> 467 58 88 77 73 86  0
#> 470 82 30 55 11 71  0
#> 471 43 88 55 11 71  0
#> 472 82 88 55 11 71  0
#> 473 43 30 77 11 71  0
#> 474 82 30 77 11 71  0
#> 475 43 88 77 11 71  0
#> 476 82 88 77 11 71  0
#> 477 43 30 55 73 71  0
#> 478 82 30 55 73 71  0
#> 479 43 88 55 73 71  0
#> 480 82 88 55 73 71  0
#> 481 43 30 77 73 71  0
#> 482 82 30 77 73 71  0
#> 483 43 88 77 73 71  0
#> 484 82 88 77 73 71  0
#> 485 43 30 55 11 86  0
#> 486 82 30 55 11 86  0
#> 487 43 88 55 11 86  0
#> 488 82 88 55 11 86  0
#> 489 43 30 77 11 86  0
#> 490 82 30 77 11 86  0
#> 491 43 88 77 11 86  0
#> 492 82 88 77 11 86  0
#> 493 43 30 55 73 86  0
#> 494 82 30 55 73 86  0
#> 495 43 88 55 73 86  0
#> 497 43 30 77 73 86  0
#> 498 82 30 77 73 86  0
#> 502 82 83 32 57 52  0
#> 503 19 88 32 57 52  0
#> 504 82 88 32 57 52  0
#> 505 19 83 77 57 52  0
#> 506 82 83 77 57 52  0
#> 507 19 88 77 57 52  0
#> 508 82 88 77 57 52  0
#> 509 19 83 32 73 52  0
#> 510 82 83 32 73 52  0
#> 511 19 88 32 73 52  0
#> 512 82 88 32 73 52  0
#> 513 19 83 77 73 52  0
#> 514 82 83 77 73 52  0
#> 515 19 88 77 73 52  0
#> 516 82 88 77 73 52  0
#> 517 19 83 32 57 86  0
#> 518 82 83 32 57 86  0
#> 519 19 88 32 57 86  0
#> 520 82 88 32 57 86  0
#> 521 19 83 77 57 86  0
#> 522 82 83 77 57 86  0
#> 523 19 88 77 57 86  0
#> 524 82 88 77 57 86  0
#> 525 19 83 32 73 86  0
#> 526 82 83 32 73 86  0
#> 527 19 88 32 73 86  0
#> 528 82 88 32 73 86  0
#> 529 19 83 77 73 86  0
#> 530 82 83 77 73 86  0
#> 531 19 88 77 73 86  0
#> 534 82 67 51 10 51  0
#> 535 28 88 51 10 51  0
#> 536 82 88 51 10 51  0
#> 537 28 67 77 10 51  0
#> 538 82 67 77 10 51  0
#> 539 28 88 77 10 51  0
#> 540 82 88 77 10 51  0
#> 541 28 67 51 73 51  0
#> 542 82 67 51 73 51  0
#> 543 28 88 51 73 51  0
#> 544 82 88 51 73 51  0
#> 545 28 67 77 73 51  0
#> 546 82 67 77 73 51  0
#> 547 28 88 77 73 51  0
#> 549 28 67 51 10 86  0
#> 550 82 67 51 10 86  0
#> 551 28 88 51 10 86  0
#> 552 82 88 51 10 86  0
#> 553 28 67 77 10 86  0
#> 554 82 67 77 10 86  0
#> 555 28 88 77 10 86  0
#> 557 28 67 51 73 86  0
#> 558 82 67 51 73 86  0
#> 559 28 88 51 73 86  0
#> 560 82 88 51 73 86  0
#> 561 28 67 77 73 86  0
#> 562 82 67 77 73 86  0
#> 563 28 88 77 73 86  0
#> 566 82 62 54  6 35  0
#> 567 56 88 54  6 35  0
#> 568 82 88 54  6 35  0
#> 569 56 62 77  6 35  0
#> 570 82 62 77  6 35  0
#> 571 56 88 77  6 35  0
#> 572 82 88 77  6 35  0
#> 573 56 62 54 73 35  0
#> 574 82 62 54 73 35  0
#> 575 56 88 54 73 35  0
#> 576 82 88 54 73 35  0
#> 577 56 62 77 73 35  0
#> 578 82 62 77 73 35  0
#> 579 56 88 77 73 35  0
#> 580 82 88 77 73 35  0
#> 581 56 62 54  6 86  0
#> 582 82 62 54  6 86  0
#> 583 56 88 54  6 86  0
#> 584 82 88 54  6 86  0
#> 585 56 62 77  6 86  0
#> 586 82 62 77  6 86  0
#> 587 56 88 77  6 86  0
#> 588 82 88 77  6 86  0
#> 589 56 62 54 73 86  0
#> 590 82 62 54 73 86  0
#> 591 56 88 54 73 86  0
#> 592 82 88 54 73 86  0
#> 593 56 62 77 73 86  0
#> 594 82 62 77 73 86  0
#> 595 56 88 77 73 86  0
#> 598 82 12 39 58 39  0
#> 599 77 88 39 58 39  0
#> 600 82 88 39 58 39  0
#> 601 77 12 77 58 39  0
#> 602 82 12 77 58 39  0
#> 603 77 88 77 58 39  0
#> 604 82 88 77 58 39  0
#> 605 77 12 39 73 39  0
#> 606 82 12 39 73 39  0
#> 607 77 88 39 73 39  0
#> 608 82 88 39 73 39  0
#> 609 77 12 77 73 39  0
#> 610 82 12 77 73 39  0
#> 611 77 88 77 73 39  0
#> 613 77 12 39 58 86  0
#> 614 82 12 39 58 86  0
#> 615 77 88 39 58 86  0
#> 616 82 88 39 58 86  0
#> 617 77 12 77 58 86  0
#> 618 82 12 77 58 86  0
#> 619 77 88 77 58 86  0
#> 620 82 88 77 58 86  0
#> 621 77 12 39 73 86  0
#> 622 82 12 39 73 86  0
#> 623 77 88 39 73 86  0
#> 624 82 88 39 73 86  0
#> 625 77 12 77 73 86  0
#> 626 82 12 77 73 86  0
#> 627 77 88 77 73 86  0
#> 630 82 10 45 64 37  0
#> 631 70 88 45 64 37  0
#> 632 82 88 45 64 37  0
#> 633 70 10 77 64 37  0
#> 634 82 10 77 64 37  0
#> 635 70 88 77 64 37  0
#> 636 82 88 77 64 37  0
#> 637 70 10 45 73 37  0
#> 638 82 10 45 73 37  0
#> 639 70 88 45 73 37  0
#> 640 82 88 45 73 37  0
#> 641 70 10 77 73 37  0
#> 642 82 10 77 73 37  0
#> 643 70 88 77 73 37  0
#> 644 82 88 77 73 37  0
#> 645 70 10 45 64 86  0
#> 646 82 10 45 64 86  0
#> 647 70 88 45 64 86  0
#> 648 82 88 45 64 86  0
#> 649 70 10 77 64 86  0
#> 650 82 10 77 64 86  0
#> 651 70 88 77 64 86  0
#> 652 82 88 77 64 86  0
#> 653 70 10 45 73 86  0
#> 654 82 10 45 73 86  0
#> 655 70 88 45 73 86  0
#> 656 82 88 45 73 86  0
#> 657 70 10 77 73 86  0
#> 658 82 10 77 73 86  0
#> 659 70 88 77 73 86  0