Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
#> Warning: Only 8 samples generated!
addRays(pts)
#>    z1 z2 z3 pt
#> 4   2 44 60  1
#> 1   8 44 26  1
#> 7  26 70 12  1
#> 5  39  8 73  1
#> 2  48 35  3  1
#> 6  58 43  2  1
#> 3  79 11 46  1
#> 21 84 44 60  0
#> 31  2 75 60  0
#> 41 84 75 60  0
#> 51  2 44 78  0
#> 61 84 44 78  0
#> 71  2 75 78  0
#> 8  84 75 78  0
#> 22 84 44 26  0
#> 32  8 75 26  0
#> 42 84 75 26  0
#> 52  8 44 78  0
#> 72  8 75 78  0
#> 23 84 70 12  0
#> 33 26 75 12  0
#> 43 84 75 12  0
#> 53 26 70 78  0
#> 63 84 70 78  0
#> 73 26 75 78  0
#> 24 84  8 73  0
#> 34 39 75 73  0
#> 44 84 75 73  0
#> 54 39  8 78  0
#> 64 84  8 78  0
#> 74 39 75 78  0
#> 25 84 35  3  0
#> 35 48 75  3  0
#> 45 84 75  3  0
#> 55 48 35 78  0
#> 65 84 35 78  0
#> 75 48 75 78  0
#> 26 84 43  2  0
#> 36 58 75  2  0
#> 46 84 75  2  0
#> 56 58 43 78  0
#> 66 84 43 78  0
#> 76 58 75 78  0
#> 27 84 11 46  0
#> 37 79 75 46  0
#> 47 84 75 46  0
#> 57 79 11 78  0
#> 67 84 11 78  0
#> 77 79 75 78  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 4   2 44 60  1
#> 1   8 44 26  1
#> 7  26 70 12  1
#> 5  39  8 73  1
#> 2  48 35  3  1
#> 6  58 43  2  1
#> 3  79 11 46  1
#> 21 84 44 60  0
#> 31  2  3 60  0
#> 41 84  3 60  0
#> 51  2 44 78  0
#> 61 84 44 78  0
#> 71  2  3 78  0
#> 8  84  3 78  0
#> 22 84 44 26  0
#> 32  8  3 26  0
#> 42 84  3 26  0
#> 52  8 44 78  0
#> 72  8  3 78  0
#> 23 84 70 12  0
#> 33 26  3 12  0
#> 43 84  3 12  0
#> 53 26 70 78  0
#> 63 84 70 78  0
#> 73 26  3 78  0
#> 24 84  8 73  0
#> 34 39  3 73  0
#> 44 84  3 73  0
#> 54 39  8 78  0
#> 64 84  8 78  0
#> 74 39  3 78  0
#> 25 84 35  3  0
#> 35 48  3  3  0
#> 45 84  3  3  0
#> 55 48 35 78  0
#> 65 84 35 78  0
#> 75 48  3 78  0
#> 26 84 43  2  0
#> 36 58  3  2  0
#> 46 84  3  2  0
#> 56 58 43 78  0
#> 66 84 43 78  0
#> 76 58  3 78  0
#> 27 84 11 46  0
#> 37 79  3 46  0
#> 47 84  3 46  0
#> 57 79 11 78  0
#> 67 84 11 78  0
#> 77 79  3 78  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 4   2 44  60  1
#> 1   8 44  26  1
#> 7  26 70  12  1
#> 5  39  8  73  1
#> 2  48 35   3  1
#> 6  58 43   2  1
#> 3  79 11  46  1
#> 21  0 44  60  0
#> 31  2  0  60  0
#> 41  0  0  60  0
#> 51  2 44 100  0
#> 61  0 44 100  0
#> 71  2  0 100  0
#> 8   0  0 100  0
#> 22  0 44  26  0
#> 32  8  0  26  0
#> 42  0  0  26  0
#> 52  8 44 100  0
#> 72  8  0 100  0
#> 23  0 70  12  0
#> 33 26  0  12  0
#> 43  0  0  12  0
#> 53 26 70 100  0
#> 63  0 70 100  0
#> 73 26  0 100  0
#> 24  0  8  73  0
#> 34 39  0  73  0
#> 44  0  0  73  0
#> 54 39  8 100  0
#> 64  0  8 100  0
#> 74 39  0 100  0
#> 25  0 35   3  0
#> 35 48  0   3  0
#> 45  0  0   3  0
#> 55 48 35 100  0
#> 65  0 35 100  0
#> 75 48  0 100  0
#> 26  0 43   2  0
#> 36 58  0   2  0
#> 46  0  0   2  0
#> 56 58 43 100  0
#> 66  0 43 100  0
#> 76 58  0 100  0
#> 27  0 11  46  0
#> 37 79  0  46  0
#> 47  0  0  46  0
#> 57 79 11 100  0
#> 67  0 11 100  0
#> 77 79  0 100  0
pts <- genSample(5,20)[,1:5]
#> Warning: Only 17 samples generated!
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   47  3 58 47 42  1
#> 2   78 44 38 13 61  1
#> 3   31 21 16 50 48  1
#> 4   37 48 92 36 34  1
#> 5   38 20 56 45 86  1
#> 6   49 21 68 59 15  1
#> 7   82 54 26 23 59  1
#> 8   37  5 58 45 60  1
#> 9   36 48 18 20 66  1
#> 10  56 35 16 22 64  1
#> 11  68 16 77 35 47  1
#> 12  24 11 53 40 42  1
#> 13  36 18 71 38 75  1
#> 14  32 83 41 30 26  1
#> 15  58 12 49 35 77  1
#> 16  56 53  7 35 33  1
#> 17  28 17 43 52 77  1
#> 19  87  3 58 47 42  0
#> 20  47 88 58 47 42  0
#> 21  87 88 58 47 42  0
#> 22  47  3 97 47 42  0
#> 23  87  3 97 47 42  0
#> 24  47 88 97 47 42  0
#> 25  87 88 97 47 42  0
#> 26  47  3 58 64 42  0
#> 27  87  3 58 64 42  0
#> 28  47 88 58 64 42  0
#> 29  87 88 58 64 42  0
#> 30  47  3 97 64 42  0
#> 31  87  3 97 64 42  0
#> 32  47 88 97 64 42  0
#> 33  87 88 97 64 42  0
#> 34  47  3 58 47 91  0
#> 35  87  3 58 47 91  0
#> 36  47 88 58 47 91  0
#> 37  87 88 58 47 91  0
#> 38  47  3 97 47 91  0
#> 39  87  3 97 47 91  0
#> 40  47 88 97 47 91  0
#> 41  87 88 97 47 91  0
#> 42  47  3 58 64 91  0
#> 43  87  3 58 64 91  0
#> 44  47 88 58 64 91  0
#> 45  87 88 58 64 91  0
#> 46  47  3 97 64 91  0
#> 47  87  3 97 64 91  0
#> 48  47 88 97 64 91  0
#> 49  87 88 97 64 91  0
#> 51  87 44 38 13 61  0
#> 52  78 88 38 13 61  0
#> 53  87 88 38 13 61  0
#> 54  78 44 97 13 61  0
#> 55  87 44 97 13 61  0
#> 56  78 88 97 13 61  0
#> 57  87 88 97 13 61  0
#> 58  78 44 38 64 61  0
#> 59  87 44 38 64 61  0
#> 60  78 88 38 64 61  0
#> 61  87 88 38 64 61  0
#> 62  78 44 97 64 61  0
#> 63  87 44 97 64 61  0
#> 64  78 88 97 64 61  0
#> 65  87 88 97 64 61  0
#> 66  78 44 38 13 91  0
#> 67  87 44 38 13 91  0
#> 68  78 88 38 13 91  0
#> 69  87 88 38 13 91  0
#> 70  78 44 97 13 91  0
#> 71  87 44 97 13 91  0
#> 72  78 88 97 13 91  0
#> 73  87 88 97 13 91  0
#> 74  78 44 38 64 91  0
#> 75  87 44 38 64 91  0
#> 76  78 88 38 64 91  0
#> 77  87 88 38 64 91  0
#> 78  78 44 97 64 91  0
#> 79  87 44 97 64 91  0
#> 80  78 88 97 64 91  0
#> 83  87 21 16 50 48  0
#> 84  31 88 16 50 48  0
#> 85  87 88 16 50 48  0
#> 86  31 21 97 50 48  0
#> 87  87 21 97 50 48  0
#> 88  31 88 97 50 48  0
#> 89  87 88 97 50 48  0
#> 90  31 21 16 64 48  0
#> 91  87 21 16 64 48  0
#> 92  31 88 16 64 48  0
#> 93  87 88 16 64 48  0
#> 94  31 21 97 64 48  0
#> 95  87 21 97 64 48  0
#> 96  31 88 97 64 48  0
#> 97  87 88 97 64 48  0
#> 98  31 21 16 50 91  0
#> 99  87 21 16 50 91  0
#> 100 31 88 16 50 91  0
#> 101 87 88 16 50 91  0
#> 102 31 21 97 50 91  0
#> 103 87 21 97 50 91  0
#> 104 31 88 97 50 91  0
#> 105 87 88 97 50 91  0
#> 106 31 21 16 64 91  0
#> 107 87 21 16 64 91  0
#> 108 31 88 16 64 91  0
#> 109 87 88 16 64 91  0
#> 110 31 21 97 64 91  0
#> 111 87 21 97 64 91  0
#> 112 31 88 97 64 91  0
#> 115 87 48 92 36 34  0
#> 116 37 88 92 36 34  0
#> 117 87 88 92 36 34  0
#> 118 37 48 97 36 34  0
#> 119 87 48 97 36 34  0
#> 120 37 88 97 36 34  0
#> 121 87 88 97 36 34  0
#> 122 37 48 92 64 34  0
#> 123 87 48 92 64 34  0
#> 124 37 88 92 64 34  0
#> 125 87 88 92 64 34  0
#> 126 37 48 97 64 34  0
#> 127 87 48 97 64 34  0
#> 128 37 88 97 64 34  0
#> 129 87 88 97 64 34  0
#> 130 37 48 92 36 91  0
#> 131 87 48 92 36 91  0
#> 132 37 88 92 36 91  0
#> 133 87 88 92 36 91  0
#> 134 37 48 97 36 91  0
#> 135 87 48 97 36 91  0
#> 136 37 88 97 36 91  0
#> 137 87 88 97 36 91  0
#> 138 37 48 92 64 91  0
#> 139 87 48 92 64 91  0
#> 140 37 88 92 64 91  0
#> 141 87 88 92 64 91  0
#> 142 37 48 97 64 91  0
#> 143 87 48 97 64 91  0
#> 144 37 88 97 64 91  0
#> 147 87 20 56 45 86  0
#> 148 38 88 56 45 86  0
#> 149 87 88 56 45 86  0
#> 150 38 20 97 45 86  0
#> 151 87 20 97 45 86  0
#> 152 38 88 97 45 86  0
#> 153 87 88 97 45 86  0
#> 154 38 20 56 64 86  0
#> 155 87 20 56 64 86  0
#> 156 38 88 56 64 86  0
#> 157 87 88 56 64 86  0
#> 158 38 20 97 64 86  0
#> 159 87 20 97 64 86  0
#> 160 38 88 97 64 86  0
#> 161 87 88 97 64 86  0
#> 162 38 20 56 45 91  0
#> 163 87 20 56 45 91  0
#> 164 38 88 56 45 91  0
#> 165 87 88 56 45 91  0
#> 166 38 20 97 45 91  0
#> 167 87 20 97 45 91  0
#> 168 38 88 97 45 91  0
#> 169 87 88 97 45 91  0
#> 170 38 20 56 64 91  0
#> 171 87 20 56 64 91  0
#> 172 38 88 56 64 91  0
#> 173 87 88 56 64 91  0
#> 174 38 20 97 64 91  0
#> 175 87 20 97 64 91  0
#> 176 38 88 97 64 91  0
#> 179 87 21 68 59 15  0
#> 180 49 88 68 59 15  0
#> 181 87 88 68 59 15  0
#> 182 49 21 97 59 15  0
#> 183 87 21 97 59 15  0
#> 184 49 88 97 59 15  0
#> 185 87 88 97 59 15  0
#> 186 49 21 68 64 15  0
#> 187 87 21 68 64 15  0
#> 188 49 88 68 64 15  0
#> 189 87 88 68 64 15  0
#> 190 49 21 97 64 15  0
#> 191 87 21 97 64 15  0
#> 192 49 88 97 64 15  0
#> 193 87 88 97 64 15  0
#> 194 49 21 68 59 91  0
#> 195 87 21 68 59 91  0
#> 196 49 88 68 59 91  0
#> 197 87 88 68 59 91  0
#> 198 49 21 97 59 91  0
#> 199 87 21 97 59 91  0
#> 200 49 88 97 59 91  0
#> 201 87 88 97 59 91  0
#> 202 49 21 68 64 91  0
#> 203 87 21 68 64 91  0
#> 204 49 88 68 64 91  0
#> 205 87 88 68 64 91  0
#> 206 49 21 97 64 91  0
#> 208 49 88 97 64 91  0
#> 211 87 54 26 23 59  0
#> 212 82 88 26 23 59  0
#> 213 87 88 26 23 59  0
#> 214 82 54 97 23 59  0
#> 215 87 54 97 23 59  0
#> 216 82 88 97 23 59  0
#> 217 87 88 97 23 59  0
#> 218 82 54 26 64 59  0
#> 219 87 54 26 64 59  0
#> 220 82 88 26 64 59  0
#> 221 87 88 26 64 59  0
#> 222 82 54 97 64 59  0
#> 223 87 54 97 64 59  0
#> 224 82 88 97 64 59  0
#> 225 87 88 97 64 59  0
#> 226 82 54 26 23 91  0
#> 227 87 54 26 23 91  0
#> 228 82 88 26 23 91  0
#> 229 87 88 26 23 91  0
#> 230 82 54 97 23 91  0
#> 231 87 54 97 23 91  0
#> 232 82 88 97 23 91  0
#> 233 87 88 97 23 91  0
#> 234 82 54 26 64 91  0
#> 235 87 54 26 64 91  0
#> 236 82 88 26 64 91  0
#> 237 87 88 26 64 91  0
#> 238 82 54 97 64 91  0
#> 239 87 54 97 64 91  0
#> 240 82 88 97 64 91  0
#> 243 87  5 58 45 60  0
#> 244 37 88 58 45 60  0
#> 245 87 88 58 45 60  0
#> 246 37  5 97 45 60  0
#> 247 87  5 97 45 60  0
#> 248 37 88 97 45 60  0
#> 249 87 88 97 45 60  0
#> 250 37  5 58 64 60  0
#> 251 87  5 58 64 60  0
#> 252 37 88 58 64 60  0
#> 253 87 88 58 64 60  0
#> 254 37  5 97 64 60  0
#> 255 87  5 97 64 60  0
#> 256 37 88 97 64 60  0
#> 257 87 88 97 64 60  0
#> 258 37  5 58 45 91  0
#> 259 87  5 58 45 91  0
#> 260 37 88 58 45 91  0
#> 261 87 88 58 45 91  0
#> 262 37  5 97 45 91  0
#> 263 87  5 97 45 91  0
#> 264 37 88 97 45 91  0
#> 266 37  5 58 64 91  0
#> 267 87  5 58 64 91  0
#> 268 37 88 58 64 91  0
#> 270 37  5 97 64 91  0
#> 271 87  5 97 64 91  0
#> 275 87 48 18 20 66  0
#> 276 36 88 18 20 66  0
#> 277 87 88 18 20 66  0
#> 278 36 48 97 20 66  0
#> 279 87 48 97 20 66  0
#> 280 36 88 97 20 66  0
#> 281 87 88 97 20 66  0
#> 282 36 48 18 64 66  0
#> 283 87 48 18 64 66  0
#> 284 36 88 18 64 66  0
#> 285 87 88 18 64 66  0
#> 286 36 48 97 64 66  0
#> 287 87 48 97 64 66  0
#> 288 36 88 97 64 66  0
#> 289 87 88 97 64 66  0
#> 290 36 48 18 20 91  0
#> 291 87 48 18 20 91  0
#> 292 36 88 18 20 91  0
#> 293 87 88 18 20 91  0
#> 294 36 48 97 20 91  0
#> 295 87 48 97 20 91  0
#> 296 36 88 97 20 91  0
#> 297 87 88 97 20 91  0
#> 298 36 48 18 64 91  0
#> 299 87 48 18 64 91  0
#> 300 36 88 18 64 91  0
#> 301 87 88 18 64 91  0
#> 302 36 48 97 64 91  0
#> 304 36 88 97 64 91  0
#> 307 87 35 16 22 64  0
#> 308 56 88 16 22 64  0
#> 309 87 88 16 22 64  0
#> 310 56 35 97 22 64  0
#> 311 87 35 97 22 64  0
#> 312 56 88 97 22 64  0
#> 313 87 88 97 22 64  0
#> 314 56 35 16 64 64  0
#> 315 87 35 16 64 64  0
#> 316 56 88 16 64 64  0
#> 317 87 88 16 64 64  0
#> 318 56 35 97 64 64  0
#> 319 87 35 97 64 64  0
#> 320 56 88 97 64 64  0
#> 321 87 88 97 64 64  0
#> 322 56 35 16 22 91  0
#> 323 87 35 16 22 91  0
#> 324 56 88 16 22 91  0
#> 325 87 88 16 22 91  0
#> 326 56 35 97 22 91  0
#> 327 87 35 97 22 91  0
#> 328 56 88 97 22 91  0
#> 329 87 88 97 22 91  0
#> 330 56 35 16 64 91  0
#> 331 87 35 16 64 91  0
#> 332 56 88 16 64 91  0
#> 334 56 35 97 64 91  0
#> 335 87 35 97 64 91  0
#> 336 56 88 97 64 91  0
#> 339 87 16 77 35 47  0
#> 340 68 88 77 35 47  0
#> 341 87 88 77 35 47  0
#> 342 68 16 97 35 47  0
#> 343 87 16 97 35 47  0
#> 344 68 88 97 35 47  0
#> 345 87 88 97 35 47  0
#> 346 68 16 77 64 47  0
#> 347 87 16 77 64 47  0
#> 348 68 88 77 64 47  0
#> 349 87 88 77 64 47  0
#> 350 68 16 97 64 47  0
#> 351 87 16 97 64 47  0
#> 352 68 88 97 64 47  0
#> 353 87 88 97 64 47  0
#> 354 68 16 77 35 91  0
#> 355 87 16 77 35 91  0
#> 356 68 88 77 35 91  0
#> 357 87 88 77 35 91  0
#> 358 68 16 97 35 91  0
#> 359 87 16 97 35 91  0
#> 360 68 88 97 35 91  0
#> 361 87 88 97 35 91  0
#> 362 68 16 77 64 91  0
#> 363 87 16 77 64 91  0
#> 364 68 88 77 64 91  0
#> 365 87 88 77 64 91  0
#> 366 68 16 97 64 91  0
#> 367 87 16 97 64 91  0
#> 368 68 88 97 64 91  0
#> 371 87 11 53 40 42  0
#> 372 24 88 53 40 42  0
#> 373 87 88 53 40 42  0
#> 374 24 11 97 40 42  0
#> 375 87 11 97 40 42  0
#> 376 24 88 97 40 42  0
#> 377 87 88 97 40 42  0
#> 378 24 11 53 64 42  0
#> 379 87 11 53 64 42  0
#> 380 24 88 53 64 42  0
#> 381 87 88 53 64 42  0
#> 382 24 11 97 64 42  0
#> 383 87 11 97 64 42  0
#> 384 24 88 97 64 42  0
#> 386 24 11 53 40 91  0
#> 387 87 11 53 40 91  0
#> 388 24 88 53 40 91  0
#> 389 87 88 53 40 91  0
#> 390 24 11 97 40 91  0
#> 391 87 11 97 40 91  0
#> 392 24 88 97 40 91  0
#> 393 87 88 97 40 91  0
#> 394 24 11 53 64 91  0
#> 395 87 11 53 64 91  0
#> 396 24 88 53 64 91  0
#> 397 87 88 53 64 91  0
#> 398 24 11 97 64 91  0
#> 399 87 11 97 64 91  0
#> 400 24 88 97 64 91  0
#> 403 87 18 71 38 75  0
#> 404 36 88 71 38 75  0
#> 405 87 88 71 38 75  0
#> 406 36 18 97 38 75  0
#> 407 87 18 97 38 75  0
#> 408 36 88 97 38 75  0
#> 409 87 88 97 38 75  0
#> 410 36 18 71 64 75  0
#> 411 87 18 71 64 75  0
#> 412 36 88 71 64 75  0
#> 413 87 88 71 64 75  0
#> 414 36 18 97 64 75  0
#> 415 87 18 97 64 75  0
#> 416 36 88 97 64 75  0
#> 417 87 88 97 64 75  0
#> 418 36 18 71 38 91  0
#> 419 87 18 71 38 91  0
#> 420 36 88 71 38 91  0
#> 421 87 88 71 38 91  0
#> 422 36 18 97 38 91  0
#> 423 87 18 97 38 91  0
#> 424 36 88 97 38 91  0
#> 425 87 88 97 38 91  0
#> 426 36 18 71 64 91  0
#> 427 87 18 71 64 91  0
#> 428 36 88 71 64 91  0
#> 429 87 88 71 64 91  0
#> 430 36 18 97 64 91  0
#> 431 87 18 97 64 91  0
#> 435 87 83 41 30 26  0
#> 436 32 88 41 30 26  0
#> 437 87 88 41 30 26  0
#> 438 32 83 97 30 26  0
#> 439 87 83 97 30 26  0
#> 440 32 88 97 30 26  0
#> 441 87 88 97 30 26  0
#> 442 32 83 41 64 26  0
#> 443 87 83 41 64 26  0
#> 444 32 88 41 64 26  0
#> 445 87 88 41 64 26  0
#> 446 32 83 97 64 26  0
#> 447 87 83 97 64 26  0
#> 448 32 88 97 64 26  0
#> 449 87 88 97 64 26  0
#> 450 32 83 41 30 91  0
#> 451 87 83 41 30 91  0
#> 452 32 88 41 30 91  0
#> 453 87 88 41 30 91  0
#> 454 32 83 97 30 91  0
#> 455 87 83 97 30 91  0
#> 456 32 88 97 30 91  0
#> 457 87 88 97 30 91  0
#> 458 32 83 41 64 91  0
#> 459 87 83 41 64 91  0
#> 460 32 88 41 64 91  0
#> 461 87 88 41 64 91  0
#> 462 32 83 97 64 91  0
#> 463 87 83 97 64 91  0
#> 464 32 88 97 64 91  0
#> 467 87 12 49 35 77  0
#> 468 58 88 49 35 77  0
#> 469 87 88 49 35 77  0
#> 470 58 12 97 35 77  0
#> 471 87 12 97 35 77  0
#> 472 58 88 97 35 77  0
#> 473 87 88 97 35 77  0
#> 474 58 12 49 64 77  0
#> 475 87 12 49 64 77  0
#> 476 58 88 49 64 77  0
#> 477 87 88 49 64 77  0
#> 478 58 12 97 64 77  0
#> 479 87 12 97 64 77  0
#> 480 58 88 97 64 77  0
#> 481 87 88 97 64 77  0
#> 482 58 12 49 35 91  0
#> 483 87 12 49 35 91  0
#> 484 58 88 49 35 91  0
#> 485 87 88 49 35 91  0
#> 486 58 12 97 35 91  0
#> 487 87 12 97 35 91  0
#> 488 58 88 97 35 91  0
#> 490 58 12 49 64 91  0
#> 491 87 12 49 64 91  0
#> 492 58 88 49 64 91  0
#> 493 87 88 49 64 91  0
#> 494 58 12 97 64 91  0
#> 495 87 12 97 64 91  0
#> 496 58 88 97 64 91  0
#> 499 87 53  7 35 33  0
#> 500 56 88  7 35 33  0
#> 501 87 88  7 35 33  0
#> 502 56 53 97 35 33  0
#> 503 87 53 97 35 33  0
#> 504 56 88 97 35 33  0
#> 505 87 88 97 35 33  0
#> 506 56 53  7 64 33  0
#> 507 87 53  7 64 33  0
#> 508 56 88  7 64 33  0
#> 509 87 88  7 64 33  0
#> 510 56 53 97 64 33  0
#> 511 87 53 97 64 33  0
#> 512 56 88 97 64 33  0
#> 513 87 88 97 64 33  0
#> 514 56 53  7 35 91  0
#> 515 87 53  7 35 91  0
#> 516 56 88  7 35 91  0
#> 517 87 88  7 35 91  0
#> 518 56 53 97 35 91  0
#> 519 87 53 97 35 91  0
#> 520 56 88 97 35 91  0
#> 522 56 53  7 64 91  0
#> 523 87 53  7 64 91  0
#> 524 56 88  7 64 91  0
#> 525 87 88  7 64 91  0
#> 526 56 53 97 64 91  0
#> 527 87 53 97 64 91  0
#> 531 87 17 43 52 77  0
#> 532 28 88 43 52 77  0
#> 533 87 88 43 52 77  0
#> 534 28 17 97 52 77  0
#> 535 87 17 97 52 77  0
#> 536 28 88 97 52 77  0
#> 537 87 88 97 52 77  0
#> 538 28 17 43 64 77  0
#> 539 87 17 43 64 77  0
#> 540 28 88 43 64 77  0
#> 541 87 88 43 64 77  0
#> 542 28 17 97 64 77  0
#> 543 87 17 97 64 77  0
#> 544 28 88 97 64 77  0
#> 546 28 17 43 52 91  0
#> 547 87 17 43 52 91  0
#> 548 28 88 43 52 91  0
#> 549 87 88 43 52 91  0
#> 550 28 17 97 52 91  0
#> 551 87 17 97 52 91  0
#> 552 28 88 97 52 91  0
#> 553 87 88 97 52 91  0
#> 554 28 17 43 64 91  0
#> 555 87 17 43 64 91  0
#> 556 28 88 43 64 91  0
#> 557 87 88 43 64 91  0
#> 558 28 17 97 64 91  0
#> 559 87 17 97 64 91  0
#> 560 28 88 97 64 91  0