Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 10 18 16 36  1
#> 2  39  4 37  1
#> 3  51  2 58  1
#> 5  56 39  3  1
#> 7  65 11 25  1
#> 6  77 19 23  1
#> 21 82 16 36  0
#> 31 18 44 36  0
#> 4  82 44 36  0
#> 51 18 16 63  0
#> 61 82 16 63  0
#> 71 18 44 63  0
#> 8  82 44 63  0
#> 22 82  4 37  0
#> 32 39 44 37  0
#> 41 82 44 37  0
#> 52 39  4 63  0
#> 62 82  4 63  0
#> 72 39 44 63  0
#> 23 82  2 58  0
#> 33 51 44 58  0
#> 42 82 44 58  0
#> 53 51  2 63  0
#> 63 82  2 63  0
#> 73 51 44 63  0
#> 24 82 39  3  0
#> 34 56 44  3  0
#> 43 82 44  3  0
#> 54 56 39 63  0
#> 64 82 39 63  0
#> 74 56 44 63  0
#> 25 82 11 25  0
#> 35 65 44 25  0
#> 44 82 44 25  0
#> 55 65 11 63  0
#> 65 82 11 63  0
#> 75 65 44 63  0
#> 26 82 19 23  0
#> 36 77 44 23  0
#> 45 82 44 23  0
#> 56 77 19 63  0
#> 66 82 19 63  0
#> 76 77 44 63  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 10 18 16 36  1
#> 2  39  4 37  1
#> 3  51  2 58  1
#> 5  56 39  3  1
#> 7  65 11 25  1
#> 6  77 19 23  1
#> 21 82 16 36  0
#> 31 18 -3 36  0
#> 4  82 -3 36  0
#> 51 18 16 63  0
#> 61 82 16 63  0
#> 71 18 -3 63  0
#> 8  82 -3 63  0
#> 22 82  4 37  0
#> 32 39 -3 37  0
#> 41 82 -3 37  0
#> 52 39  4 63  0
#> 62 82  4 63  0
#> 72 39 -3 63  0
#> 23 82  2 58  0
#> 33 51 -3 58  0
#> 42 82 -3 58  0
#> 53 51  2 63  0
#> 63 82  2 63  0
#> 73 51 -3 63  0
#> 24 82 39  3  0
#> 34 56 -3  3  0
#> 43 82 -3  3  0
#> 54 56 39 63  0
#> 64 82 39 63  0
#> 74 56 -3 63  0
#> 25 82 11 25  0
#> 35 65 -3 25  0
#> 44 82 -3 25  0
#> 55 65 11 63  0
#> 65 82 11 63  0
#> 75 65 -3 63  0
#> 26 82 19 23  0
#> 36 77 -3 23  0
#> 45 82 -3 23  0
#> 56 77 19 63  0
#> 66 82 19 63  0
#> 76 77 -3 63  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 10 18 16  36  1
#> 2  39  4  37  1
#> 3  51  2  58  1
#> 5  56 39   3  1
#> 7  65 11  25  1
#> 6  77 19  23  1
#> 21  0 16  36  0
#> 31 18  0  36  0
#> 4   0  0  36  0
#> 51 18 16 100  0
#> 61  0 16 100  0
#> 71 18  0 100  0
#> 8   0  0 100  0
#> 22  0  4  37  0
#> 32 39  0  37  0
#> 41  0  0  37  0
#> 52 39  4 100  0
#> 62  0  4 100  0
#> 72 39  0 100  0
#> 23  0  2  58  0
#> 33 51  0  58  0
#> 42  0  0  58  0
#> 53 51  2 100  0
#> 63  0  2 100  0
#> 73 51  0 100  0
#> 24  0 39   3  0
#> 34 56  0   3  0
#> 43  0  0   3  0
#> 54 56 39 100  0
#> 64  0 39 100  0
#> 74 56  0 100  0
#> 25  0 11  25  0
#> 35 65  0  25  0
#> 44  0  0  25  0
#> 55 65 11 100  0
#> 65  0 11 100  0
#> 75 65  0 100  0
#> 26  0 19  23  0
#> 36 77  0  23  0
#> 45  0  0  23  0
#> 56 77 19 100  0
#> 66  0 19 100  0
#> 76 77  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   72 26 18 48 69  1
#> 2   66 77 45 52 12  1
#> 3   21 35 16 39 47  1
#> 4   27 44 44 17 24  1
#> 5    9 67 54 38 33  1
#> 6   27 31 55 23 78  1
#> 7   58 76 18 25 45  1
#> 8   62 84 50 29 24  1
#> 9   42 24 18 30 64  1
#> 10  44 53  4 65 55  1
#> 11  73 53  9 35 54  1
#> 12  38 74 44 61 11  1
#> 13  21 26 26 49 31  1
#> 14  46 28 56 30 12  1
#> 15  30 28 33 40 84  1
#> 16  17 23 61 45 71  1
#> 17  31 70 63 43 12  1
#> 18  73 67 17 61 29  1
#> 19  78 49 32 53 14  1
#> 20  30 37 77 24 71  1
#> 22  83 26 18 48 69  0
#> 23  72 89 18 48 69  0
#> 24  83 89 18 48 69  0
#> 25  72 26 82 48 69  0
#> 26  83 26 82 48 69  0
#> 27  72 89 82 48 69  0
#> 28  83 89 82 48 69  0
#> 29  72 26 18 70 69  0
#> 30  83 26 18 70 69  0
#> 31  72 89 18 70 69  0
#> 32  83 89 18 70 69  0
#> 33  72 26 82 70 69  0
#> 34  83 26 82 70 69  0
#> 35  72 89 82 70 69  0
#> 36  83 89 82 70 69  0
#> 37  72 26 18 48 89  0
#> 38  83 26 18 48 89  0
#> 39  72 89 18 48 89  0
#> 40  83 89 18 48 89  0
#> 41  72 26 82 48 89  0
#> 42  83 26 82 48 89  0
#> 43  72 89 82 48 89  0
#> 44  83 89 82 48 89  0
#> 45  72 26 18 70 89  0
#> 46  83 26 18 70 89  0
#> 47  72 89 18 70 89  0
#> 48  83 89 18 70 89  0
#> 49  72 26 82 70 89  0
#> 50  83 26 82 70 89  0
#> 51  72 89 82 70 89  0
#> 52  83 89 82 70 89  0
#> 54  83 77 45 52 12  0
#> 55  66 89 45 52 12  0
#> 56  83 89 45 52 12  0
#> 57  66 77 82 52 12  0
#> 58  83 77 82 52 12  0
#> 59  66 89 82 52 12  0
#> 60  83 89 82 52 12  0
#> 61  66 77 45 70 12  0
#> 62  83 77 45 70 12  0
#> 63  66 89 45 70 12  0
#> 64  83 89 45 70 12  0
#> 65  66 77 82 70 12  0
#> 66  83 77 82 70 12  0
#> 67  66 89 82 70 12  0
#> 68  83 89 82 70 12  0
#> 69  66 77 45 52 89  0
#> 70  83 77 45 52 89  0
#> 71  66 89 45 52 89  0
#> 72  83 89 45 52 89  0
#> 73  66 77 82 52 89  0
#> 74  83 77 82 52 89  0
#> 75  66 89 82 52 89  0
#> 76  83 89 82 52 89  0
#> 77  66 77 45 70 89  0
#> 78  83 77 45 70 89  0
#> 79  66 89 45 70 89  0
#> 80  83 89 45 70 89  0
#> 81  66 77 82 70 89  0
#> 82  83 77 82 70 89  0
#> 83  66 89 82 70 89  0
#> 86  83 35 16 39 47  0
#> 87  21 89 16 39 47  0
#> 88  83 89 16 39 47  0
#> 89  21 35 82 39 47  0
#> 90  83 35 82 39 47  0
#> 91  21 89 82 39 47  0
#> 92  83 89 82 39 47  0
#> 93  21 35 16 70 47  0
#> 94  83 35 16 70 47  0
#> 95  21 89 16 70 47  0
#> 96  83 89 16 70 47  0
#> 97  21 35 82 70 47  0
#> 98  83 35 82 70 47  0
#> 99  21 89 82 70 47  0
#> 100 83 89 82 70 47  0
#> 101 21 35 16 39 89  0
#> 102 83 35 16 39 89  0
#> 103 21 89 16 39 89  0
#> 104 83 89 16 39 89  0
#> 105 21 35 82 39 89  0
#> 106 83 35 82 39 89  0
#> 107 21 89 82 39 89  0
#> 108 83 89 82 39 89  0
#> 109 21 35 16 70 89  0
#> 110 83 35 16 70 89  0
#> 111 21 89 16 70 89  0
#> 112 83 89 16 70 89  0
#> 113 21 35 82 70 89  0
#> 114 83 35 82 70 89  0
#> 115 21 89 82 70 89  0
#> 118 83 44 44 17 24  0
#> 119 27 89 44 17 24  0
#> 120 83 89 44 17 24  0
#> 121 27 44 82 17 24  0
#> 122 83 44 82 17 24  0
#> 123 27 89 82 17 24  0
#> 124 83 89 82 17 24  0
#> 125 27 44 44 70 24  0
#> 126 83 44 44 70 24  0
#> 127 27 89 44 70 24  0
#> 128 83 89 44 70 24  0
#> 129 27 44 82 70 24  0
#> 130 83 44 82 70 24  0
#> 131 27 89 82 70 24  0
#> 132 83 89 82 70 24  0
#> 133 27 44 44 17 89  0
#> 134 83 44 44 17 89  0
#> 135 27 89 44 17 89  0
#> 136 83 89 44 17 89  0
#> 137 27 44 82 17 89  0
#> 138 83 44 82 17 89  0
#> 139 27 89 82 17 89  0
#> 140 83 89 82 17 89  0
#> 141 27 44 44 70 89  0
#> 142 83 44 44 70 89  0
#> 143 27 89 44 70 89  0
#> 144 83 89 44 70 89  0
#> 145 27 44 82 70 89  0
#> 146 83 44 82 70 89  0
#> 147 27 89 82 70 89  0
#> 150 83 67 54 38 33  0
#> 151  9 89 54 38 33  0
#> 152 83 89 54 38 33  0
#> 153  9 67 82 38 33  0
#> 154 83 67 82 38 33  0
#> 155  9 89 82 38 33  0
#> 156 83 89 82 38 33  0
#> 157  9 67 54 70 33  0
#> 158 83 67 54 70 33  0
#> 159  9 89 54 70 33  0
#> 160 83 89 54 70 33  0
#> 161  9 67 82 70 33  0
#> 162 83 67 82 70 33  0
#> 163  9 89 82 70 33  0
#> 164 83 89 82 70 33  0
#> 165  9 67 54 38 89  0
#> 166 83 67 54 38 89  0
#> 167  9 89 54 38 89  0
#> 168 83 89 54 38 89  0
#> 169  9 67 82 38 89  0
#> 170 83 67 82 38 89  0
#> 171  9 89 82 38 89  0
#> 172 83 89 82 38 89  0
#> 173  9 67 54 70 89  0
#> 174 83 67 54 70 89  0
#> 175  9 89 54 70 89  0
#> 176 83 89 54 70 89  0
#> 177  9 67 82 70 89  0
#> 178 83 67 82 70 89  0
#> 179  9 89 82 70 89  0
#> 182 83 31 55 23 78  0
#> 183 27 89 55 23 78  0
#> 184 83 89 55 23 78  0
#> 185 27 31 82 23 78  0
#> 186 83 31 82 23 78  0
#> 187 27 89 82 23 78  0
#> 188 83 89 82 23 78  0
#> 189 27 31 55 70 78  0
#> 190 83 31 55 70 78  0
#> 191 27 89 55 70 78  0
#> 192 83 89 55 70 78  0
#> 193 27 31 82 70 78  0
#> 194 83 31 82 70 78  0
#> 195 27 89 82 70 78  0
#> 196 83 89 82 70 78  0
#> 197 27 31 55 23 89  0
#> 198 83 31 55 23 89  0
#> 199 27 89 55 23 89  0
#> 200 83 89 55 23 89  0
#> 201 27 31 82 23 89  0
#> 202 83 31 82 23 89  0
#> 203 27 89 82 23 89  0
#> 204 83 89 82 23 89  0
#> 205 27 31 55 70 89  0
#> 206 83 31 55 70 89  0
#> 207 27 89 55 70 89  0
#> 208 83 89 55 70 89  0
#> 209 27 31 82 70 89  0
#> 210 83 31 82 70 89  0
#> 214 83 76 18 25 45  0
#> 215 58 89 18 25 45  0
#> 216 83 89 18 25 45  0
#> 217 58 76 82 25 45  0
#> 218 83 76 82 25 45  0
#> 219 58 89 82 25 45  0
#> 220 83 89 82 25 45  0
#> 221 58 76 18 70 45  0
#> 222 83 76 18 70 45  0
#> 223 58 89 18 70 45  0
#> 224 83 89 18 70 45  0
#> 225 58 76 82 70 45  0
#> 226 83 76 82 70 45  0
#> 227 58 89 82 70 45  0
#> 228 83 89 82 70 45  0
#> 229 58 76 18 25 89  0
#> 230 83 76 18 25 89  0
#> 231 58 89 18 25 89  0
#> 232 83 89 18 25 89  0
#> 233 58 76 82 25 89  0
#> 234 83 76 82 25 89  0
#> 235 58 89 82 25 89  0
#> 236 83 89 82 25 89  0
#> 237 58 76 18 70 89  0
#> 238 83 76 18 70 89  0
#> 239 58 89 18 70 89  0
#> 241 58 76 82 70 89  0
#> 242 83 76 82 70 89  0
#> 243 58 89 82 70 89  0
#> 246 83 84 50 29 24  0
#> 247 62 89 50 29 24  0
#> 248 83 89 50 29 24  0
#> 249 62 84 82 29 24  0
#> 250 83 84 82 29 24  0
#> 251 62 89 82 29 24  0
#> 252 83 89 82 29 24  0
#> 253 62 84 50 70 24  0
#> 254 83 84 50 70 24  0
#> 255 62 89 50 70 24  0
#> 256 83 89 50 70 24  0
#> 257 62 84 82 70 24  0
#> 258 83 84 82 70 24  0
#> 259 62 89 82 70 24  0
#> 261 62 84 50 29 89  0
#> 262 83 84 50 29 89  0
#> 263 62 89 50 29 89  0
#> 264 83 89 50 29 89  0
#> 265 62 84 82 29 89  0
#> 266 83 84 82 29 89  0
#> 267 62 89 82 29 89  0
#> 268 83 89 82 29 89  0
#> 269 62 84 50 70 89  0
#> 270 83 84 50 70 89  0
#> 271 62 89 50 70 89  0
#> 272 83 89 50 70 89  0
#> 273 62 84 82 70 89  0
#> 274 83 84 82 70 89  0
#> 275 62 89 82 70 89  0
#> 278 83 24 18 30 64  0
#> 279 42 89 18 30 64  0
#> 280 83 89 18 30 64  0
#> 281 42 24 82 30 64  0
#> 282 83 24 82 30 64  0
#> 283 42 89 82 30 64  0
#> 284 83 89 82 30 64  0
#> 285 42 24 18 70 64  0
#> 286 83 24 18 70 64  0
#> 287 42 89 18 70 64  0
#> 288 83 89 18 70 64  0
#> 289 42 24 82 70 64  0
#> 290 83 24 82 70 64  0
#> 291 42 89 82 70 64  0
#> 292 83 89 82 70 64  0
#> 293 42 24 18 30 89  0
#> 294 83 24 18 30 89  0
#> 295 42 89 18 30 89  0
#> 296 83 89 18 30 89  0
#> 297 42 24 82 30 89  0
#> 298 83 24 82 30 89  0
#> 299 42 89 82 30 89  0
#> 300 83 89 82 30 89  0
#> 301 42 24 18 70 89  0
#> 302 83 24 18 70 89  0
#> 303 42 89 18 70 89  0
#> 305 42 24 82 70 89  0
#> 306 83 24 82 70 89  0
#> 307 42 89 82 70 89  0
#> 310 83 53  4 65 55  0
#> 311 44 89  4 65 55  0
#> 312 83 89  4 65 55  0
#> 313 44 53 82 65 55  0
#> 314 83 53 82 65 55  0
#> 315 44 89 82 65 55  0
#> 316 83 89 82 65 55  0
#> 317 44 53  4 70 55  0
#> 318 83 53  4 70 55  0
#> 319 44 89  4 70 55  0
#> 320 83 89  4 70 55  0
#> 321 44 53 82 70 55  0
#> 322 83 53 82 70 55  0
#> 323 44 89 82 70 55  0
#> 324 83 89 82 70 55  0
#> 325 44 53  4 65 89  0
#> 326 83 53  4 65 89  0
#> 327 44 89  4 65 89  0
#> 328 83 89  4 65 89  0
#> 329 44 53 82 65 89  0
#> 330 83 53 82 65 89  0
#> 331 44 89 82 65 89  0
#> 332 83 89 82 65 89  0
#> 333 44 53  4 70 89  0
#> 334 83 53  4 70 89  0
#> 335 44 89  4 70 89  0
#> 336 83 89  4 70 89  0
#> 337 44 53 82 70 89  0
#> 338 83 53 82 70 89  0
#> 339 44 89 82 70 89  0
#> 342 83 53  9 35 54  0
#> 343 73 89  9 35 54  0
#> 344 83 89  9 35 54  0
#> 345 73 53 82 35 54  0
#> 346 83 53 82 35 54  0
#> 347 73 89 82 35 54  0
#> 348 83 89 82 35 54  0
#> 349 73 53  9 70 54  0
#> 350 83 53  9 70 54  0
#> 351 73 89  9 70 54  0
#> 352 83 89  9 70 54  0
#> 353 73 53 82 70 54  0
#> 354 83 53 82 70 54  0
#> 355 73 89 82 70 54  0
#> 356 83 89 82 70 54  0
#> 357 73 53  9 35 89  0
#> 358 83 53  9 35 89  0
#> 359 73 89  9 35 89  0
#> 360 83 89  9 35 89  0
#> 361 73 53 82 35 89  0
#> 362 83 53 82 35 89  0
#> 363 73 89 82 35 89  0
#> 364 83 89 82 35 89  0
#> 365 73 53  9 70 89  0
#> 366 83 53  9 70 89  0
#> 367 73 89  9 70 89  0
#> 368 83 89  9 70 89  0
#> 369 73 53 82 70 89  0
#> 371 73 89 82 70 89  0
#> 374 83 74 44 61 11  0
#> 375 38 89 44 61 11  0
#> 376 83 89 44 61 11  0
#> 377 38 74 82 61 11  0
#> 378 83 74 82 61 11  0
#> 379 38 89 82 61 11  0
#> 380 83 89 82 61 11  0
#> 381 38 74 44 70 11  0
#> 382 83 74 44 70 11  0
#> 383 38 89 44 70 11  0
#> 384 83 89 44 70 11  0
#> 385 38 74 82 70 11  0
#> 386 83 74 82 70 11  0
#> 387 38 89 82 70 11  0
#> 388 83 89 82 70 11  0
#> 389 38 74 44 61 89  0
#> 390 83 74 44 61 89  0
#> 391 38 89 44 61 89  0
#> 392 83 89 44 61 89  0
#> 393 38 74 82 61 89  0
#> 394 83 74 82 61 89  0
#> 395 38 89 82 61 89  0
#> 396 83 89 82 61 89  0
#> 397 38 74 44 70 89  0
#> 398 83 74 44 70 89  0
#> 399 38 89 44 70 89  0
#> 401 38 74 82 70 89  0
#> 402 83 74 82 70 89  0
#> 403 38 89 82 70 89  0
#> 406 83 26 26 49 31  0
#> 407 21 89 26 49 31  0
#> 408 83 89 26 49 31  0
#> 409 21 26 82 49 31  0
#> 410 83 26 82 49 31  0
#> 411 21 89 82 49 31  0
#> 412 83 89 82 49 31  0
#> 413 21 26 26 70 31  0
#> 414 83 26 26 70 31  0
#> 415 21 89 26 70 31  0
#> 416 83 89 26 70 31  0
#> 417 21 26 82 70 31  0
#> 418 83 26 82 70 31  0
#> 419 21 89 82 70 31  0
#> 420 83 89 82 70 31  0
#> 421 21 26 26 49 89  0
#> 422 83 26 26 49 89  0
#> 423 21 89 26 49 89  0
#> 424 83 89 26 49 89  0
#> 425 21 26 82 49 89  0
#> 426 83 26 82 49 89  0
#> 427 21 89 82 49 89  0
#> 428 83 89 82 49 89  0
#> 429 21 26 26 70 89  0
#> 430 83 26 26 70 89  0
#> 431 21 89 26 70 89  0
#> 432 83 89 26 70 89  0
#> 433 21 26 82 70 89  0
#> 438 83 28 56 30 12  0
#> 439 46 89 56 30 12  0
#> 440 83 89 56 30 12  0
#> 441 46 28 82 30 12  0
#> 442 83 28 82 30 12  0
#> 443 46 89 82 30 12  0
#> 444 83 89 82 30 12  0
#> 445 46 28 56 70 12  0
#> 446 83 28 56 70 12  0
#> 447 46 89 56 70 12  0
#> 448 83 89 56 70 12  0
#> 449 46 28 82 70 12  0
#> 450 83 28 82 70 12  0
#> 451 46 89 82 70 12  0
#> 453 46 28 56 30 89  0
#> 454 83 28 56 30 89  0
#> 455 46 89 56 30 89  0
#> 456 83 89 56 30 89  0
#> 457 46 28 82 30 89  0
#> 458 83 28 82 30 89  0
#> 459 46 89 82 30 89  0
#> 461 46 28 56 70 89  0
#> 462 83 28 56 70 89  0
#> 463 46 89 56 70 89  0
#> 464 83 89 56 70 89  0
#> 465 46 28 82 70 89  0
#> 466 83 28 82 70 89  0
#> 467 46 89 82 70 89  0
#> 470 83 28 33 40 84  0
#> 471 30 89 33 40 84  0
#> 472 83 89 33 40 84  0
#> 473 30 28 82 40 84  0
#> 474 83 28 82 40 84  0
#> 475 30 89 82 40 84  0
#> 476 83 89 82 40 84  0
#> 477 30 28 33 70 84  0
#> 478 83 28 33 70 84  0
#> 479 30 89 33 70 84  0
#> 480 83 89 33 70 84  0
#> 481 30 28 82 70 84  0
#> 482 83 28 82 70 84  0
#> 483 30 89 82 70 84  0
#> 484 83 89 82 70 84  0
#> 485 30 28 33 40 89  0
#> 486 83 28 33 40 89  0
#> 487 30 89 33 40 89  0
#> 488 83 89 33 40 89  0
#> 489 30 28 82 40 89  0
#> 490 83 28 82 40 89  0
#> 491 30 89 82 40 89  0
#> 492 83 89 82 40 89  0
#> 493 30 28 33 70 89  0
#> 494 83 28 33 70 89  0
#> 495 30 89 33 70 89  0
#> 496 83 89 33 70 89  0
#> 497 30 28 82 70 89  0
#> 499 30 89 82 70 89  0
#> 502 83 23 61 45 71  0
#> 503 17 89 61 45 71  0
#> 504 83 89 61 45 71  0
#> 505 17 23 82 45 71  0
#> 506 83 23 82 45 71  0
#> 507 17 89 82 45 71  0
#> 508 83 89 82 45 71  0
#> 509 17 23 61 70 71  0
#> 510 83 23 61 70 71  0
#> 511 17 89 61 70 71  0
#> 512 83 89 61 70 71  0
#> 513 17 23 82 70 71  0
#> 514 83 23 82 70 71  0
#> 515 17 89 82 70 71  0
#> 516 83 89 82 70 71  0
#> 517 17 23 61 45 89  0
#> 518 83 23 61 45 89  0
#> 519 17 89 61 45 89  0
#> 520 83 89 61 45 89  0
#> 521 17 23 82 45 89  0
#> 522 83 23 82 45 89  0
#> 523 17 89 82 45 89  0
#> 524 83 89 82 45 89  0
#> 525 17 23 61 70 89  0
#> 526 83 23 61 70 89  0
#> 527 17 89 61 70 89  0
#> 528 83 89 61 70 89  0
#> 529 17 23 82 70 89  0
#> 530 83 23 82 70 89  0
#> 531 17 89 82 70 89  0
#> 534 83 70 63 43 12  0
#> 535 31 89 63 43 12  0
#> 536 83 89 63 43 12  0
#> 537 31 70 82 43 12  0
#> 538 83 70 82 43 12  0
#> 539 31 89 82 43 12  0
#> 540 83 89 82 43 12  0
#> 541 31 70 63 70 12  0
#> 542 83 70 63 70 12  0
#> 543 31 89 63 70 12  0
#> 544 83 89 63 70 12  0
#> 545 31 70 82 70 12  0
#> 546 83 70 82 70 12  0
#> 547 31 89 82 70 12  0
#> 549 31 70 63 43 89  0
#> 550 83 70 63 43 89  0
#> 551 31 89 63 43 89  0
#> 552 83 89 63 43 89  0
#> 553 31 70 82 43 89  0
#> 554 83 70 82 43 89  0
#> 555 31 89 82 43 89  0
#> 556 83 89 82 43 89  0
#> 557 31 70 63 70 89  0
#> 558 83 70 63 70 89  0
#> 559 31 89 63 70 89  0
#> 560 83 89 63 70 89  0
#> 561 31 70 82 70 89  0
#> 562 83 70 82 70 89  0
#> 563 31 89 82 70 89  0
#> 566 83 67 17 61 29  0
#> 567 73 89 17 61 29  0
#> 568 83 89 17 61 29  0
#> 569 73 67 82 61 29  0
#> 570 83 67 82 61 29  0
#> 571 73 89 82 61 29  0
#> 572 83 89 82 61 29  0
#> 573 73 67 17 70 29  0
#> 574 83 67 17 70 29  0
#> 575 73 89 17 70 29  0
#> 576 83 89 17 70 29  0
#> 577 73 67 82 70 29  0
#> 578 83 67 82 70 29  0
#> 579 73 89 82 70 29  0
#> 580 83 89 82 70 29  0
#> 581 73 67 17 61 89  0
#> 582 83 67 17 61 89  0
#> 583 73 89 17 61 89  0
#> 584 83 89 17 61 89  0
#> 585 73 67 82 61 89  0
#> 586 83 67 82 61 89  0
#> 587 73 89 82 61 89  0
#> 589 73 67 17 70 89  0
#> 590 83 67 17 70 89  0
#> 591 73 89 17 70 89  0
#> 592 83 89 17 70 89  0
#> 593 73 67 82 70 89  0
#> 598 83 49 32 53 14  0
#> 599 78 89 32 53 14  0
#> 600 83 89 32 53 14  0
#> 601 78 49 82 53 14  0
#> 602 83 49 82 53 14  0
#> 603 78 89 82 53 14  0
#> 604 83 89 82 53 14  0
#> 605 78 49 32 70 14  0
#> 606 83 49 32 70 14  0
#> 607 78 89 32 70 14  0
#> 608 83 89 32 70 14  0
#> 609 78 49 82 70 14  0
#> 610 83 49 82 70 14  0
#> 611 78 89 82 70 14  0
#> 612 83 89 82 70 14  0
#> 613 78 49 32 53 89  0
#> 614 83 49 32 53 89  0
#> 615 78 89 32 53 89  0
#> 616 83 89 32 53 89  0
#> 617 78 49 82 53 89  0
#> 618 83 49 82 53 89  0
#> 619 78 89 82 53 89  0
#> 620 83 89 82 53 89  0
#> 621 78 49 32 70 89  0
#> 622 83 49 32 70 89  0
#> 623 78 89 32 70 89  0
#> 624 83 89 32 70 89  0
#> 625 78 49 82 70 89  0
#> 626 83 49 82 70 89  0
#> 627 78 89 82 70 89  0
#> 630 83 37 77 24 71  0
#> 631 30 89 77 24 71  0
#> 632 83 89 77 24 71  0
#> 633 30 37 82 24 71  0
#> 634 83 37 82 24 71  0
#> 635 30 89 82 24 71  0
#> 636 83 89 82 24 71  0
#> 637 30 37 77 70 71  0
#> 638 83 37 77 70 71  0
#> 639 30 89 77 70 71  0
#> 640 83 89 77 70 71  0
#> 641 30 37 82 70 71  0
#> 642 83 37 82 70 71  0
#> 643 30 89 82 70 71  0
#> 645 30 37 77 24 89  0
#> 646 83 37 77 24 89  0
#> 647 30 89 77 24 89  0
#> 648 83 89 77 24 89  0
#> 649 30 37 82 24 89  0
#> 650 83 37 82 24 89  0
#> 651 30 89 82 24 89  0
#> 652 83 89 82 24 89  0
#> 653 30 37 77 70 89  0
#> 654 83 37 77 70 89  0
#> 655 30 89 77 70 89  0
#> 656 83 89 77 70 89  0
#> 657 30 37 82 70 89  0
#> 658 83 37 82 70 89  0