Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 1  56  1 47  1
#> 2  16 18 64  1
#> 3   6 71 45  1
#> 4  30  8 66  1
#> 5  54 41  2  1
#> 6  53 47  1  1
#> 7  55  6 29  1
#> 8  31  8 34  1
#> 9   2 55 57  1
#> 10  7 72 43  1
#> 12 61  1 47  0
#> 13 56 77 47  0
#> 14 61 77 47  0
#> 15 56  1 71  0
#> 16 61  1 71  0
#> 17 56 77 71  0
#> 18 61 77 71  0
#> 20 61 18 64  0
#> 21 16 77 64  0
#> 22 61 77 64  0
#> 23 16 18 71  0
#> 24 61 18 71  0
#> 25 16 77 71  0
#> 28 61 71 45  0
#> 29  6 77 45  0
#> 30 61 77 45  0
#> 31  6 71 71  0
#> 32 61 71 71  0
#> 33  6 77 71  0
#> 36 61  8 66  0
#> 37 30 77 66  0
#> 38 61 77 66  0
#> 39 30  8 71  0
#> 40 61  8 71  0
#> 41 30 77 71  0
#> 44 61 41  2  0
#> 45 54 77  2  0
#> 46 61 77  2  0
#> 47 54 41 71  0
#> 48 61 41 71  0
#> 49 54 77 71  0
#> 52 61 47  1  0
#> 53 53 77  1  0
#> 54 61 77  1  0
#> 55 53 47 71  0
#> 56 61 47 71  0
#> 57 53 77 71  0
#> 60 61  6 29  0
#> 61 55 77 29  0
#> 62 61 77 29  0
#> 63 55  6 71  0
#> 64 61  6 71  0
#> 65 55 77 71  0
#> 68 61  8 34  0
#> 69 31 77 34  0
#> 70 61 77 34  0
#> 71 31  8 71  0
#> 73 31 77 71  0
#> 76 61 55 57  0
#> 77  2 77 57  0
#> 78 61 77 57  0
#> 79  2 55 71  0
#> 80 61 55 71  0
#> 81  2 77 71  0
#> 84 61 72 43  0
#> 85  7 77 43  0
#> 86 61 77 43  0
#> 87  7 72 71  0
#> 88 61 72 71  0
#> 89  7 77 71  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 1  56  1 47  1
#> 2  16 18 64  1
#> 3   6 71 45  1
#> 4  30  8 66  1
#> 5  54 41  2  1
#> 6  53 47  1  1
#> 7  55  6 29  1
#> 8  31  8 34  1
#> 9   2 55 57  1
#> 10  7 72 43  1
#> 12 61  1 47  0
#> 13 56 -4 47  0
#> 14 61 -4 47  0
#> 15 56  1 71  0
#> 16 61  1 71  0
#> 17 56 -4 71  0
#> 18 61 -4 71  0
#> 20 61 18 64  0
#> 21 16 -4 64  0
#> 22 61 -4 64  0
#> 23 16 18 71  0
#> 24 61 18 71  0
#> 25 16 -4 71  0
#> 28 61 71 45  0
#> 29  6 -4 45  0
#> 30 61 -4 45  0
#> 31  6 71 71  0
#> 32 61 71 71  0
#> 33  6 -4 71  0
#> 36 61  8 66  0
#> 37 30 -4 66  0
#> 38 61 -4 66  0
#> 39 30  8 71  0
#> 40 61  8 71  0
#> 41 30 -4 71  0
#> 44 61 41  2  0
#> 45 54 -4  2  0
#> 46 61 -4  2  0
#> 47 54 41 71  0
#> 48 61 41 71  0
#> 49 54 -4 71  0
#> 52 61 47  1  0
#> 53 53 -4  1  0
#> 54 61 -4  1  0
#> 55 53 47 71  0
#> 56 61 47 71  0
#> 57 53 -4 71  0
#> 60 61  6 29  0
#> 61 55 -4 29  0
#> 62 61 -4 29  0
#> 63 55  6 71  0
#> 64 61  6 71  0
#> 65 55 -4 71  0
#> 68 61  8 34  0
#> 69 31 -4 34  0
#> 70 61 -4 34  0
#> 71 31  8 71  0
#> 73 31 -4 71  0
#> 76 61 55 57  0
#> 77  2 -4 57  0
#> 78 61 -4 57  0
#> 79  2 55 71  0
#> 80 61 55 71  0
#> 81  2 -4 71  0
#> 84 61 72 43  0
#> 85  7 -4 43  0
#> 86 61 -4 43  0
#> 87  7 72 71  0
#> 88 61 72 71  0
#> 89  7 -4 71  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 1  56  1  47  1
#> 2  16 18  64  1
#> 3   6 71  45  1
#> 4  30  8  66  1
#> 5  54 41   2  1
#> 6  53 47   1  1
#> 7  55  6  29  1
#> 8  31  8  34  1
#> 9   2 55  57  1
#> 10  7 72  43  1
#> 12  0  1  47  0
#> 13 56  0  47  0
#> 14  0  0  47  0
#> 15 56  1 100  0
#> 16  0  1 100  0
#> 17 56  0 100  0
#> 18  0  0 100  0
#> 20  0 18  64  0
#> 21 16  0  64  0
#> 22  0  0  64  0
#> 23 16 18 100  0
#> 24  0 18 100  0
#> 25 16  0 100  0
#> 28  0 71  45  0
#> 29  6  0  45  0
#> 30  0  0  45  0
#> 31  6 71 100  0
#> 32  0 71 100  0
#> 33  6  0 100  0
#> 36  0  8  66  0
#> 37 30  0  66  0
#> 38  0  0  66  0
#> 39 30  8 100  0
#> 40  0  8 100  0
#> 41 30  0 100  0
#> 44  0 41   2  0
#> 45 54  0   2  0
#> 46  0  0   2  0
#> 47 54 41 100  0
#> 48  0 41 100  0
#> 49 54  0 100  0
#> 52  0 47   1  0
#> 53 53  0   1  0
#> 54  0  0   1  0
#> 55 53 47 100  0
#> 56  0 47 100  0
#> 57 53  0 100  0
#> 60  0  6  29  0
#> 61 55  0  29  0
#> 62  0  0  29  0
#> 63 55  6 100  0
#> 64  0  6 100  0
#> 65 55  0 100  0
#> 68  0  8  34  0
#> 69 31  0  34  0
#> 70  0  0  34  0
#> 71 31  8 100  0
#> 73 31  0 100  0
#> 76  0 55  57  0
#> 77  2  0  57  0
#> 78  0  0  57  0
#> 79  2 55 100  0
#> 80  0 55 100  0
#> 81  2  0 100  0
#> 84  0 72  43  0
#> 85  7  0  43  0
#> 86  0  0  43  0
#> 87  7 72 100  0
#> 88  0 72 100  0
#> 89  7  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     z1 z2 z3 z4 z5 pt
#> 1   20 55 52 84 32  1
#> 2   52 51 63  7 69  1
#> 3   69  8 38 60 46  1
#> 4   72 13 65 33 44  1
#> 5   22 38 87 39 50  1
#> 6   22 83 26 49 50  1
#> 7   69 25 61 20 31  1
#> 8    6 38 40 64 46  1
#> 9   79 25 55 66 24  1
#> 10  68 45 74 36 14  1
#> 11  26 79 36 47 22  1
#> 12  58 12 72 34 38  1
#> 13  47 19 64 52 15  1
#> 14  39  4 63 53 56  1
#> 15  57 44 49 84 15  1
#> 16  58 12 60 53 22  1
#> 17  47 12 54 22 61  1
#> 18  32 21 69 35 25  1
#> 19  33 91 51 41 31  1
#> 20  76 27 62 60 19  1
#> 22  84 55 52 84 32  0
#> 23  20 96 52 84 32  0
#> 24  84 96 52 84 32  0
#> 25  20 55 92 84 32  0
#> 26  84 55 92 84 32  0
#> 27  20 96 92 84 32  0
#> 28  84 96 92 84 32  0
#> 29  20 55 52 89 32  0
#> 30  84 55 52 89 32  0
#> 31  20 96 52 89 32  0
#> 32  84 96 52 89 32  0
#> 33  20 55 92 89 32  0
#> 34  84 55 92 89 32  0
#> 35  20 96 92 89 32  0
#> 36  84 96 92 89 32  0
#> 37  20 55 52 84 74  0
#> 38  84 55 52 84 74  0
#> 39  20 96 52 84 74  0
#> 40  84 96 52 84 74  0
#> 41  20 55 92 84 74  0
#> 42  84 55 92 84 74  0
#> 43  20 96 92 84 74  0
#> 44  84 96 92 84 74  0
#> 45  20 55 52 89 74  0
#> 46  84 55 52 89 74  0
#> 47  20 96 52 89 74  0
#> 48  84 96 52 89 74  0
#> 49  20 55 92 89 74  0
#> 50  84 55 92 89 74  0
#> 51  20 96 92 89 74  0
#> 52  84 96 92 89 74  0
#> 54  84 51 63  7 69  0
#> 55  52 96 63  7 69  0
#> 56  84 96 63  7 69  0
#> 57  52 51 92  7 69  0
#> 58  84 51 92  7 69  0
#> 59  52 96 92  7 69  0
#> 60  84 96 92  7 69  0
#> 61  52 51 63 89 69  0
#> 62  84 51 63 89 69  0
#> 63  52 96 63 89 69  0
#> 64  84 96 63 89 69  0
#> 65  52 51 92 89 69  0
#> 66  84 51 92 89 69  0
#> 67  52 96 92 89 69  0
#> 68  84 96 92 89 69  0
#> 69  52 51 63  7 74  0
#> 70  84 51 63  7 74  0
#> 71  52 96 63  7 74  0
#> 72  84 96 63  7 74  0
#> 73  52 51 92  7 74  0
#> 74  84 51 92  7 74  0
#> 75  52 96 92  7 74  0
#> 76  84 96 92  7 74  0
#> 77  52 51 63 89 74  0
#> 78  84 51 63 89 74  0
#> 79  52 96 63 89 74  0
#> 80  84 96 63 89 74  0
#> 81  52 51 92 89 74  0
#> 82  84 51 92 89 74  0
#> 83  52 96 92 89 74  0
#> 86  84  8 38 60 46  0
#> 87  69 96 38 60 46  0
#> 88  84 96 38 60 46  0
#> 89  69  8 92 60 46  0
#> 90  84  8 92 60 46  0
#> 91  69 96 92 60 46  0
#> 92  84 96 92 60 46  0
#> 93  69  8 38 89 46  0
#> 94  84  8 38 89 46  0
#> 95  69 96 38 89 46  0
#> 96  84 96 38 89 46  0
#> 97  69  8 92 89 46  0
#> 98  84  8 92 89 46  0
#> 99  69 96 92 89 46  0
#> 100 84 96 92 89 46  0
#> 101 69  8 38 60 74  0
#> 102 84  8 38 60 74  0
#> 103 69 96 38 60 74  0
#> 104 84 96 38 60 74  0
#> 105 69  8 92 60 74  0
#> 106 84  8 92 60 74  0
#> 107 69 96 92 60 74  0
#> 108 84 96 92 60 74  0
#> 109 69  8 38 89 74  0
#> 110 84  8 38 89 74  0
#> 111 69 96 38 89 74  0
#> 112 84 96 38 89 74  0
#> 113 69  8 92 89 74  0
#> 114 84  8 92 89 74  0
#> 115 69 96 92 89 74  0
#> 118 84 13 65 33 44  0
#> 119 72 96 65 33 44  0
#> 120 84 96 65 33 44  0
#> 121 72 13 92 33 44  0
#> 122 84 13 92 33 44  0
#> 123 72 96 92 33 44  0
#> 124 84 96 92 33 44  0
#> 125 72 13 65 89 44  0
#> 126 84 13 65 89 44  0
#> 127 72 96 65 89 44  0
#> 128 84 96 65 89 44  0
#> 129 72 13 92 89 44  0
#> 130 84 13 92 89 44  0
#> 131 72 96 92 89 44  0
#> 132 84 96 92 89 44  0
#> 133 72 13 65 33 74  0
#> 134 84 13 65 33 74  0
#> 135 72 96 65 33 74  0
#> 136 84 96 65 33 74  0
#> 137 72 13 92 33 74  0
#> 138 84 13 92 33 74  0
#> 139 72 96 92 33 74  0
#> 140 84 96 92 33 74  0
#> 141 72 13 65 89 74  0
#> 142 84 13 65 89 74  0
#> 143 72 96 65 89 74  0
#> 144 84 96 65 89 74  0
#> 145 72 13 92 89 74  0
#> 146 84 13 92 89 74  0
#> 147 72 96 92 89 74  0
#> 150 84 38 87 39 50  0
#> 151 22 96 87 39 50  0
#> 152 84 96 87 39 50  0
#> 153 22 38 92 39 50  0
#> 154 84 38 92 39 50  0
#> 155 22 96 92 39 50  0
#> 156 84 96 92 39 50  0
#> 157 22 38 87 89 50  0
#> 158 84 38 87 89 50  0
#> 159 22 96 87 89 50  0
#> 160 84 96 87 89 50  0
#> 161 22 38 92 89 50  0
#> 162 84 38 92 89 50  0
#> 163 22 96 92 89 50  0
#> 164 84 96 92 89 50  0
#> 165 22 38 87 39 74  0
#> 166 84 38 87 39 74  0
#> 167 22 96 87 39 74  0
#> 168 84 96 87 39 74  0
#> 169 22 38 92 39 74  0
#> 170 84 38 92 39 74  0
#> 171 22 96 92 39 74  0
#> 172 84 96 92 39 74  0
#> 173 22 38 87 89 74  0
#> 174 84 38 87 89 74  0
#> 175 22 96 87 89 74  0
#> 176 84 96 87 89 74  0
#> 177 22 38 92 89 74  0
#> 178 84 38 92 89 74  0
#> 179 22 96 92 89 74  0
#> 182 84 83 26 49 50  0
#> 183 22 96 26 49 50  0
#> 184 84 96 26 49 50  0
#> 185 22 83 92 49 50  0
#> 186 84 83 92 49 50  0
#> 187 22 96 92 49 50  0
#> 188 84 96 92 49 50  0
#> 189 22 83 26 89 50  0
#> 190 84 83 26 89 50  0
#> 191 22 96 26 89 50  0
#> 192 84 96 26 89 50  0
#> 193 22 83 92 89 50  0
#> 194 84 83 92 89 50  0
#> 197 22 83 26 49 74  0
#> 198 84 83 26 49 74  0
#> 199 22 96 26 49 74  0
#> 200 84 96 26 49 74  0
#> 201 22 83 92 49 74  0
#> 202 84 83 92 49 74  0
#> 203 22 96 92 49 74  0
#> 204 84 96 92 49 74  0
#> 205 22 83 26 89 74  0
#> 206 84 83 26 89 74  0
#> 207 22 96 26 89 74  0
#> 208 84 96 26 89 74  0
#> 209 22 83 92 89 74  0
#> 210 84 83 92 89 74  0
#> 214 84 25 61 20 31  0
#> 215 69 96 61 20 31  0
#> 216 84 96 61 20 31  0
#> 217 69 25 92 20 31  0
#> 218 84 25 92 20 31  0
#> 219 69 96 92 20 31  0
#> 220 84 96 92 20 31  0
#> 221 69 25 61 89 31  0
#> 222 84 25 61 89 31  0
#> 223 69 96 61 89 31  0
#> 224 84 96 61 89 31  0
#> 225 69 25 92 89 31  0
#> 226 84 25 92 89 31  0
#> 227 69 96 92 89 31  0
#> 228 84 96 92 89 31  0
#> 229 69 25 61 20 74  0
#> 230 84 25 61 20 74  0
#> 231 69 96 61 20 74  0
#> 232 84 96 61 20 74  0
#> 233 69 25 92 20 74  0
#> 234 84 25 92 20 74  0
#> 235 69 96 92 20 74  0
#> 236 84 96 92 20 74  0
#> 237 69 25 61 89 74  0
#> 238 84 25 61 89 74  0
#> 239 69 96 61 89 74  0
#> 240 84 96 61 89 74  0
#> 241 69 25 92 89 74  0
#> 242 84 25 92 89 74  0
#> 246 84 38 40 64 46  0
#> 247  6 96 40 64 46  0
#> 248 84 96 40 64 46  0
#> 249  6 38 92 64 46  0
#> 250 84 38 92 64 46  0
#> 251  6 96 92 64 46  0
#> 252 84 96 92 64 46  0
#> 253  6 38 40 89 46  0
#> 254 84 38 40 89 46  0
#> 255  6 96 40 89 46  0
#> 256 84 96 40 89 46  0
#> 257  6 38 92 89 46  0
#> 258 84 38 92 89 46  0
#> 259  6 96 92 89 46  0
#> 261  6 38 40 64 74  0
#> 262 84 38 40 64 74  0
#> 263  6 96 40 64 74  0
#> 264 84 96 40 64 74  0
#> 265  6 38 92 64 74  0
#> 266 84 38 92 64 74  0
#> 267  6 96 92 64 74  0
#> 268 84 96 92 64 74  0
#> 269  6 38 40 89 74  0
#> 270 84 38 40 89 74  0
#> 271  6 96 40 89 74  0
#> 272 84 96 40 89 74  0
#> 273  6 38 92 89 74  0
#> 275  6 96 92 89 74  0
#> 278 84 25 55 66 24  0
#> 279 79 96 55 66 24  0
#> 280 84 96 55 66 24  0
#> 281 79 25 92 66 24  0
#> 282 84 25 92 66 24  0
#> 283 79 96 92 66 24  0
#> 284 84 96 92 66 24  0
#> 285 79 25 55 89 24  0
#> 286 84 25 55 89 24  0
#> 287 79 96 55 89 24  0
#> 288 84 96 55 89 24  0
#> 289 79 25 92 89 24  0
#> 290 84 25 92 89 24  0
#> 291 79 96 92 89 24  0
#> 292 84 96 92 89 24  0
#> 293 79 25 55 66 74  0
#> 294 84 25 55 66 74  0
#> 295 79 96 55 66 74  0
#> 296 84 96 55 66 74  0
#> 297 79 25 92 66 74  0
#> 298 84 25 92 66 74  0
#> 299 79 96 92 66 74  0
#> 300 84 96 92 66 74  0
#> 301 79 25 55 89 74  0
#> 302 84 25 55 89 74  0
#> 303 79 96 55 89 74  0
#> 304 84 96 55 89 74  0
#> 305 79 25 92 89 74  0
#> 307 79 96 92 89 74  0
#> 310 84 45 74 36 14  0
#> 311 68 96 74 36 14  0
#> 312 84 96 74 36 14  0
#> 313 68 45 92 36 14  0
#> 314 84 45 92 36 14  0
#> 315 68 96 92 36 14  0
#> 316 84 96 92 36 14  0
#> 317 68 45 74 89 14  0
#> 318 84 45 74 89 14  0
#> 319 68 96 74 89 14  0
#> 320 84 96 74 89 14  0
#> 321 68 45 92 89 14  0
#> 322 84 45 92 89 14  0
#> 323 68 96 92 89 14  0
#> 324 84 96 92 89 14  0
#> 325 68 45 74 36 74  0
#> 326 84 45 74 36 74  0
#> 327 68 96 74 36 74  0
#> 328 84 96 74 36 74  0
#> 329 68 45 92 36 74  0
#> 330 84 45 92 36 74  0
#> 331 68 96 92 36 74  0
#> 332 84 96 92 36 74  0
#> 333 68 45 74 89 74  0
#> 334 84 45 74 89 74  0
#> 335 68 96 74 89 74  0
#> 336 84 96 74 89 74  0
#> 337 68 45 92 89 74  0
#> 338 84 45 92 89 74  0
#> 339 68 96 92 89 74  0
#> 342 84 79 36 47 22  0
#> 343 26 96 36 47 22  0
#> 344 84 96 36 47 22  0
#> 345 26 79 92 47 22  0
#> 346 84 79 92 47 22  0
#> 347 26 96 92 47 22  0
#> 348 84 96 92 47 22  0
#> 349 26 79 36 89 22  0
#> 350 84 79 36 89 22  0
#> 351 26 96 36 89 22  0
#> 352 84 96 36 89 22  0
#> 353 26 79 92 89 22  0
#> 354 84 79 92 89 22  0
#> 355 26 96 92 89 22  0
#> 356 84 96 92 89 22  0
#> 357 26 79 36 47 74  0
#> 358 84 79 36 47 74  0
#> 359 26 96 36 47 74  0
#> 360 84 96 36 47 74  0
#> 361 26 79 92 47 74  0
#> 362 84 79 92 47 74  0
#> 363 26 96 92 47 74  0
#> 364 84 96 92 47 74  0
#> 365 26 79 36 89 74  0
#> 366 84 79 36 89 74  0
#> 367 26 96 36 89 74  0
#> 368 84 96 36 89 74  0
#> 369 26 79 92 89 74  0
#> 370 84 79 92 89 74  0
#> 371 26 96 92 89 74  0
#> 374 84 12 72 34 38  0
#> 375 58 96 72 34 38  0
#> 376 84 96 72 34 38  0
#> 377 58 12 92 34 38  0
#> 378 84 12 92 34 38  0
#> 379 58 96 92 34 38  0
#> 380 84 96 92 34 38  0
#> 381 58 12 72 89 38  0
#> 382 84 12 72 89 38  0
#> 383 58 96 72 89 38  0
#> 384 84 96 72 89 38  0
#> 385 58 12 92 89 38  0
#> 386 84 12 92 89 38  0
#> 387 58 96 92 89 38  0
#> 388 84 96 92 89 38  0
#> 389 58 12 72 34 74  0
#> 390 84 12 72 34 74  0
#> 391 58 96 72 34 74  0
#> 392 84 96 72 34 74  0
#> 393 58 12 92 34 74  0
#> 394 84 12 92 34 74  0
#> 395 58 96 92 34 74  0
#> 396 84 96 92 34 74  0
#> 397 58 12 72 89 74  0
#> 398 84 12 72 89 74  0
#> 399 58 96 72 89 74  0
#> 400 84 96 72 89 74  0
#> 401 58 12 92 89 74  0
#> 402 84 12 92 89 74  0
#> 403 58 96 92 89 74  0
#> 406 84 19 64 52 15  0
#> 407 47 96 64 52 15  0
#> 408 84 96 64 52 15  0
#> 409 47 19 92 52 15  0
#> 410 84 19 92 52 15  0
#> 411 47 96 92 52 15  0
#> 412 84 96 92 52 15  0
#> 413 47 19 64 89 15  0
#> 414 84 19 64 89 15  0
#> 415 47 96 64 89 15  0
#> 416 84 96 64 89 15  0
#> 417 47 19 92 89 15  0
#> 418 84 19 92 89 15  0
#> 419 47 96 92 89 15  0
#> 420 84 96 92 89 15  0
#> 421 47 19 64 52 74  0
#> 422 84 19 64 52 74  0
#> 423 47 96 64 52 74  0
#> 424 84 96 64 52 74  0
#> 425 47 19 92 52 74  0
#> 426 84 19 92 52 74  0
#> 427 47 96 92 52 74  0
#> 428 84 96 92 52 74  0
#> 429 47 19 64 89 74  0
#> 430 84 19 64 89 74  0
#> 431 47 96 64 89 74  0
#> 432 84 96 64 89 74  0
#> 433 47 19 92 89 74  0
#> 434 84 19 92 89 74  0
#> 435 47 96 92 89 74  0
#> 438 84  4 63 53 56  0
#> 439 39 96 63 53 56  0
#> 440 84 96 63 53 56  0
#> 441 39  4 92 53 56  0
#> 442 84  4 92 53 56  0
#> 443 39 96 92 53 56  0
#> 444 84 96 92 53 56  0
#> 445 39  4 63 89 56  0
#> 446 84  4 63 89 56  0
#> 447 39 96 63 89 56  0
#> 448 84 96 63 89 56  0
#> 449 39  4 92 89 56  0
#> 450 84  4 92 89 56  0
#> 451 39 96 92 89 56  0
#> 452 84 96 92 89 56  0
#> 453 39  4 63 53 74  0
#> 454 84  4 63 53 74  0
#> 455 39 96 63 53 74  0
#> 456 84 96 63 53 74  0
#> 457 39  4 92 53 74  0
#> 458 84  4 92 53 74  0
#> 459 39 96 92 53 74  0
#> 460 84 96 92 53 74  0
#> 461 39  4 63 89 74  0
#> 462 84  4 63 89 74  0
#> 463 39 96 63 89 74  0
#> 465 39  4 92 89 74  0
#> 466 84  4 92 89 74  0
#> 467 39 96 92 89 74  0
#> 470 84 44 49 84 15  0
#> 471 57 96 49 84 15  0
#> 472 84 96 49 84 15  0
#> 473 57 44 92 84 15  0
#> 474 84 44 92 84 15  0
#> 475 57 96 92 84 15  0
#> 476 84 96 92 84 15  0
#> 477 57 44 49 89 15  0
#> 478 84 44 49 89 15  0
#> 479 57 96 49 89 15  0
#> 480 84 96 49 89 15  0
#> 481 57 44 92 89 15  0
#> 482 84 44 92 89 15  0
#> 483 57 96 92 89 15  0
#> 485 57 44 49 84 74  0
#> 486 84 44 49 84 74  0
#> 487 57 96 49 84 74  0
#> 488 84 96 49 84 74  0
#> 489 57 44 92 84 74  0
#> 490 84 44 92 84 74  0
#> 491 57 96 92 84 74  0
#> 493 57 44 49 89 74  0
#> 494 84 44 49 89 74  0
#> 495 57 96 49 89 74  0
#> 496 84 96 49 89 74  0
#> 497 57 44 92 89 74  0
#> 498 84 44 92 89 74  0
#> 499 57 96 92 89 74  0
#> 502 84 12 60 53 22  0
#> 503 58 96 60 53 22  0
#> 504 84 96 60 53 22  0
#> 505 58 12 92 53 22  0
#> 506 84 12 92 53 22  0
#> 507 58 96 92 53 22  0
#> 508 84 96 92 53 22  0
#> 509 58 12 60 89 22  0
#> 510 84 12 60 89 22  0
#> 511 58 96 60 89 22  0
#> 512 84 96 60 89 22  0
#> 513 58 12 92 89 22  0
#> 514 84 12 92 89 22  0
#> 515 58 96 92 89 22  0
#> 517 58 12 60 53 74  0
#> 518 84 12 60 53 74  0
#> 519 58 96 60 53 74  0
#> 520 84 96 60 53 74  0
#> 521 58 12 92 53 74  0
#> 522 84 12 92 53 74  0
#> 523 58 96 92 53 74  0
#> 525 58 12 60 89 74  0
#> 526 84 12 60 89 74  0
#> 527 58 96 60 89 74  0
#> 528 84 96 60 89 74  0
#> 534 84 12 54 22 61  0
#> 535 47 96 54 22 61  0
#> 536 84 96 54 22 61  0
#> 537 47 12 92 22 61  0
#> 538 84 12 92 22 61  0
#> 539 47 96 92 22 61  0
#> 540 84 96 92 22 61  0
#> 541 47 12 54 89 61  0
#> 542 84 12 54 89 61  0
#> 543 47 96 54 89 61  0
#> 544 84 96 54 89 61  0
#> 545 47 12 92 89 61  0
#> 546 84 12 92 89 61  0
#> 547 47 96 92 89 61  0
#> 548 84 96 92 89 61  0
#> 549 47 12 54 22 74  0
#> 550 84 12 54 22 74  0
#> 551 47 96 54 22 74  0
#> 552 84 96 54 22 74  0
#> 553 47 12 92 22 74  0
#> 554 84 12 92 22 74  0
#> 555 47 96 92 22 74  0
#> 556 84 96 92 22 74  0
#> 557 47 12 54 89 74  0
#> 558 84 12 54 89 74  0
#> 559 47 96 54 89 74  0
#> 560 84 96 54 89 74  0
#> 561 47 12 92 89 74  0
#> 566 84 21 69 35 25  0
#> 567 32 96 69 35 25  0
#> 568 84 96 69 35 25  0
#> 569 32 21 92 35 25  0
#> 570 84 21 92 35 25  0
#> 571 32 96 92 35 25  0
#> 572 84 96 92 35 25  0
#> 573 32 21 69 89 25  0
#> 574 84 21 69 89 25  0
#> 575 32 96 69 89 25  0
#> 576 84 96 69 89 25  0
#> 577 32 21 92 89 25  0
#> 578 84 21 92 89 25  0
#> 579 32 96 92 89 25  0
#> 580 84 96 92 89 25  0
#> 581 32 21 69 35 74  0
#> 582 84 21 69 35 74  0
#> 583 32 96 69 35 74  0
#> 584 84 96 69 35 74  0
#> 585 32 21 92 35 74  0
#> 586 84 21 92 35 74  0
#> 587 32 96 92 35 74  0
#> 588 84 96 92 35 74  0
#> 589 32 21 69 89 74  0
#> 590 84 21 69 89 74  0
#> 591 32 96 69 89 74  0
#> 592 84 96 69 89 74  0
#> 593 32 21 92 89 74  0
#> 594 84 21 92 89 74  0
#> 595 32 96 92 89 74  0
#> 598 84 91 51 41 31  0
#> 599 33 96 51 41 31  0
#> 600 84 96 51 41 31  0
#> 601 33 91 92 41 31  0
#> 602 84 91 92 41 31  0
#> 603 33 96 92 41 31  0
#> 604 84 96 92 41 31  0
#> 605 33 91 51 89 31  0
#> 606 84 91 51 89 31  0
#> 607 33 96 51 89 31  0
#> 608 84 96 51 89 31  0
#> 609 33 91 92 89 31  0
#> 610 84 91 92 89 31  0
#> 611 33 96 92 89 31  0
#> 613 33 91 51 41 74  0
#> 614 84 91 51 41 74  0
#> 615 33 96 51 41 74  0
#> 616 84 96 51 41 74  0
#> 617 33 91 92 41 74  0
#> 618 84 91 92 41 74  0
#> 619 33 96 92 41 74  0
#> 620 84 96 92 41 74  0
#> 621 33 91 51 89 74  0
#> 622 84 91 51 89 74  0
#> 623 33 96 51 89 74  0
#> 624 84 96 51 89 74  0
#> 625 33 91 92 89 74  0
#> 626 84 91 92 89 74  0
#> 627 33 96 92 89 74  0
#> 630 84 27 62 60 19  0
#> 631 76 96 62 60 19  0
#> 632 84 96 62 60 19  0
#> 633 76 27 92 60 19  0
#> 634 84 27 92 60 19  0
#> 635 76 96 92 60 19  0
#> 636 84 96 92 60 19  0
#> 637 76 27 62 89 19  0
#> 638 84 27 62 89 19  0
#> 639 76 96 62 89 19  0
#> 640 84 96 62 89 19  0
#> 641 76 27 92 89 19  0
#> 642 84 27 92 89 19  0
#> 643 76 96 92 89 19  0
#> 644 84 96 92 89 19  0
#> 645 76 27 62 60 74  0
#> 646 84 27 62 60 74  0
#> 647 76 96 62 60 74  0
#> 648 84 96 62 60 74  0
#> 649 76 27 92 60 74  0
#> 650 84 27 92 60 74  0
#> 651 76 96 92 60 74  0
#> 653 76 27 62 89 74  0
#> 654 84 27 62 89 74  0
#> 655 76 96 62 89 74  0
#> 656 84 96 62 89 74  0
#> 657 76 27 92 89 74  0
#> 658 84 27 92 89 74  0
#> 659 76 96 92 89 74  0