Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 10  2 60 45  1
#> 5   3 37 54  1
#> 4  20 81 25  1
#> 2  41 76  9  1
#> 7  42 11 79  1
#> 1  45 45  2  1
#> 9  46  3 65  1
#> 21 51 60 45  0
#> 3   2 86 45  0
#> 41 51 86 45  0
#> 51  2 60 84  0
#> 6  51 60 84  0
#> 71  2 86 84  0
#> 8  51 86 84  0
#> 22 51 37 54  0
#> 31  3 86 54  0
#> 42 51 86 54  0
#> 52  3 37 84  0
#> 61 51 37 84  0
#> 72  3 86 84  0
#> 23 51 81 25  0
#> 32 20 86 25  0
#> 43 51 86 25  0
#> 53 20 81 84  0
#> 62 51 81 84  0
#> 73 20 86 84  0
#> 24 51 76  9  0
#> 33 41 86  9  0
#> 44 51 86  9  0
#> 54 41 76 84  0
#> 63 51 76 84  0
#> 74 41 86 84  0
#> 25 51 11 79  0
#> 34 42 86 79  0
#> 45 51 86 79  0
#> 55 42 11 84  0
#> 64 51 11 84  0
#> 75 42 86 84  0
#> 26 51 45  2  0
#> 35 45 86  2  0
#> 46 51 86  2  0
#> 56 45 45 84  0
#> 65 51 45 84  0
#> 76 45 86 84  0
#> 27 51  3 65  0
#> 36 46 86 65  0
#> 47 51 86 65  0
#> 57 46  3 84  0
#> 66 51  3 84  0
#> 77 46 86 84  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 10  2 60 45  1
#> 5   3 37 54  1
#> 4  20 81 25  1
#> 2  41 76  9  1
#> 7  42 11 79  1
#> 1  45 45  2  1
#> 9  46  3 65  1
#> 21 51 60 45  0
#> 3   2 -2 45  0
#> 41 51 -2 45  0
#> 51  2 60 84  0
#> 6  51 60 84  0
#> 71  2 -2 84  0
#> 8  51 -2 84  0
#> 22 51 37 54  0
#> 31  3 -2 54  0
#> 42 51 -2 54  0
#> 52  3 37 84  0
#> 61 51 37 84  0
#> 72  3 -2 84  0
#> 23 51 81 25  0
#> 32 20 -2 25  0
#> 43 51 -2 25  0
#> 53 20 81 84  0
#> 62 51 81 84  0
#> 73 20 -2 84  0
#> 24 51 76  9  0
#> 33 41 -2  9  0
#> 44 51 -2  9  0
#> 54 41 76 84  0
#> 63 51 76 84  0
#> 74 41 -2 84  0
#> 25 51 11 79  0
#> 34 42 -2 79  0
#> 45 51 -2 79  0
#> 55 42 11 84  0
#> 64 51 11 84  0
#> 75 42 -2 84  0
#> 26 51 45  2  0
#> 35 45 -2  2  0
#> 46 51 -2  2  0
#> 56 45 45 84  0
#> 65 51 45 84  0
#> 76 45 -2 84  0
#> 27 51  3 65  0
#> 36 46 -2 65  0
#> 47 51 -2 65  0
#> 57 46  3 84  0
#> 66 51  3 84  0
#> 77 46 -2 84  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 10  2 60  45  1
#> 5   3 37  54  1
#> 4  20 81  25  1
#> 2  41 76   9  1
#> 7  42 11  79  1
#> 1  45 45   2  1
#> 9  46  3  65  1
#> 21  0 60  45  0
#> 3   2  0  45  0
#> 41  0  0  45  0
#> 51  2 60 100  0
#> 6   0 60 100  0
#> 71  2  0 100  0
#> 8   0  0 100  0
#> 22  0 37  54  0
#> 31  3  0  54  0
#> 42  0  0  54  0
#> 52  3 37 100  0
#> 61  0 37 100  0
#> 72  3  0 100  0
#> 23  0 81  25  0
#> 32 20  0  25  0
#> 43  0  0  25  0
#> 53 20 81 100  0
#> 62  0 81 100  0
#> 73 20  0 100  0
#> 24  0 76   9  0
#> 33 41  0   9  0
#> 44  0  0   9  0
#> 54 41 76 100  0
#> 63  0 76 100  0
#> 74 41  0 100  0
#> 25  0 11  79  0
#> 34 42  0  79  0
#> 45  0  0  79  0
#> 55 42 11 100  0
#> 64  0 11 100  0
#> 75 42  0 100  0
#> 26  0 45   2  0
#> 35 45  0   2  0
#> 46  0  0   2  0
#> 56 45 45 100  0
#> 65  0 45 100  0
#> 76 45  0 100  0
#> 27  0  3  65  0
#> 36 46  0  65  0
#> 47  0  0  65  0
#> 57 46  3 100  0
#> 66  0  3 100  0
#> 77 46  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   57 32 58 59  7  1
#> 2   67 20 18 39 55  1
#> 3   66 77 18 31 43  1
#> 4   18 18 54 34 44  1
#> 5   44 31  7 64 52  1
#> 6   19 35 25 39 71  1
#> 7   79 55 28 32 23  1
#> 8   30 25 66 27 26  1
#> 9   48 39 70 76 14  1
#> 10  32 39 37  9 45  1
#> 11  54 25 21 23 62  1
#> 12  19 33 48 76 28  1
#> 13  10 38 29 41 60  1
#> 14  47 65 15 30 74  1
#> 15  48 58  2 54 53  1
#> 16  10 54 46 56 23  1
#> 17  38 34 27 16 68  1
#> 18  40  9 60 43 73  1
#> 19  80 23 49 29 70  1
#> 20  59 30 15 65 28  1
#> 22  85 32 58 59  7  0
#> 23  57 82 58 59  7  0
#> 24  85 82 58 59  7  0
#> 25  57 32 75 59  7  0
#> 26  85 32 75 59  7  0
#> 27  57 82 75 59  7  0
#> 28  85 82 75 59  7  0
#> 29  57 32 58 81  7  0
#> 30  85 32 58 81  7  0
#> 31  57 82 58 81  7  0
#> 32  85 82 58 81  7  0
#> 33  57 32 75 81  7  0
#> 34  85 32 75 81  7  0
#> 35  57 82 75 81  7  0
#> 36  85 82 75 81  7  0
#> 37  57 32 58 59 79  0
#> 38  85 32 58 59 79  0
#> 39  57 82 58 59 79  0
#> 40  85 82 58 59 79  0
#> 41  57 32 75 59 79  0
#> 42  85 32 75 59 79  0
#> 43  57 82 75 59 79  0
#> 44  85 82 75 59 79  0
#> 45  57 32 58 81 79  0
#> 46  85 32 58 81 79  0
#> 47  57 82 58 81 79  0
#> 48  85 82 58 81 79  0
#> 49  57 32 75 81 79  0
#> 50  85 32 75 81 79  0
#> 51  57 82 75 81 79  0
#> 52  85 82 75 81 79  0
#> 54  85 20 18 39 55  0
#> 55  67 82 18 39 55  0
#> 56  85 82 18 39 55  0
#> 57  67 20 75 39 55  0
#> 58  85 20 75 39 55  0
#> 59  67 82 75 39 55  0
#> 60  85 82 75 39 55  0
#> 61  67 20 18 81 55  0
#> 62  85 20 18 81 55  0
#> 63  67 82 18 81 55  0
#> 64  85 82 18 81 55  0
#> 65  67 20 75 81 55  0
#> 66  85 20 75 81 55  0
#> 67  67 82 75 81 55  0
#> 68  85 82 75 81 55  0
#> 69  67 20 18 39 79  0
#> 70  85 20 18 39 79  0
#> 71  67 82 18 39 79  0
#> 72  85 82 18 39 79  0
#> 73  67 20 75 39 79  0
#> 74  85 20 75 39 79  0
#> 75  67 82 75 39 79  0
#> 76  85 82 75 39 79  0
#> 77  67 20 18 81 79  0
#> 78  85 20 18 81 79  0
#> 79  67 82 18 81 79  0
#> 80  85 82 18 81 79  0
#> 81  67 20 75 81 79  0
#> 82  85 20 75 81 79  0
#> 83  67 82 75 81 79  0
#> 86  85 77 18 31 43  0
#> 87  66 82 18 31 43  0
#> 88  85 82 18 31 43  0
#> 89  66 77 75 31 43  0
#> 90  85 77 75 31 43  0
#> 91  66 82 75 31 43  0
#> 92  85 82 75 31 43  0
#> 93  66 77 18 81 43  0
#> 94  85 77 18 81 43  0
#> 95  66 82 18 81 43  0
#> 96  85 82 18 81 43  0
#> 97  66 77 75 81 43  0
#> 98  85 77 75 81 43  0
#> 99  66 82 75 81 43  0
#> 100 85 82 75 81 43  0
#> 101 66 77 18 31 79  0
#> 102 85 77 18 31 79  0
#> 103 66 82 18 31 79  0
#> 104 85 82 18 31 79  0
#> 105 66 77 75 31 79  0
#> 106 85 77 75 31 79  0
#> 107 66 82 75 31 79  0
#> 108 85 82 75 31 79  0
#> 109 66 77 18 81 79  0
#> 110 85 77 18 81 79  0
#> 111 66 82 18 81 79  0
#> 113 66 77 75 81 79  0
#> 114 85 77 75 81 79  0
#> 115 66 82 75 81 79  0
#> 118 85 18 54 34 44  0
#> 119 18 82 54 34 44  0
#> 120 85 82 54 34 44  0
#> 121 18 18 75 34 44  0
#> 122 85 18 75 34 44  0
#> 123 18 82 75 34 44  0
#> 124 85 82 75 34 44  0
#> 125 18 18 54 81 44  0
#> 126 85 18 54 81 44  0
#> 127 18 82 54 81 44  0
#> 128 85 82 54 81 44  0
#> 129 18 18 75 81 44  0
#> 130 85 18 75 81 44  0
#> 131 18 82 75 81 44  0
#> 132 85 82 75 81 44  0
#> 133 18 18 54 34 79  0
#> 134 85 18 54 34 79  0
#> 135 18 82 54 34 79  0
#> 136 85 82 54 34 79  0
#> 137 18 18 75 34 79  0
#> 138 85 18 75 34 79  0
#> 139 18 82 75 34 79  0
#> 140 85 82 75 34 79  0
#> 141 18 18 54 81 79  0
#> 142 85 18 54 81 79  0
#> 143 18 82 54 81 79  0
#> 144 85 82 54 81 79  0
#> 145 18 18 75 81 79  0
#> 146 85 18 75 81 79  0
#> 147 18 82 75 81 79  0
#> 150 85 31  7 64 52  0
#> 151 44 82  7 64 52  0
#> 152 85 82  7 64 52  0
#> 153 44 31 75 64 52  0
#> 154 85 31 75 64 52  0
#> 155 44 82 75 64 52  0
#> 156 85 82 75 64 52  0
#> 157 44 31  7 81 52  0
#> 158 85 31  7 81 52  0
#> 159 44 82  7 81 52  0
#> 160 85 82  7 81 52  0
#> 161 44 31 75 81 52  0
#> 162 85 31 75 81 52  0
#> 163 44 82 75 81 52  0
#> 164 85 82 75 81 52  0
#> 165 44 31  7 64 79  0
#> 166 85 31  7 64 79  0
#> 167 44 82  7 64 79  0
#> 168 85 82  7 64 79  0
#> 169 44 31 75 64 79  0
#> 170 85 31 75 64 79  0
#> 171 44 82 75 64 79  0
#> 172 85 82 75 64 79  0
#> 173 44 31  7 81 79  0
#> 174 85 31  7 81 79  0
#> 175 44 82  7 81 79  0
#> 176 85 82  7 81 79  0
#> 177 44 31 75 81 79  0
#> 178 85 31 75 81 79  0
#> 179 44 82 75 81 79  0
#> 182 85 35 25 39 71  0
#> 183 19 82 25 39 71  0
#> 184 85 82 25 39 71  0
#> 185 19 35 75 39 71  0
#> 186 85 35 75 39 71  0
#> 187 19 82 75 39 71  0
#> 188 85 82 75 39 71  0
#> 189 19 35 25 81 71  0
#> 190 85 35 25 81 71  0
#> 191 19 82 25 81 71  0
#> 192 85 82 25 81 71  0
#> 193 19 35 75 81 71  0
#> 194 85 35 75 81 71  0
#> 195 19 82 75 81 71  0
#> 196 85 82 75 81 71  0
#> 197 19 35 25 39 79  0
#> 198 85 35 25 39 79  0
#> 199 19 82 25 39 79  0
#> 200 85 82 25 39 79  0
#> 201 19 35 75 39 79  0
#> 202 85 35 75 39 79  0
#> 203 19 82 75 39 79  0
#> 205 19 35 25 81 79  0
#> 206 85 35 25 81 79  0
#> 207 19 82 25 81 79  0
#> 208 85 82 25 81 79  0
#> 209 19 35 75 81 79  0
#> 210 85 35 75 81 79  0
#> 211 19 82 75 81 79  0
#> 214 85 55 28 32 23  0
#> 215 79 82 28 32 23  0
#> 216 85 82 28 32 23  0
#> 217 79 55 75 32 23  0
#> 218 85 55 75 32 23  0
#> 219 79 82 75 32 23  0
#> 220 85 82 75 32 23  0
#> 221 79 55 28 81 23  0
#> 222 85 55 28 81 23  0
#> 223 79 82 28 81 23  0
#> 224 85 82 28 81 23  0
#> 225 79 55 75 81 23  0
#> 226 85 55 75 81 23  0
#> 227 79 82 75 81 23  0
#> 228 85 82 75 81 23  0
#> 229 79 55 28 32 79  0
#> 230 85 55 28 32 79  0
#> 231 79 82 28 32 79  0
#> 232 85 82 28 32 79  0
#> 233 79 55 75 32 79  0
#> 234 85 55 75 32 79  0
#> 235 79 82 75 32 79  0
#> 236 85 82 75 32 79  0
#> 237 79 55 28 81 79  0
#> 238 85 55 28 81 79  0
#> 239 79 82 28 81 79  0
#> 240 85 82 28 81 79  0
#> 241 79 55 75 81 79  0
#> 242 85 55 75 81 79  0
#> 243 79 82 75 81 79  0
#> 246 85 25 66 27 26  0
#> 247 30 82 66 27 26  0
#> 248 85 82 66 27 26  0
#> 249 30 25 75 27 26  0
#> 250 85 25 75 27 26  0
#> 251 30 82 75 27 26  0
#> 252 85 82 75 27 26  0
#> 253 30 25 66 81 26  0
#> 254 85 25 66 81 26  0
#> 255 30 82 66 81 26  0
#> 256 85 82 66 81 26  0
#> 257 30 25 75 81 26  0
#> 258 85 25 75 81 26  0
#> 259 30 82 75 81 26  0
#> 260 85 82 75 81 26  0
#> 261 30 25 66 27 79  0
#> 262 85 25 66 27 79  0
#> 263 30 82 66 27 79  0
#> 264 85 82 66 27 79  0
#> 265 30 25 75 27 79  0
#> 266 85 25 75 27 79  0
#> 267 30 82 75 27 79  0
#> 268 85 82 75 27 79  0
#> 269 30 25 66 81 79  0
#> 270 85 25 66 81 79  0
#> 271 30 82 66 81 79  0
#> 272 85 82 66 81 79  0
#> 273 30 25 75 81 79  0
#> 274 85 25 75 81 79  0
#> 275 30 82 75 81 79  0
#> 278 85 39 70 76 14  0
#> 279 48 82 70 76 14  0
#> 280 85 82 70 76 14  0
#> 281 48 39 75 76 14  0
#> 282 85 39 75 76 14  0
#> 283 48 82 75 76 14  0
#> 284 85 82 75 76 14  0
#> 285 48 39 70 81 14  0
#> 286 85 39 70 81 14  0
#> 287 48 82 70 81 14  0
#> 288 85 82 70 81 14  0
#> 289 48 39 75 81 14  0
#> 290 85 39 75 81 14  0
#> 291 48 82 75 81 14  0
#> 292 85 82 75 81 14  0
#> 293 48 39 70 76 79  0
#> 294 85 39 70 76 79  0
#> 295 48 82 70 76 79  0
#> 296 85 82 70 76 79  0
#> 297 48 39 75 76 79  0
#> 298 85 39 75 76 79  0
#> 299 48 82 75 76 79  0
#> 300 85 82 75 76 79  0
#> 301 48 39 70 81 79  0
#> 302 85 39 70 81 79  0
#> 303 48 82 70 81 79  0
#> 304 85 82 70 81 79  0
#> 305 48 39 75 81 79  0
#> 306 85 39 75 81 79  0
#> 307 48 82 75 81 79  0
#> 310 85 39 37  9 45  0
#> 311 32 82 37  9 45  0
#> 312 85 82 37  9 45  0
#> 313 32 39 75  9 45  0
#> 314 85 39 75  9 45  0
#> 315 32 82 75  9 45  0
#> 316 85 82 75  9 45  0
#> 317 32 39 37 81 45  0
#> 318 85 39 37 81 45  0
#> 319 32 82 37 81 45  0
#> 320 85 82 37 81 45  0
#> 321 32 39 75 81 45  0
#> 322 85 39 75 81 45  0
#> 323 32 82 75 81 45  0
#> 324 85 82 75 81 45  0
#> 325 32 39 37  9 79  0
#> 326 85 39 37  9 79  0
#> 327 32 82 37  9 79  0
#> 328 85 82 37  9 79  0
#> 329 32 39 75  9 79  0
#> 330 85 39 75  9 79  0
#> 331 32 82 75  9 79  0
#> 332 85 82 75  9 79  0
#> 333 32 39 37 81 79  0
#> 334 85 39 37 81 79  0
#> 335 32 82 37 81 79  0
#> 336 85 82 37 81 79  0
#> 337 32 39 75 81 79  0
#> 339 32 82 75 81 79  0
#> 342 85 25 21 23 62  0
#> 343 54 82 21 23 62  0
#> 344 85 82 21 23 62  0
#> 345 54 25 75 23 62  0
#> 346 85 25 75 23 62  0
#> 347 54 82 75 23 62  0
#> 348 85 82 75 23 62  0
#> 349 54 25 21 81 62  0
#> 350 85 25 21 81 62  0
#> 351 54 82 21 81 62  0
#> 352 85 82 21 81 62  0
#> 353 54 25 75 81 62  0
#> 354 85 25 75 81 62  0
#> 355 54 82 75 81 62  0
#> 356 85 82 75 81 62  0
#> 357 54 25 21 23 79  0
#> 358 85 25 21 23 79  0
#> 359 54 82 21 23 79  0
#> 360 85 82 21 23 79  0
#> 361 54 25 75 23 79  0
#> 362 85 25 75 23 79  0
#> 363 54 82 75 23 79  0
#> 364 85 82 75 23 79  0
#> 365 54 25 21 81 79  0
#> 366 85 25 21 81 79  0
#> 367 54 82 21 81 79  0
#> 368 85 82 21 81 79  0
#> 369 54 25 75 81 79  0
#> 371 54 82 75 81 79  0
#> 374 85 33 48 76 28  0
#> 375 19 82 48 76 28  0
#> 376 85 82 48 76 28  0
#> 377 19 33 75 76 28  0
#> 378 85 33 75 76 28  0
#> 379 19 82 75 76 28  0
#> 380 85 82 75 76 28  0
#> 381 19 33 48 81 28  0
#> 382 85 33 48 81 28  0
#> 383 19 82 48 81 28  0
#> 384 85 82 48 81 28  0
#> 385 19 33 75 81 28  0
#> 386 85 33 75 81 28  0
#> 387 19 82 75 81 28  0
#> 388 85 82 75 81 28  0
#> 389 19 33 48 76 79  0
#> 390 85 33 48 76 79  0
#> 391 19 82 48 76 79  0
#> 392 85 82 48 76 79  0
#> 393 19 33 75 76 79  0
#> 394 85 33 75 76 79  0
#> 395 19 82 75 76 79  0
#> 397 19 33 48 81 79  0
#> 398 85 33 48 81 79  0
#> 399 19 82 48 81 79  0
#> 400 85 82 48 81 79  0
#> 401 19 33 75 81 79  0
#> 402 85 33 75 81 79  0
#> 406 85 38 29 41 60  0
#> 407 10 82 29 41 60  0
#> 408 85 82 29 41 60  0
#> 409 10 38 75 41 60  0
#> 410 85 38 75 41 60  0
#> 411 10 82 75 41 60  0
#> 412 85 82 75 41 60  0
#> 413 10 38 29 81 60  0
#> 414 85 38 29 81 60  0
#> 415 10 82 29 81 60  0
#> 416 85 82 29 81 60  0
#> 417 10 38 75 81 60  0
#> 418 85 38 75 81 60  0
#> 419 10 82 75 81 60  0
#> 420 85 82 75 81 60  0
#> 421 10 38 29 41 79  0
#> 422 85 38 29 41 79  0
#> 423 10 82 29 41 79  0
#> 424 85 82 29 41 79  0
#> 425 10 38 75 41 79  0
#> 426 85 38 75 41 79  0
#> 427 10 82 75 41 79  0
#> 428 85 82 75 41 79  0
#> 429 10 38 29 81 79  0
#> 430 85 38 29 81 79  0
#> 431 10 82 29 81 79  0
#> 432 85 82 29 81 79  0
#> 433 10 38 75 81 79  0
#> 434 85 38 75 81 79  0
#> 435 10 82 75 81 79  0
#> 438 85 65 15 30 74  0
#> 439 47 82 15 30 74  0
#> 440 85 82 15 30 74  0
#> 441 47 65 75 30 74  0
#> 442 85 65 75 30 74  0
#> 443 47 82 75 30 74  0
#> 444 85 82 75 30 74  0
#> 445 47 65 15 81 74  0
#> 446 85 65 15 81 74  0
#> 447 47 82 15 81 74  0
#> 448 85 82 15 81 74  0
#> 449 47 65 75 81 74  0
#> 450 85 65 75 81 74  0
#> 451 47 82 75 81 74  0
#> 452 85 82 75 81 74  0
#> 453 47 65 15 30 79  0
#> 454 85 65 15 30 79  0
#> 455 47 82 15 30 79  0
#> 456 85 82 15 30 79  0
#> 457 47 65 75 30 79  0
#> 458 85 65 75 30 79  0
#> 459 47 82 75 30 79  0
#> 460 85 82 75 30 79  0
#> 461 47 65 15 81 79  0
#> 462 85 65 15 81 79  0
#> 463 47 82 15 81 79  0
#> 464 85 82 15 81 79  0
#> 465 47 65 75 81 79  0
#> 466 85 65 75 81 79  0
#> 467 47 82 75 81 79  0
#> 470 85 58  2 54 53  0
#> 471 48 82  2 54 53  0
#> 472 85 82  2 54 53  0
#> 473 48 58 75 54 53  0
#> 474 85 58 75 54 53  0
#> 475 48 82 75 54 53  0
#> 476 85 82 75 54 53  0
#> 477 48 58  2 81 53  0
#> 478 85 58  2 81 53  0
#> 479 48 82  2 81 53  0
#> 480 85 82  2 81 53  0
#> 481 48 58 75 81 53  0
#> 482 85 58 75 81 53  0
#> 483 48 82 75 81 53  0
#> 484 85 82 75 81 53  0
#> 485 48 58  2 54 79  0
#> 486 85 58  2 54 79  0
#> 487 48 82  2 54 79  0
#> 488 85 82  2 54 79  0
#> 489 48 58 75 54 79  0
#> 490 85 58 75 54 79  0
#> 491 48 82 75 54 79  0
#> 492 85 82 75 54 79  0
#> 493 48 58  2 81 79  0
#> 494 85 58  2 81 79  0
#> 495 48 82  2 81 79  0
#> 496 85 82  2 81 79  0
#> 497 48 58 75 81 79  0
#> 498 85 58 75 81 79  0
#> 502 85 54 46 56 23  0
#> 503 10 82 46 56 23  0
#> 504 85 82 46 56 23  0
#> 505 10 54 75 56 23  0
#> 506 85 54 75 56 23  0
#> 507 10 82 75 56 23  0
#> 508 85 82 75 56 23  0
#> 509 10 54 46 81 23  0
#> 510 85 54 46 81 23  0
#> 511 10 82 46 81 23  0
#> 512 85 82 46 81 23  0
#> 513 10 54 75 81 23  0
#> 514 85 54 75 81 23  0
#> 515 10 82 75 81 23  0
#> 517 10 54 46 56 79  0
#> 518 85 54 46 56 79  0
#> 519 10 82 46 56 79  0
#> 520 85 82 46 56 79  0
#> 521 10 54 75 56 79  0
#> 522 85 54 75 56 79  0
#> 523 10 82 75 56 79  0
#> 524 85 82 75 56 79  0
#> 525 10 54 46 81 79  0
#> 526 85 54 46 81 79  0
#> 527 10 82 46 81 79  0
#> 528 85 82 46 81 79  0
#> 529 10 54 75 81 79  0
#> 530 85 54 75 81 79  0
#> 534 85 34 27 16 68  0
#> 535 38 82 27 16 68  0
#> 536 85 82 27 16 68  0
#> 537 38 34 75 16 68  0
#> 538 85 34 75 16 68  0
#> 539 38 82 75 16 68  0
#> 540 85 82 75 16 68  0
#> 541 38 34 27 81 68  0
#> 542 85 34 27 81 68  0
#> 543 38 82 27 81 68  0
#> 544 85 82 27 81 68  0
#> 545 38 34 75 81 68  0
#> 546 85 34 75 81 68  0
#> 547 38 82 75 81 68  0
#> 548 85 82 75 81 68  0
#> 549 38 34 27 16 79  0
#> 550 85 34 27 16 79  0
#> 551 38 82 27 16 79  0
#> 552 85 82 27 16 79  0
#> 553 38 34 75 16 79  0
#> 554 85 34 75 16 79  0
#> 555 38 82 75 16 79  0
#> 556 85 82 75 16 79  0
#> 557 38 34 27 81 79  0
#> 558 85 34 27 81 79  0
#> 559 38 82 27 81 79  0
#> 560 85 82 27 81 79  0
#> 561 38 34 75 81 79  0
#> 562 85 34 75 81 79  0
#> 563 38 82 75 81 79  0
#> 566 85  9 60 43 73  0
#> 567 40 82 60 43 73  0
#> 568 85 82 60 43 73  0
#> 569 40  9 75 43 73  0
#> 570 85  9 75 43 73  0
#> 571 40 82 75 43 73  0
#> 572 85 82 75 43 73  0
#> 573 40  9 60 81 73  0
#> 574 85  9 60 81 73  0
#> 575 40 82 60 81 73  0
#> 576 85 82 60 81 73  0
#> 577 40  9 75 81 73  0
#> 578 85  9 75 81 73  0
#> 579 40 82 75 81 73  0
#> 580 85 82 75 81 73  0
#> 581 40  9 60 43 79  0
#> 582 85  9 60 43 79  0
#> 583 40 82 60 43 79  0
#> 584 85 82 60 43 79  0
#> 585 40  9 75 43 79  0
#> 586 85  9 75 43 79  0
#> 587 40 82 75 43 79  0
#> 588 85 82 75 43 79  0
#> 589 40  9 60 81 79  0
#> 590 85  9 60 81 79  0
#> 591 40 82 60 81 79  0
#> 592 85 82 60 81 79  0
#> 593 40  9 75 81 79  0
#> 594 85  9 75 81 79  0
#> 595 40 82 75 81 79  0
#> 598 85 23 49 29 70  0
#> 599 80 82 49 29 70  0
#> 600 85 82 49 29 70  0
#> 601 80 23 75 29 70  0
#> 602 85 23 75 29 70  0
#> 603 80 82 75 29 70  0
#> 604 85 82 75 29 70  0
#> 605 80 23 49 81 70  0
#> 606 85 23 49 81 70  0
#> 607 80 82 49 81 70  0
#> 608 85 82 49 81 70  0
#> 609 80 23 75 81 70  0
#> 610 85 23 75 81 70  0
#> 611 80 82 75 81 70  0
#> 612 85 82 75 81 70  0
#> 613 80 23 49 29 79  0
#> 614 85 23 49 29 79  0
#> 615 80 82 49 29 79  0
#> 616 85 82 49 29 79  0
#> 617 80 23 75 29 79  0
#> 618 85 23 75 29 79  0
#> 619 80 82 75 29 79  0
#> 620 85 82 75 29 79  0
#> 621 80 23 49 81 79  0
#> 622 85 23 49 81 79  0
#> 623 80 82 49 81 79  0
#> 624 85 82 49 81 79  0
#> 625 80 23 75 81 79  0
#> 626 85 23 75 81 79  0
#> 627 80 82 75 81 79  0
#> 630 85 30 15 65 28  0
#> 631 59 82 15 65 28  0
#> 632 85 82 15 65 28  0
#> 633 59 30 75 65 28  0
#> 634 85 30 75 65 28  0
#> 635 59 82 75 65 28  0
#> 636 85 82 75 65 28  0
#> 637 59 30 15 81 28  0
#> 638 85 30 15 81 28  0
#> 639 59 82 15 81 28  0
#> 640 85 82 15 81 28  0
#> 641 59 30 75 81 28  0
#> 642 85 30 75 81 28  0
#> 643 59 82 75 81 28  0
#> 645 59 30 15 65 79  0
#> 646 85 30 15 65 79  0
#> 647 59 82 15 65 79  0
#> 648 85 82 15 65 79  0
#> 649 59 30 75 65 79  0
#> 650 85 30 75 65 79  0
#> 651 59 82 75 65 79  0
#> 652 85 82 75 65 79  0
#> 653 59 30 15 81 79  0
#> 654 85 30 15 81 79  0
#> 655 59 82 15 81 79  0
#> 657 59 30 75 81 79  0
#> 658 85 30 75 81 79  0
#> 659 59 82 75 81 79  0