Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 5  11 59 22  1
#> 4  16 24 74  1
#> 2  33 28 10  1
#> 8  36 73  9  1
#> 10 41  2 47  1
#> 3  44 48  2  1
#> 6  88 20 38  1
#> 21 93 59 22  0
#> 31 11 78 22  0
#> 41 93 78 22  0
#> 51 11 59 79  0
#> 61 93 59 79  0
#> 7  11 78 79  0
#> 81 93 78 79  0
#> 22 93 24 74  0
#> 32 16 78 74  0
#> 42 93 78 74  0
#> 52 16 24 79  0
#> 62 93 24 79  0
#> 71 16 78 79  0
#> 23 93 28 10  0
#> 33 33 78 10  0
#> 43 93 78 10  0
#> 53 33 28 79  0
#> 63 93 28 79  0
#> 72 33 78 79  0
#> 24 93 73  9  0
#> 34 36 78  9  0
#> 44 93 78  9  0
#> 54 36 73 79  0
#> 64 93 73 79  0
#> 73 36 78 79  0
#> 25 93  2 47  0
#> 35 41 78 47  0
#> 45 93 78 47  0
#> 55 41  2 79  0
#> 65 93  2 79  0
#> 74 41 78 79  0
#> 26 93 48  2  0
#> 36 44 78  2  0
#> 46 93 78  2  0
#> 56 44 48 79  0
#> 66 93 48 79  0
#> 75 44 78 79  0
#> 27 93 20 38  0
#> 37 88 78 38  0
#> 47 93 78 38  0
#> 57 88 20 79  0
#> 67 93 20 79  0
#> 76 88 78 79  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 5  11 59 22  1
#> 4  16 24 74  1
#> 2  33 28 10  1
#> 8  36 73  9  1
#> 10 41  2 47  1
#> 3  44 48  2  1
#> 6  88 20 38  1
#> 21 93 59 22  0
#> 31 11 -3 22  0
#> 41 93 -3 22  0
#> 51 11 59 79  0
#> 61 93 59 79  0
#> 7  11 -3 79  0
#> 81 93 -3 79  0
#> 22 93 24 74  0
#> 32 16 -3 74  0
#> 42 93 -3 74  0
#> 52 16 24 79  0
#> 62 93 24 79  0
#> 71 16 -3 79  0
#> 23 93 28 10  0
#> 33 33 -3 10  0
#> 43 93 -3 10  0
#> 53 33 28 79  0
#> 63 93 28 79  0
#> 72 33 -3 79  0
#> 24 93 73  9  0
#> 34 36 -3  9  0
#> 44 93 -3  9  0
#> 54 36 73 79  0
#> 64 93 73 79  0
#> 73 36 -3 79  0
#> 25 93  2 47  0
#> 35 41 -3 47  0
#> 45 93 -3 47  0
#> 55 41  2 79  0
#> 65 93  2 79  0
#> 74 41 -3 79  0
#> 26 93 48  2  0
#> 36 44 -3  2  0
#> 46 93 -3  2  0
#> 56 44 48 79  0
#> 66 93 48 79  0
#> 75 44 -3 79  0
#> 27 93 20 38  0
#> 37 88 -3 38  0
#> 47 93 -3 38  0
#> 57 88 20 79  0
#> 67 93 20 79  0
#> 76 88 -3 79  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 5  11 59  22  1
#> 4  16 24  74  1
#> 2  33 28  10  1
#> 8  36 73   9  1
#> 10 41  2  47  1
#> 3  44 48   2  1
#> 6  88 20  38  1
#> 21  0 59  22  0
#> 31 11  0  22  0
#> 41  0  0  22  0
#> 51 11 59 100  0
#> 61  0 59 100  0
#> 7  11  0 100  0
#> 81  0  0 100  0
#> 22  0 24  74  0
#> 32 16  0  74  0
#> 42  0  0  74  0
#> 52 16 24 100  0
#> 62  0 24 100  0
#> 71 16  0 100  0
#> 23  0 28  10  0
#> 33 33  0  10  0
#> 43  0  0  10  0
#> 53 33 28 100  0
#> 63  0 28 100  0
#> 72 33  0 100  0
#> 24  0 73   9  0
#> 34 36  0   9  0
#> 44  0  0   9  0
#> 54 36 73 100  0
#> 64  0 73 100  0
#> 73 36  0 100  0
#> 25  0  2  47  0
#> 35 41  0  47  0
#> 45  0  0  47  0
#> 55 41  2 100  0
#> 65  0  2 100  0
#> 74 41  0 100  0
#> 26  0 48   2  0
#> 36 44  0   2  0
#> 46  0  0   2  0
#> 56 44 48 100  0
#> 66  0 48 100  0
#> 75 44  0 100  0
#> 27  0 20  38  0
#> 37 88  0  38  0
#> 47  0  0  38  0
#> 57 88 20 100  0
#> 67  0 20 100  0
#> 76 88  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   39 59 45 12 77  1
#> 2   30 91 44 35 41  1
#> 3    6 52 53 30 45  1
#> 4   22 55 85 31 54  1
#> 5   36 51 74 16 28  1
#> 6   16 23 57 45 70  1
#> 7   38 20 16 38 44  1
#> 8   85 48 18 50 37  1
#> 9   41 88 24 34 52  1
#> 10  44 36 67 57  7  1
#> 11  68 29 33 54 14  1
#> 12  25 66 18 72 56  1
#> 13  68 81 31 31 30  1
#> 14  10 60 37 28 54  1
#> 15  14 60 61 48 20  1
#> 16  76 19 62 29 65  1
#> 17  54 52 25 13 31  1
#> 18  48 54  4 66 50  1
#> 19  35 45  5 48 40  1
#> 20  61 51 46  7 30  1
#> 22  90 59 45 12 77  0
#> 23  39 96 45 12 77  0
#> 24  90 96 45 12 77  0
#> 25  39 59 90 12 77  0
#> 26  90 59 90 12 77  0
#> 27  39 96 90 12 77  0
#> 28  90 96 90 12 77  0
#> 29  39 59 45 77 77  0
#> 30  90 59 45 77 77  0
#> 31  39 96 45 77 77  0
#> 32  90 96 45 77 77  0
#> 33  39 59 90 77 77  0
#> 34  90 59 90 77 77  0
#> 35  39 96 90 77 77  0
#> 36  90 96 90 77 77  0
#> 37  39 59 45 12 82  0
#> 38  90 59 45 12 82  0
#> 39  39 96 45 12 82  0
#> 40  90 96 45 12 82  0
#> 41  39 59 90 12 82  0
#> 42  90 59 90 12 82  0
#> 43  39 96 90 12 82  0
#> 44  90 96 90 12 82  0
#> 45  39 59 45 77 82  0
#> 46  90 59 45 77 82  0
#> 47  39 96 45 77 82  0
#> 48  90 96 45 77 82  0
#> 49  39 59 90 77 82  0
#> 50  90 59 90 77 82  0
#> 51  39 96 90 77 82  0
#> 52  90 96 90 77 82  0
#> 54  90 91 44 35 41  0
#> 55  30 96 44 35 41  0
#> 56  90 96 44 35 41  0
#> 57  30 91 90 35 41  0
#> 58  90 91 90 35 41  0
#> 59  30 96 90 35 41  0
#> 60  90 96 90 35 41  0
#> 61  30 91 44 77 41  0
#> 62  90 91 44 77 41  0
#> 63  30 96 44 77 41  0
#> 64  90 96 44 77 41  0
#> 65  30 91 90 77 41  0
#> 66  90 91 90 77 41  0
#> 67  30 96 90 77 41  0
#> 68  90 96 90 77 41  0
#> 69  30 91 44 35 82  0
#> 70  90 91 44 35 82  0
#> 71  30 96 44 35 82  0
#> 72  90 96 44 35 82  0
#> 73  30 91 90 35 82  0
#> 74  90 91 90 35 82  0
#> 75  30 96 90 35 82  0
#> 76  90 96 90 35 82  0
#> 77  30 91 44 77 82  0
#> 78  90 91 44 77 82  0
#> 79  30 96 44 77 82  0
#> 80  90 96 44 77 82  0
#> 81  30 91 90 77 82  0
#> 82  90 91 90 77 82  0
#> 83  30 96 90 77 82  0
#> 86  90 52 53 30 45  0
#> 87   6 96 53 30 45  0
#> 88  90 96 53 30 45  0
#> 89   6 52 90 30 45  0
#> 90  90 52 90 30 45  0
#> 91   6 96 90 30 45  0
#> 92  90 96 90 30 45  0
#> 93   6 52 53 77 45  0
#> 94  90 52 53 77 45  0
#> 95   6 96 53 77 45  0
#> 96  90 96 53 77 45  0
#> 97   6 52 90 77 45  0
#> 98  90 52 90 77 45  0
#> 99   6 96 90 77 45  0
#> 100 90 96 90 77 45  0
#> 101  6 52 53 30 82  0
#> 102 90 52 53 30 82  0
#> 103  6 96 53 30 82  0
#> 104 90 96 53 30 82  0
#> 105  6 52 90 30 82  0
#> 106 90 52 90 30 82  0
#> 107  6 96 90 30 82  0
#> 108 90 96 90 30 82  0
#> 109  6 52 53 77 82  0
#> 110 90 52 53 77 82  0
#> 111  6 96 53 77 82  0
#> 112 90 96 53 77 82  0
#> 113  6 52 90 77 82  0
#> 114 90 52 90 77 82  0
#> 115  6 96 90 77 82  0
#> 118 90 55 85 31 54  0
#> 119 22 96 85 31 54  0
#> 120 90 96 85 31 54  0
#> 121 22 55 90 31 54  0
#> 122 90 55 90 31 54  0
#> 123 22 96 90 31 54  0
#> 124 90 96 90 31 54  0
#> 125 22 55 85 77 54  0
#> 126 90 55 85 77 54  0
#> 127 22 96 85 77 54  0
#> 128 90 96 85 77 54  0
#> 129 22 55 90 77 54  0
#> 130 90 55 90 77 54  0
#> 131 22 96 90 77 54  0
#> 132 90 96 90 77 54  0
#> 133 22 55 85 31 82  0
#> 134 90 55 85 31 82  0
#> 135 22 96 85 31 82  0
#> 136 90 96 85 31 82  0
#> 137 22 55 90 31 82  0
#> 138 90 55 90 31 82  0
#> 139 22 96 90 31 82  0
#> 140 90 96 90 31 82  0
#> 141 22 55 85 77 82  0
#> 142 90 55 85 77 82  0
#> 143 22 96 85 77 82  0
#> 144 90 96 85 77 82  0
#> 145 22 55 90 77 82  0
#> 146 90 55 90 77 82  0
#> 147 22 96 90 77 82  0
#> 150 90 51 74 16 28  0
#> 151 36 96 74 16 28  0
#> 152 90 96 74 16 28  0
#> 153 36 51 90 16 28  0
#> 154 90 51 90 16 28  0
#> 155 36 96 90 16 28  0
#> 156 90 96 90 16 28  0
#> 157 36 51 74 77 28  0
#> 158 90 51 74 77 28  0
#> 159 36 96 74 77 28  0
#> 160 90 96 74 77 28  0
#> 161 36 51 90 77 28  0
#> 162 90 51 90 77 28  0
#> 163 36 96 90 77 28  0
#> 164 90 96 90 77 28  0
#> 165 36 51 74 16 82  0
#> 166 90 51 74 16 82  0
#> 167 36 96 74 16 82  0
#> 168 90 96 74 16 82  0
#> 169 36 51 90 16 82  0
#> 170 90 51 90 16 82  0
#> 171 36 96 90 16 82  0
#> 172 90 96 90 16 82  0
#> 173 36 51 74 77 82  0
#> 174 90 51 74 77 82  0
#> 175 36 96 74 77 82  0
#> 176 90 96 74 77 82  0
#> 177 36 51 90 77 82  0
#> 178 90 51 90 77 82  0
#> 179 36 96 90 77 82  0
#> 182 90 23 57 45 70  0
#> 183 16 96 57 45 70  0
#> 184 90 96 57 45 70  0
#> 185 16 23 90 45 70  0
#> 186 90 23 90 45 70  0
#> 187 16 96 90 45 70  0
#> 188 90 96 90 45 70  0
#> 189 16 23 57 77 70  0
#> 190 90 23 57 77 70  0
#> 191 16 96 57 77 70  0
#> 192 90 96 57 77 70  0
#> 193 16 23 90 77 70  0
#> 194 90 23 90 77 70  0
#> 195 16 96 90 77 70  0
#> 196 90 96 90 77 70  0
#> 197 16 23 57 45 82  0
#> 198 90 23 57 45 82  0
#> 199 16 96 57 45 82  0
#> 200 90 96 57 45 82  0
#> 201 16 23 90 45 82  0
#> 202 90 23 90 45 82  0
#> 203 16 96 90 45 82  0
#> 204 90 96 90 45 82  0
#> 205 16 23 57 77 82  0
#> 206 90 23 57 77 82  0
#> 207 16 96 57 77 82  0
#> 208 90 96 57 77 82  0
#> 209 16 23 90 77 82  0
#> 210 90 23 90 77 82  0
#> 211 16 96 90 77 82  0
#> 214 90 20 16 38 44  0
#> 215 38 96 16 38 44  0
#> 216 90 96 16 38 44  0
#> 217 38 20 90 38 44  0
#> 218 90 20 90 38 44  0
#> 219 38 96 90 38 44  0
#> 220 90 96 90 38 44  0
#> 221 38 20 16 77 44  0
#> 222 90 20 16 77 44  0
#> 223 38 96 16 77 44  0
#> 224 90 96 16 77 44  0
#> 225 38 20 90 77 44  0
#> 226 90 20 90 77 44  0
#> 227 38 96 90 77 44  0
#> 228 90 96 90 77 44  0
#> 229 38 20 16 38 82  0
#> 230 90 20 16 38 82  0
#> 231 38 96 16 38 82  0
#> 232 90 96 16 38 82  0
#> 233 38 20 90 38 82  0
#> 234 90 20 90 38 82  0
#> 235 38 96 90 38 82  0
#> 236 90 96 90 38 82  0
#> 237 38 20 16 77 82  0
#> 238 90 20 16 77 82  0
#> 239 38 96 16 77 82  0
#> 240 90 96 16 77 82  0
#> 241 38 20 90 77 82  0
#> 242 90 20 90 77 82  0
#> 243 38 96 90 77 82  0
#> 246 90 48 18 50 37  0
#> 247 85 96 18 50 37  0
#> 248 90 96 18 50 37  0
#> 249 85 48 90 50 37  0
#> 250 90 48 90 50 37  0
#> 251 85 96 90 50 37  0
#> 252 90 96 90 50 37  0
#> 253 85 48 18 77 37  0
#> 254 90 48 18 77 37  0
#> 255 85 96 18 77 37  0
#> 256 90 96 18 77 37  0
#> 257 85 48 90 77 37  0
#> 258 90 48 90 77 37  0
#> 259 85 96 90 77 37  0
#> 260 90 96 90 77 37  0
#> 261 85 48 18 50 82  0
#> 262 90 48 18 50 82  0
#> 263 85 96 18 50 82  0
#> 264 90 96 18 50 82  0
#> 265 85 48 90 50 82  0
#> 266 90 48 90 50 82  0
#> 267 85 96 90 50 82  0
#> 268 90 96 90 50 82  0
#> 269 85 48 18 77 82  0
#> 270 90 48 18 77 82  0
#> 271 85 96 18 77 82  0
#> 272 90 96 18 77 82  0
#> 273 85 48 90 77 82  0
#> 274 90 48 90 77 82  0
#> 275 85 96 90 77 82  0
#> 278 90 88 24 34 52  0
#> 279 41 96 24 34 52  0
#> 280 90 96 24 34 52  0
#> 281 41 88 90 34 52  0
#> 282 90 88 90 34 52  0
#> 283 41 96 90 34 52  0
#> 284 90 96 90 34 52  0
#> 285 41 88 24 77 52  0
#> 286 90 88 24 77 52  0
#> 287 41 96 24 77 52  0
#> 288 90 96 24 77 52  0
#> 289 41 88 90 77 52  0
#> 290 90 88 90 77 52  0
#> 291 41 96 90 77 52  0
#> 292 90 96 90 77 52  0
#> 293 41 88 24 34 82  0
#> 294 90 88 24 34 82  0
#> 295 41 96 24 34 82  0
#> 296 90 96 24 34 82  0
#> 297 41 88 90 34 82  0
#> 298 90 88 90 34 82  0
#> 299 41 96 90 34 82  0
#> 300 90 96 90 34 82  0
#> 301 41 88 24 77 82  0
#> 302 90 88 24 77 82  0
#> 303 41 96 24 77 82  0
#> 304 90 96 24 77 82  0
#> 305 41 88 90 77 82  0
#> 306 90 88 90 77 82  0
#> 307 41 96 90 77 82  0
#> 310 90 36 67 57  7  0
#> 311 44 96 67 57  7  0
#> 312 90 96 67 57  7  0
#> 313 44 36 90 57  7  0
#> 314 90 36 90 57  7  0
#> 315 44 96 90 57  7  0
#> 316 90 96 90 57  7  0
#> 317 44 36 67 77  7  0
#> 318 90 36 67 77  7  0
#> 319 44 96 67 77  7  0
#> 320 90 96 67 77  7  0
#> 321 44 36 90 77  7  0
#> 322 90 36 90 77  7  0
#> 323 44 96 90 77  7  0
#> 324 90 96 90 77  7  0
#> 325 44 36 67 57 82  0
#> 326 90 36 67 57 82  0
#> 327 44 96 67 57 82  0
#> 328 90 96 67 57 82  0
#> 329 44 36 90 57 82  0
#> 330 90 36 90 57 82  0
#> 331 44 96 90 57 82  0
#> 332 90 96 90 57 82  0
#> 333 44 36 67 77 82  0
#> 334 90 36 67 77 82  0
#> 335 44 96 67 77 82  0
#> 336 90 96 67 77 82  0
#> 337 44 36 90 77 82  0
#> 338 90 36 90 77 82  0
#> 339 44 96 90 77 82  0
#> 342 90 29 33 54 14  0
#> 343 68 96 33 54 14  0
#> 344 90 96 33 54 14  0
#> 345 68 29 90 54 14  0
#> 346 90 29 90 54 14  0
#> 347 68 96 90 54 14  0
#> 348 90 96 90 54 14  0
#> 349 68 29 33 77 14  0
#> 350 90 29 33 77 14  0
#> 351 68 96 33 77 14  0
#> 352 90 96 33 77 14  0
#> 353 68 29 90 77 14  0
#> 354 90 29 90 77 14  0
#> 355 68 96 90 77 14  0
#> 356 90 96 90 77 14  0
#> 357 68 29 33 54 82  0
#> 358 90 29 33 54 82  0
#> 359 68 96 33 54 82  0
#> 360 90 96 33 54 82  0
#> 361 68 29 90 54 82  0
#> 362 90 29 90 54 82  0
#> 363 68 96 90 54 82  0
#> 364 90 96 90 54 82  0
#> 365 68 29 33 77 82  0
#> 366 90 29 33 77 82  0
#> 367 68 96 33 77 82  0
#> 368 90 96 33 77 82  0
#> 369 68 29 90 77 82  0
#> 370 90 29 90 77 82  0
#> 371 68 96 90 77 82  0
#> 374 90 66 18 72 56  0
#> 375 25 96 18 72 56  0
#> 376 90 96 18 72 56  0
#> 377 25 66 90 72 56  0
#> 378 90 66 90 72 56  0
#> 379 25 96 90 72 56  0
#> 380 90 96 90 72 56  0
#> 381 25 66 18 77 56  0
#> 382 90 66 18 77 56  0
#> 383 25 96 18 77 56  0
#> 384 90 96 18 77 56  0
#> 385 25 66 90 77 56  0
#> 386 90 66 90 77 56  0
#> 387 25 96 90 77 56  0
#> 388 90 96 90 77 56  0
#> 389 25 66 18 72 82  0
#> 390 90 66 18 72 82  0
#> 391 25 96 18 72 82  0
#> 392 90 96 18 72 82  0
#> 393 25 66 90 72 82  0
#> 394 90 66 90 72 82  0
#> 395 25 96 90 72 82  0
#> 396 90 96 90 72 82  0
#> 397 25 66 18 77 82  0
#> 398 90 66 18 77 82  0
#> 399 25 96 18 77 82  0
#> 401 25 66 90 77 82  0
#> 402 90 66 90 77 82  0
#> 403 25 96 90 77 82  0
#> 406 90 81 31 31 30  0
#> 407 68 96 31 31 30  0
#> 408 90 96 31 31 30  0
#> 409 68 81 90 31 30  0
#> 410 90 81 90 31 30  0
#> 411 68 96 90 31 30  0
#> 412 90 96 90 31 30  0
#> 413 68 81 31 77 30  0
#> 414 90 81 31 77 30  0
#> 415 68 96 31 77 30  0
#> 416 90 96 31 77 30  0
#> 417 68 81 90 77 30  0
#> 418 90 81 90 77 30  0
#> 419 68 96 90 77 30  0
#> 420 90 96 90 77 30  0
#> 421 68 81 31 31 82  0
#> 422 90 81 31 31 82  0
#> 423 68 96 31 31 82  0
#> 424 90 96 31 31 82  0
#> 425 68 81 90 31 82  0
#> 426 90 81 90 31 82  0
#> 427 68 96 90 31 82  0
#> 429 68 81 31 77 82  0
#> 430 90 81 31 77 82  0
#> 431 68 96 31 77 82  0
#> 432 90 96 31 77 82  0
#> 433 68 81 90 77 82  0
#> 434 90 81 90 77 82  0
#> 438 90 60 37 28 54  0
#> 439 10 96 37 28 54  0
#> 440 90 96 37 28 54  0
#> 441 10 60 90 28 54  0
#> 442 90 60 90 28 54  0
#> 443 10 96 90 28 54  0
#> 444 90 96 90 28 54  0
#> 445 10 60 37 77 54  0
#> 446 90 60 37 77 54  0
#> 447 10 96 37 77 54  0
#> 448 90 96 37 77 54  0
#> 449 10 60 90 77 54  0
#> 450 90 60 90 77 54  0
#> 451 10 96 90 77 54  0
#> 453 10 60 37 28 82  0
#> 454 90 60 37 28 82  0
#> 455 10 96 37 28 82  0
#> 456 90 96 37 28 82  0
#> 457 10 60 90 28 82  0
#> 458 90 60 90 28 82  0
#> 459 10 96 90 28 82  0
#> 460 90 96 90 28 82  0
#> 461 10 60 37 77 82  0
#> 462 90 60 37 77 82  0
#> 463 10 96 37 77 82  0
#> 464 90 96 37 77 82  0
#> 465 10 60 90 77 82  0
#> 466 90 60 90 77 82  0
#> 467 10 96 90 77 82  0
#> 470 90 60 61 48 20  0
#> 471 14 96 61 48 20  0
#> 472 90 96 61 48 20  0
#> 473 14 60 90 48 20  0
#> 474 90 60 90 48 20  0
#> 475 14 96 90 48 20  0
#> 476 90 96 90 48 20  0
#> 477 14 60 61 77 20  0
#> 478 90 60 61 77 20  0
#> 479 14 96 61 77 20  0
#> 480 90 96 61 77 20  0
#> 481 14 60 90 77 20  0
#> 482 90 60 90 77 20  0
#> 483 14 96 90 77 20  0
#> 484 90 96 90 77 20  0
#> 485 14 60 61 48 82  0
#> 486 90 60 61 48 82  0
#> 487 14 96 61 48 82  0
#> 488 90 96 61 48 82  0
#> 489 14 60 90 48 82  0
#> 490 90 60 90 48 82  0
#> 491 14 96 90 48 82  0
#> 492 90 96 90 48 82  0
#> 493 14 60 61 77 82  0
#> 494 90 60 61 77 82  0
#> 495 14 96 61 77 82  0
#> 496 90 96 61 77 82  0
#> 497 14 60 90 77 82  0
#> 499 14 96 90 77 82  0
#> 502 90 19 62 29 65  0
#> 503 76 96 62 29 65  0
#> 504 90 96 62 29 65  0
#> 505 76 19 90 29 65  0
#> 506 90 19 90 29 65  0
#> 507 76 96 90 29 65  0
#> 508 90 96 90 29 65  0
#> 509 76 19 62 77 65  0
#> 510 90 19 62 77 65  0
#> 511 76 96 62 77 65  0
#> 512 90 96 62 77 65  0
#> 513 76 19 90 77 65  0
#> 514 90 19 90 77 65  0
#> 515 76 96 90 77 65  0
#> 516 90 96 90 77 65  0
#> 517 76 19 62 29 82  0
#> 518 90 19 62 29 82  0
#> 519 76 96 62 29 82  0
#> 520 90 96 62 29 82  0
#> 521 76 19 90 29 82  0
#> 522 90 19 90 29 82  0
#> 523 76 96 90 29 82  0
#> 524 90 96 90 29 82  0
#> 525 76 19 62 77 82  0
#> 526 90 19 62 77 82  0
#> 527 76 96 62 77 82  0
#> 528 90 96 62 77 82  0
#> 529 76 19 90 77 82  0
#> 530 90 19 90 77 82  0
#> 531 76 96 90 77 82  0
#> 534 90 52 25 13 31  0
#> 535 54 96 25 13 31  0
#> 536 90 96 25 13 31  0
#> 537 54 52 90 13 31  0
#> 538 90 52 90 13 31  0
#> 539 54 96 90 13 31  0
#> 540 90 96 90 13 31  0
#> 541 54 52 25 77 31  0
#> 542 90 52 25 77 31  0
#> 543 54 96 25 77 31  0
#> 544 90 96 25 77 31  0
#> 545 54 52 90 77 31  0
#> 546 90 52 90 77 31  0
#> 547 54 96 90 77 31  0
#> 548 90 96 90 77 31  0
#> 549 54 52 25 13 82  0
#> 550 90 52 25 13 82  0
#> 551 54 96 25 13 82  0
#> 552 90 96 25 13 82  0
#> 553 54 52 90 13 82  0
#> 554 90 52 90 13 82  0
#> 555 54 96 90 13 82  0
#> 556 90 96 90 13 82  0
#> 557 54 52 25 77 82  0
#> 558 90 52 25 77 82  0
#> 559 54 96 25 77 82  0
#> 560 90 96 25 77 82  0
#> 561 54 52 90 77 82  0
#> 563 54 96 90 77 82  0
#> 566 90 54  4 66 50  0
#> 567 48 96  4 66 50  0
#> 568 90 96  4 66 50  0
#> 569 48 54 90 66 50  0
#> 570 90 54 90 66 50  0
#> 571 48 96 90 66 50  0
#> 572 90 96 90 66 50  0
#> 573 48 54  4 77 50  0
#> 574 90 54  4 77 50  0
#> 575 48 96  4 77 50  0
#> 576 90 96  4 77 50  0
#> 577 48 54 90 77 50  0
#> 578 90 54 90 77 50  0
#> 579 48 96 90 77 50  0
#> 580 90 96 90 77 50  0
#> 581 48 54  4 66 82  0
#> 582 90 54  4 66 82  0
#> 583 48 96  4 66 82  0
#> 584 90 96  4 66 82  0
#> 585 48 54 90 66 82  0
#> 586 90 54 90 66 82  0
#> 587 48 96 90 66 82  0
#> 588 90 96 90 66 82  0
#> 589 48 54  4 77 82  0
#> 590 90 54  4 77 82  0
#> 591 48 96  4 77 82  0
#> 592 90 96  4 77 82  0
#> 593 48 54 90 77 82  0
#> 594 90 54 90 77 82  0
#> 595 48 96 90 77 82  0
#> 598 90 45  5 48 40  0
#> 599 35 96  5 48 40  0
#> 600 90 96  5 48 40  0
#> 601 35 45 90 48 40  0
#> 602 90 45 90 48 40  0
#> 603 35 96 90 48 40  0
#> 604 90 96 90 48 40  0
#> 605 35 45  5 77 40  0
#> 606 90 45  5 77 40  0
#> 607 35 96  5 77 40  0
#> 608 90 96  5 77 40  0
#> 609 35 45 90 77 40  0
#> 610 90 45 90 77 40  0
#> 611 35 96 90 77 40  0
#> 612 90 96 90 77 40  0
#> 613 35 45  5 48 82  0
#> 614 90 45  5 48 82  0
#> 615 35 96  5 48 82  0
#> 616 90 96  5 48 82  0
#> 617 35 45 90 48 82  0
#> 618 90 45 90 48 82  0
#> 619 35 96 90 48 82  0
#> 621 35 45  5 77 82  0
#> 622 90 45  5 77 82  0
#> 623 35 96  5 77 82  0
#> 624 90 96  5 77 82  0
#> 625 35 45 90 77 82  0
#> 626 90 45 90 77 82  0
#> 627 35 96 90 77 82  0
#> 630 90 51 46  7 30  0
#> 631 61 96 46  7 30  0
#> 632 90 96 46  7 30  0
#> 633 61 51 90  7 30  0
#> 634 90 51 90  7 30  0
#> 635 61 96 90  7 30  0
#> 636 90 96 90  7 30  0
#> 637 61 51 46 77 30  0
#> 638 90 51 46 77 30  0
#> 639 61 96 46 77 30  0
#> 640 90 96 46 77 30  0
#> 641 61 51 90 77 30  0
#> 642 90 51 90 77 30  0
#> 643 61 96 90 77 30  0
#> 645 61 51 46  7 82  0
#> 646 90 51 46  7 82  0
#> 647 61 96 46  7 82  0
#> 648 90 96 46  7 82  0
#> 649 61 51 90  7 82  0
#> 650 90 51 90  7 82  0
#> 651 61 96 90  7 82  0
#> 652 90 96 90  7 82  0
#> 653 61 51 46 77 82  0
#> 654 90 51 46 77 82  0
#> 655 61 96 46 77 82  0
#> 656 90 96 46 77 82  0
#> 657 61 51 90 77 82  0
#> 659 61 96 90 77 82  0