Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts
A data frame with all points

m
Minimum values of the bounding box.

M
Maximum values of the bounding box.

direction
Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts.

Examples

pts <- genNDSet(3,10)[,1:3]

#> Warning: Only 8 samples generated!

addRays(pts)

#>    z1 z2 z3 pt
#> 5   2 55 57  1
#> 4   6 71 45  1
#> 6  19 25 79  1
#> 1  30 12 74  1
#> 2  53 47  1  1
#> 3  56  1 47  1
#> 21 61 55 57  0
#> 31  2 76 57  0
#> 41 61 76 57  0
#> 51  2 55 84  0
#> 61 61 55 84  0
#> 7   2 76 84  0
#> 8  61 76 84  0
#> 22 61 71 45  0
#> 32  6 76 45  0
#> 42 61 76 45  0
#> 52  6 71 84  0
#> 62 61 71 84  0
#> 71  6 76 84  0
#> 23 61 25 79  0
#> 33 19 76 79  0
#> 43 61 76 79  0
#> 53 19 25 84  0
#> 63 61 25 84  0
#> 72 19 76 84  0
#> 24 61 12 74  0
#> 34 30 76 74  0
#> 44 61 76 74  0
#> 54 30 12 84  0
#> 64 61 12 84  0
#> 73 30 76 84  0
#> 25 61 47  1  0
#> 35 53 76  1  0
#> 45 61 76  1  0
#> 55 53 47 84  0
#> 65 61 47 84  0
#> 74 53 76 84  0
#> 26 61  1 47  0
#> 36 56 76 47  0
#> 46 61 76 47  0
#> 56 56  1 84  0
#> 66 61  1 84  0
#> 75 56 76 84  0

addRays(pts, dir = c(1,-1,1))

#>    z1 z2 z3 pt
#> 5   2 55 57  1
#> 4   6 71 45  1
#> 6  19 25 79  1
#> 1  30 12 74  1
#> 2  53 47  1  1
#> 3  56  1 47  1
#> 21 61 55 57  0
#> 31  2 -4 57  0
#> 41 61 -4 57  0
#> 51  2 55 84  0
#> 61 61 55 84  0
#> 7   2 -4 84  0
#> 8  61 -4 84  0
#> 22 61 71 45  0
#> 32  6 -4 45  0
#> 42 61 -4 45  0
#> 52  6 71 84  0
#> 62 61 71 84  0
#> 71  6 -4 84  0
#> 23 61 25 79  0
#> 33 19 -4 79  0
#> 43 61 -4 79  0
#> 53 19 25 84  0
#> 63 61 25 84  0
#> 72 19 -4 84  0
#> 24 61 12 74  0
#> 34 30 -4 74  0
#> 44 61 -4 74  0
#> 54 30 12 84  0
#> 64 61 12 84  0
#> 73 30 -4 84  0
#> 25 61 47  1  0
#> 35 53 -4  1  0
#> 45 61 -4  1  0
#> 55 53 47 84  0
#> 65 61 47 84  0
#> 74 53 -4 84  0
#> 26 61  1 47  0
#> 36 56 -4 47  0
#> 46 61 -4 47  0
#> 56 56  1 84  0
#> 66 61  1 84  0
#> 75 56 -4 84  0

addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))

#>    z1 z2  z3 pt
#> 5   2 55  57  1
#> 4   6 71  45  1
#> 6  19 25  79  1
#> 1  30 12  74  1
#> 2  53 47   1  1
#> 3  56  1  47  1
#> 21  0 55  57  0
#> 31  2  0  57  0
#> 41  0  0  57  0
#> 51  2 55 100  0
#> 61  0 55 100  0
#> 7   2  0 100  0
#> 8   0  0 100  0
#> 22  0 71  45  0
#> 32  6  0  45  0
#> 42  0  0  45  0
#> 52  6 71 100  0
#> 62  0 71 100  0
#> 71  6  0 100  0
#> 23  0 25  79  0
#> 33 19  0  79  0
#> 43  0  0  79  0
#> 53 19 25 100  0
#> 63  0 25 100  0
#> 72 19  0 100  0
#> 24  0 12  74  0
#> 34 30  0  74  0
#> 44  0  0  74  0
#> 54 30 12 100  0
#> 64  0 12 100  0
#> 73 30  0 100  0
#> 25  0 47   1  0
#> 35 53  0   1  0
#> 45  0  0   1  0
#> 55 53 47 100  0
#> 65  0 47 100  0
#> 74 53  0 100  0
#> 26  0  1  47  0
#> 36 56  0  47  0
#> 46  0  0  47  0
#> 56 56  1 100  0
#> 66  0  1 100  0
#> 75 56  0 100  0

pts <- genSample(5,20)[,1:5]
addRays(pts)

#>     V1 V2 V3 V4 V5 pt
#> 1   42 34 28 34 14  1
#> 2   27 31 32 58 84  1
#> 3   44  7 58 37 68  1
#> 4   24 70 52 20 29  1
#> 5   63 54 17 34 81  1
#> 6   17 67 45 55 19  1
#> 7   25 14 62 41 67  1
#> 8   32 30 21 77 60  1
#> 9   43 74 49 46  8  1
#> 10  51 35 42 22 87  1
#> 11  22 42 70 19 64  1
#> 12  35 47 52 56  4  1
#> 13  56 21 67 15 52  1
#> 14  37 33 64 74 16  1
#> 15  32 39 24 34 83  1
#> 16  29 58 38 34 89  1
#> 17  55  9 50 44 24  1
#> 18  16 82 35 49 54  1
#> 19  52  1 48 53 51  1
#> 20  39 26 62 12 59  1
#> 22  68 34 28 34 14  0
#> 23  42 87 28 34 14  0
#> 24  68 87 28 34 14  0
#> 25  42 34 75 34 14  0
#> 26  68 34 75 34 14  0
#> 27  42 87 75 34 14  0
#> 28  68 87 75 34 14  0
#> 29  42 34 28 82 14  0
#> 30  68 34 28 82 14  0
#> 31  42 87 28 82 14  0
#> 32  68 87 28 82 14  0
#> 33  42 34 75 82 14  0
#> 34  68 34 75 82 14  0
#> 35  42 87 75 82 14  0
#> 36  68 87 75 82 14  0
#> 37  42 34 28 34 94  0
#> 38  68 34 28 34 94  0
#> 39  42 87 28 34 94  0
#> 40  68 87 28 34 94  0
#> 41  42 34 75 34 94  0
#> 42  68 34 75 34 94  0
#> 43  42 87 75 34 94  0
#> 44  68 87 75 34 94  0
#> 45  42 34 28 82 94  0
#> 46  68 34 28 82 94  0
#> 47  42 87 28 82 94  0
#> 48  68 87 28 82 94  0
#> 49  42 34 75 82 94  0
#> 50  68 34 75 82 94  0
#> 51  42 87 75 82 94  0
#> 52  68 87 75 82 94  0
#> 54  68 31 32 58 84  0
#> 55  27 87 32 58 84  0
#> 56  68 87 32 58 84  0
#> 57  27 31 75 58 84  0
#> 58  68 31 75 58 84  0
#> 59  27 87 75 58 84  0
#> 60  68 87 75 58 84  0
#> 61  27 31 32 82 84  0
#> 62  68 31 32 82 84  0
#> 63  27 87 32 82 84  0
#> 64  68 87 32 82 84  0
#> 65  27 31 75 82 84  0
#> 66  68 31 75 82 84  0
#> 67  27 87 75 82 84  0
#> 68  68 87 75 82 84  0
#> 69  27 31 32 58 94  0
#> 70  68 31 32 58 94  0
#> 71  27 87 32 58 94  0
#> 72  68 87 32 58 94  0
#> 73  27 31 75 58 94  0
#> 74  68 31 75 58 94  0
#> 75  27 87 75 58 94  0
#> 76  68 87 75 58 94  0
#> 77  27 31 32 82 94  0
#> 78  68 31 32 82 94  0
#> 79  27 87 32 82 94  0
#> 80  68 87 32 82 94  0
#> 81  27 31 75 82 94  0
#> 82  68 31 75 82 94  0
#> 83  27 87 75 82 94  0
#> 86  68  7 58 37 68  0
#> 87  44 87 58 37 68  0
#> 88  68 87 58 37 68  0
#> 89  44  7 75 37 68  0
#> 90  68  7 75 37 68  0
#> 91  44 87 75 37 68  0
#> 92  68 87 75 37 68  0
#> 93  44  7 58 82 68  0
#> 94  68  7 58 82 68  0
#> 95  44 87 58 82 68  0
#> 96  68 87 58 82 68  0
#> 97  44  7 75 82 68  0
#> 98  68  7 75 82 68  0
#> 99  44 87 75 82 68  0
#> 100 68 87 75 82 68  0
#> 101 44  7 58 37 94  0
#> 102 68  7 58 37 94  0
#> 103 44 87 58 37 94  0
#> 104 68 87 58 37 94  0
#> 105 44  7 75 37 94  0
#> 106 68  7 75 37 94  0
#> 107 44 87 75 37 94  0
#> 108 68 87 75 37 94  0
#> 109 44  7 58 82 94  0
#> 110 68  7 58 82 94  0
#> 111 44 87 58 82 94  0
#> 112 68 87 58 82 94  0
#> 113 44  7 75 82 94  0
#> 114 68  7 75 82 94  0
#> 115 44 87 75 82 94  0
#> 118 68 70 52 20 29  0
#> 119 24 87 52 20 29  0
#> 120 68 87 52 20 29  0
#> 121 24 70 75 20 29  0
#> 122 68 70 75 20 29  0
#> 123 24 87 75 20 29  0
#> 124 68 87 75 20 29  0
#> 125 24 70 52 82 29  0
#> 126 68 70 52 82 29  0
#> 127 24 87 52 82 29  0
#> 128 68 87 52 82 29  0
#> 129 24 70 75 82 29  0
#> 130 68 70 75 82 29  0
#> 131 24 87 75 82 29  0
#> 132 68 87 75 82 29  0
#> 133 24 70 52 20 94  0
#> 134 68 70 52 20 94  0
#> 135 24 87 52 20 94  0
#> 136 68 87 52 20 94  0
#> 137 24 70 75 20 94  0
#> 138 68 70 75 20 94  0
#> 139 24 87 75 20 94  0
#> 140 68 87 75 20 94  0
#> 141 24 70 52 82 94  0
#> 142 68 70 52 82 94  0
#> 143 24 87 52 82 94  0
#> 144 68 87 52 82 94  0
#> 145 24 70 75 82 94  0
#> 146 68 70 75 82 94  0
#> 147 24 87 75 82 94  0
#> 150 68 54 17 34 81  0
#> 151 63 87 17 34 81  0
#> 152 68 87 17 34 81  0
#> 153 63 54 75 34 81  0
#> 154 68 54 75 34 81  0
#> 155 63 87 75 34 81  0
#> 156 68 87 75 34 81  0
#> 157 63 54 17 82 81  0
#> 158 68 54 17 82 81  0
#> 159 63 87 17 82 81  0
#> 160 68 87 17 82 81  0
#> 161 63 54 75 82 81  0
#> 162 68 54 75 82 81  0
#> 163 63 87 75 82 81  0
#> 164 68 87 75 82 81  0
#> 165 63 54 17 34 94  0
#> 166 68 54 17 34 94  0
#> 167 63 87 17 34 94  0
#> 168 68 87 17 34 94  0
#> 169 63 54 75 34 94  0
#> 170 68 54 75 34 94  0
#> 171 63 87 75 34 94  0
#> 173 63 54 17 82 94  0
#> 174 68 54 17 82 94  0
#> 175 63 87 17 82 94  0
#> 176 68 87 17 82 94  0
#> 177 63 54 75 82 94  0
#> 178 68 54 75 82 94  0
#> 179 63 87 75 82 94  0
#> 182 68 67 45 55 19  0
#> 183 17 87 45 55 19  0
#> 184 68 87 45 55 19  0
#> 185 17 67 75 55 19  0
#> 186 68 67 75 55 19  0
#> 187 17 87 75 55 19  0
#> 188 68 87 75 55 19  0
#> 189 17 67 45 82 19  0
#> 190 68 67 45 82 19  0
#> 191 17 87 45 82 19  0
#> 192 68 87 45 82 19  0
#> 193 17 67 75 82 19  0
#> 194 68 67 75 82 19  0
#> 195 17 87 75 82 19  0
#> 196 68 87 75 82 19  0
#> 197 17 67 45 55 94  0
#> 198 68 67 45 55 94  0
#> 199 17 87 45 55 94  0
#> 200 68 87 45 55 94  0
#> 201 17 67 75 55 94  0
#> 202 68 67 75 55 94  0
#> 203 17 87 75 55 94  0
#> 204 68 87 75 55 94  0
#> 205 17 67 45 82 94  0
#> 206 68 67 45 82 94  0
#> 207 17 87 45 82 94  0
#> 208 68 87 45 82 94  0
#> 209 17 67 75 82 94  0
#> 210 68 67 75 82 94  0
#> 211 17 87 75 82 94  0
#> 214 68 14 62 41 67  0
#> 215 25 87 62 41 67  0
#> 216 68 87 62 41 67  0
#> 217 25 14 75 41 67  0
#> 218 68 14 75 41 67  0
#> 219 25 87 75 41 67  0
#> 220 68 87 75 41 67  0
#> 221 25 14 62 82 67  0
#> 222 68 14 62 82 67  0
#> 223 25 87 62 82 67  0
#> 224 68 87 62 82 67  0
#> 225 25 14 75 82 67  0
#> 226 68 14 75 82 67  0
#> 227 25 87 75 82 67  0
#> 228 68 87 75 82 67  0
#> 229 25 14 62 41 94  0
#> 230 68 14 62 41 94  0
#> 231 25 87 62 41 94  0
#> 232 68 87 62 41 94  0
#> 233 25 14 75 41 94  0
#> 234 68 14 75 41 94  0
#> 235 25 87 75 41 94  0
#> 236 68 87 75 41 94  0
#> 237 25 14 62 82 94  0
#> 238 68 14 62 82 94  0
#> 239 25 87 62 82 94  0
#> 240 68 87 62 82 94  0
#> 241 25 14 75 82 94  0
#> 242 68 14 75 82 94  0
#> 243 25 87 75 82 94  0
#> 246 68 30 21 77 60  0
#> 247 32 87 21 77 60  0
#> 248 68 87 21 77 60  0
#> 249 32 30 75 77 60  0
#> 250 68 30 75 77 60  0
#> 251 32 87 75 77 60  0
#> 252 68 87 75 77 60  0
#> 253 32 30 21 82 60  0
#> 254 68 30 21 82 60  0
#> 255 32 87 21 82 60  0
#> 256 68 87 21 82 60  0
#> 257 32 30 75 82 60  0
#> 258 68 30 75 82 60  0
#> 259 32 87 75 82 60  0
#> 260 68 87 75 82 60  0
#> 261 32 30 21 77 94  0
#> 262 68 30 21 77 94  0
#> 263 32 87 21 77 94  0
#> 264 68 87 21 77 94  0
#> 265 32 30 75 77 94  0
#> 266 68 30 75 77 94  0
#> 267 32 87 75 77 94  0
#> 268 68 87 75 77 94  0
#> 269 32 30 21 82 94  0
#> 270 68 30 21 82 94  0
#> 271 32 87 21 82 94  0
#> 272 68 87 21 82 94  0
#> 273 32 30 75 82 94  0
#> 274 68 30 75 82 94  0
#> 275 32 87 75 82 94  0
#> 278 68 74 49 46  8  0
#> 279 43 87 49 46  8  0
#> 280 68 87 49 46  8  0
#> 281 43 74 75 46  8  0
#> 282 68 74 75 46  8  0
#> 283 43 87 75 46  8  0
#> 284 68 87 75 46  8  0
#> 285 43 74 49 82  8  0
#> 286 68 74 49 82  8  0
#> 287 43 87 49 82  8  0
#> 288 68 87 49 82  8  0
#> 289 43 74 75 82  8  0
#> 290 68 74 75 82  8  0
#> 291 43 87 75 82  8  0
#> 292 68 87 75 82  8  0
#> 293 43 74 49 46 94  0
#> 294 68 74 49 46 94  0
#> 295 43 87 49 46 94  0
#> 296 68 87 49 46 94  0
#> 297 43 74 75 46 94  0
#> 298 68 74 75 46 94  0
#> 299 43 87 75 46 94  0
#> 300 68 87 75 46 94  0
#> 301 43 74 49 82 94  0
#> 302 68 74 49 82 94  0
#> 303 43 87 49 82 94  0
#> 304 68 87 49 82 94  0
#> 305 43 74 75 82 94  0
#> 306 68 74 75 82 94  0
#> 307 43 87 75 82 94  0
#> 310 68 35 42 22 87  0
#> 311 51 87 42 22 87  0
#> 312 68 87 42 22 87  0
#> 313 51 35 75 22 87  0
#> 314 68 35 75 22 87  0
#> 315 51 87 75 22 87  0
#> 316 68 87 75 22 87  0
#> 317 51 35 42 82 87  0
#> 318 68 35 42 82 87  0
#> 319 51 87 42 82 87  0
#> 320 68 87 42 82 87  0
#> 321 51 35 75 82 87  0
#> 322 68 35 75 82 87  0
#> 323 51 87 75 82 87  0
#> 324 68 87 75 82 87  0
#> 325 51 35 42 22 94  0
#> 326 68 35 42 22 94  0
#> 327 51 87 42 22 94  0
#> 328 68 87 42 22 94  0
#> 329 51 35 75 22 94  0
#> 330 68 35 75 22 94  0
#> 331 51 87 75 22 94  0
#> 332 68 87 75 22 94  0
#> 333 51 35 42 82 94  0
#> 334 68 35 42 82 94  0
#> 335 51 87 42 82 94  0
#> 336 68 87 42 82 94  0
#> 337 51 35 75 82 94  0
#> 338 68 35 75 82 94  0
#> 339 51 87 75 82 94  0
#> 342 68 42 70 19 64  0
#> 343 22 87 70 19 64  0
#> 344 68 87 70 19 64  0
#> 345 22 42 75 19 64  0
#> 346 68 42 75 19 64  0
#> 347 22 87 75 19 64  0
#> 348 68 87 75 19 64  0
#> 349 22 42 70 82 64  0
#> 350 68 42 70 82 64  0
#> 351 22 87 70 82 64  0
#> 352 68 87 70 82 64  0
#> 353 22 42 75 82 64  0
#> 354 68 42 75 82 64  0
#> 355 22 87 75 82 64  0
#> 356 68 87 75 82 64  0
#> 357 22 42 70 19 94  0
#> 358 68 42 70 19 94  0
#> 359 22 87 70 19 94  0
#> 360 68 87 70 19 94  0
#> 361 22 42 75 19 94  0
#> 362 68 42 75 19 94  0
#> 363 22 87 75 19 94  0
#> 364 68 87 75 19 94  0
#> 365 22 42 70 82 94  0
#> 366 68 42 70 82 94  0
#> 367 22 87 70 82 94  0
#> 368 68 87 70 82 94  0
#> 369 22 42 75 82 94  0
#> 370 68 42 75 82 94  0
#> 371 22 87 75 82 94  0
#> 374 68 47 52 56  4  0
#> 375 35 87 52 56  4  0
#> 376 68 87 52 56  4  0
#> 377 35 47 75 56  4  0
#> 378 68 47 75 56  4  0
#> 379 35 87 75 56  4  0
#> 380 68 87 75 56  4  0
#> 381 35 47 52 82  4  0
#> 382 68 47 52 82  4  0
#> 383 35 87 52 82  4  0
#> 384 68 87 52 82  4  0
#> 385 35 47 75 82  4  0
#> 386 68 47 75 82  4  0
#> 387 35 87 75 82  4  0
#> 388 68 87 75 82  4  0
#> 389 35 47 52 56 94  0
#> 390 68 47 52 56 94  0
#> 391 35 87 52 56 94  0
#> 392 68 87 52 56 94  0
#> 393 35 47 75 56 94  0
#> 394 68 47 75 56 94  0
#> 395 35 87 75 56 94  0
#> 396 68 87 75 56 94  0
#> 397 35 47 52 82 94  0
#> 398 68 47 52 82 94  0
#> 399 35 87 52 82 94  0
#> 401 35 47 75 82 94  0
#> 402 68 47 75 82 94  0
#> 403 35 87 75 82 94  0
#> 406 68 21 67 15 52  0
#> 407 56 87 67 15 52  0
#> 408 68 87 67 15 52  0
#> 409 56 21 75 15 52  0
#> 410 68 21 75 15 52  0
#> 411 56 87 75 15 52  0
#> 412 68 87 75 15 52  0
#> 413 56 21 67 82 52  0
#> 414 68 21 67 82 52  0
#> 415 56 87 67 82 52  0
#> 416 68 87 67 82 52  0
#> 417 56 21 75 82 52  0
#> 418 68 21 75 82 52  0
#> 419 56 87 75 82 52  0
#> 420 68 87 75 82 52  0
#> 421 56 21 67 15 94  0
#> 422 68 21 67 15 94  0
#> 423 56 87 67 15 94  0
#> 424 68 87 67 15 94  0
#> 425 56 21 75 15 94  0
#> 426 68 21 75 15 94  0
#> 427 56 87 75 15 94  0
#> 428 68 87 75 15 94  0
#> 429 56 21 67 82 94  0
#> 430 68 21 67 82 94  0
#> 431 56 87 67 82 94  0
#> 432 68 87 67 82 94  0
#> 433 56 21 75 82 94  0
#> 434 68 21 75 82 94  0
#> 435 56 87 75 82 94  0
#> 438 68 33 64 74 16  0
#> 439 37 87 64 74 16  0
#> 440 68 87 64 74 16  0
#> 441 37 33 75 74 16  0
#> 442 68 33 75 74 16  0
#> 443 37 87 75 74 16  0
#> 444 68 87 75 74 16  0
#> 445 37 33 64 82 16  0
#> 446 68 33 64 82 16  0
#> 447 37 87 64 82 16  0
#> 448 68 87 64 82 16  0
#> 449 37 33 75 82 16  0
#> 450 68 33 75 82 16  0
#> 451 37 87 75 82 16  0
#> 452 68 87 75 82 16  0
#> 453 37 33 64 74 94  0
#> 454 68 33 64 74 94  0
#> 455 37 87 64 74 94  0
#> 456 68 87 64 74 94  0
#> 457 37 33 75 74 94  0
#> 458 68 33 75 74 94  0
#> 459 37 87 75 74 94  0
#> 460 68 87 75 74 94  0
#> 461 37 33 64 82 94  0
#> 462 68 33 64 82 94  0
#> 463 37 87 64 82 94  0
#> 464 68 87 64 82 94  0
#> 465 37 33 75 82 94  0
#> 466 68 33 75 82 94  0
#> 467 37 87 75 82 94  0
#> 470 68 39 24 34 83  0
#> 471 32 87 24 34 83  0
#> 472 68 87 24 34 83  0
#> 473 32 39 75 34 83  0
#> 474 68 39 75 34 83  0
#> 475 32 87 75 34 83  0
#> 476 68 87 75 34 83  0
#> 477 32 39 24 82 83  0
#> 478 68 39 24 82 83  0
#> 479 32 87 24 82 83  0
#> 480 68 87 24 82 83  0
#> 481 32 39 75 82 83  0
#> 482 68 39 75 82 83  0
#> 483 32 87 75 82 83  0
#> 484 68 87 75 82 83  0
#> 485 32 39 24 34 94  0
#> 486 68 39 24 34 94  0
#> 487 32 87 24 34 94  0
#> 488 68 87 24 34 94  0
#> 489 32 39 75 34 94  0
#> 490 68 39 75 34 94  0
#> 491 32 87 75 34 94  0
#> 493 32 39 24 82 94  0
#> 494 68 39 24 82 94  0
#> 495 32 87 24 82 94  0
#> 496 68 87 24 82 94  0
#> 497 32 39 75 82 94  0
#> 498 68 39 75 82 94  0
#> 502 68 58 38 34 89  0
#> 503 29 87 38 34 89  0
#> 504 68 87 38 34 89  0
#> 505 29 58 75 34 89  0
#> 506 68 58 75 34 89  0
#> 507 29 87 75 34 89  0
#> 508 68 87 75 34 89  0
#> 509 29 58 38 82 89  0
#> 510 68 58 38 82 89  0
#> 511 29 87 38 82 89  0
#> 512 68 87 38 82 89  0
#> 513 29 58 75 82 89  0
#> 514 68 58 75 82 89  0
#> 515 29 87 75 82 89  0
#> 516 68 87 75 82 89  0
#> 517 29 58 38 34 94  0
#> 518 68 58 38 34 94  0
#> 519 29 87 38 34 94  0
#> 520 68 87 38 34 94  0
#> 521 29 58 75 34 94  0
#> 522 68 58 75 34 94  0
#> 523 29 87 75 34 94  0
#> 525 29 58 38 82 94  0
#> 526 68 58 38 82 94  0
#> 527 29 87 38 82 94  0
#> 528 68 87 38 82 94  0
#> 529 29 58 75 82 94  0
#> 530 68 58 75 82 94  0
#> 531 29 87 75 82 94  0
#> 534 68  9 50 44 24  0
#> 535 55 87 50 44 24  0
#> 536 68 87 50 44 24  0
#> 537 55  9 75 44 24  0
#> 538 68  9 75 44 24  0
#> 539 55 87 75 44 24  0
#> 540 68 87 75 44 24  0
#> 541 55  9 50 82 24  0
#> 542 68  9 50 82 24  0
#> 543 55 87 50 82 24  0
#> 544 68 87 50 82 24  0
#> 545 55  9 75 82 24  0
#> 546 68  9 75 82 24  0
#> 547 55 87 75 82 24  0
#> 548 68 87 75 82 24  0
#> 549 55  9 50 44 94  0
#> 550 68  9 50 44 94  0
#> 551 55 87 50 44 94  0
#> 552 68 87 50 44 94  0
#> 553 55  9 75 44 94  0
#> 554 68  9 75 44 94  0
#> 555 55 87 75 44 94  0
#> 556 68 87 75 44 94  0
#> 557 55  9 50 82 94  0
#> 558 68  9 50 82 94  0
#> 559 55 87 50 82 94  0
#> 560 68 87 50 82 94  0
#> 561 55  9 75 82 94  0
#> 562 68  9 75 82 94  0
#> 563 55 87 75 82 94  0
#> 566 68 82 35 49 54  0
#> 567 16 87 35 49 54  0
#> 568 68 87 35 49 54  0
#> 569 16 82 75 49 54  0
#> 570 68 82 75 49 54  0
#> 571 16 87 75 49 54  0
#> 572 68 87 75 49 54  0
#> 573 16 82 35 82 54  0
#> 574 68 82 35 82 54  0
#> 575 16 87 35 82 54  0
#> 576 68 87 35 82 54  0
#> 577 16 82 75 82 54  0
#> 578 68 82 75 82 54  0
#> 579 16 87 75 82 54  0
#> 580 68 87 75 82 54  0
#> 581 16 82 35 49 94  0
#> 582 68 82 35 49 94  0
#> 583 16 87 35 49 94  0
#> 584 68 87 35 49 94  0
#> 585 16 82 75 49 94  0
#> 586 68 82 75 49 94  0
#> 587 16 87 75 49 94  0
#> 588 68 87 75 49 94  0
#> 589 16 82 35 82 94  0
#> 590 68 82 35 82 94  0
#> 591 16 87 35 82 94  0
#> 592 68 87 35 82 94  0
#> 593 16 82 75 82 94  0
#> 594 68 82 75 82 94  0
#> 595 16 87 75 82 94  0
#> 598 68  1 48 53 51  0
#> 599 52 87 48 53 51  0
#> 600 68 87 48 53 51  0
#> 601 52  1 75 53 51  0
#> 602 68  1 75 53 51  0
#> 603 52 87 75 53 51  0
#> 604 68 87 75 53 51  0
#> 605 52  1 48 82 51  0
#> 606 68  1 48 82 51  0
#> 607 52 87 48 82 51  0
#> 608 68 87 48 82 51  0
#> 609 52  1 75 82 51  0
#> 610 68  1 75 82 51  0
#> 611 52 87 75 82 51  0
#> 612 68 87 75 82 51  0
#> 613 52  1 48 53 94  0
#> 614 68  1 48 53 94  0
#> 615 52 87 48 53 94  0
#> 616 68 87 48 53 94  0
#> 617 52  1 75 53 94  0
#> 618 68  1 75 53 94  0
#> 619 52 87 75 53 94  0
#> 620 68 87 75 53 94  0
#> 621 52  1 48 82 94  0
#> 622 68  1 48 82 94  0
#> 623 52 87 48 82 94  0
#> 624 68 87 48 82 94  0
#> 625 52  1 75 82 94  0
#> 626 68  1 75 82 94  0
#> 627 52 87 75 82 94  0
#> 630 68 26 62 12 59  0
#> 631 39 87 62 12 59  0
#> 632 68 87 62 12 59  0
#> 633 39 26 75 12 59  0
#> 634 68 26 75 12 59  0
#> 635 39 87 75 12 59  0
#> 636 68 87 75 12 59  0
#> 637 39 26 62 82 59  0
#> 638 68 26 62 82 59  0
#> 639 39 87 62 82 59  0
#> 640 68 87 62 82 59  0
#> 641 39 26 75 82 59  0
#> 642 68 26 75 82 59  0
#> 643 39 87 75 82 59  0
#> 644 68 87 75 82 59  0
#> 645 39 26 62 12 94  0
#> 646 68 26 62 12 94  0
#> 647 39 87 62 12 94  0
#> 648 68 87 62 12 94  0
#> 649 39 26 75 12 94  0
#> 650 68 26 75 12 94  0
#> 651 39 87 75 12 94  0
#> 652 68 87 75 12 94  0
#> 653 39 26 62 82 94  0
#> 654 68 26 62 82 94  0
#> 655 39 87 62 82 94  0
#> 657 39 26 75 82 94  0
#> 658 68 26 75 82 94  0
#> 659 39 87 75 82 94  0