Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 6   4 46 67  1
#> 10  5 34 42  1
#> 5   7 55 28  1
#> 8   9 28 35  1
#> 3  16 50 15  1
#> 1  40  2 51  1
#> 4  66  6 36  1
#> 2  75 35 10  1
#> 21 80 46 67  0
#> 31  4 60 67  0
#> 41 80 60 67  0
#> 51  4 46 72  0
#> 61 80 46 72  0
#> 7   4 60 72  0
#> 81 80 60 72  0
#> 22 80 34 42  0
#> 32  5 60 42  0
#> 42 80 60 42  0
#> 52  5 34 72  0
#> 62 80 34 72  0
#> 71  5 60 72  0
#> 23 80 55 28  0
#> 33  7 60 28  0
#> 43 80 60 28  0
#> 53  7 55 72  0
#> 63 80 55 72  0
#> 72  7 60 72  0
#> 24 80 28 35  0
#> 34  9 60 35  0
#> 44 80 60 35  0
#> 54  9 28 72  0
#> 64 80 28 72  0
#> 73  9 60 72  0
#> 25 80 50 15  0
#> 35 16 60 15  0
#> 45 80 60 15  0
#> 55 16 50 72  0
#> 65 80 50 72  0
#> 74 16 60 72  0
#> 26 80  2 51  0
#> 36 40 60 51  0
#> 46 80 60 51  0
#> 56 40  2 72  0
#> 66 80  2 72  0
#> 75 40 60 72  0
#> 27 80  6 36  0
#> 37 66 60 36  0
#> 47 80 60 36  0
#> 57 66  6 72  0
#> 67 80  6 72  0
#> 76 66 60 72  0
#> 28 80 35 10  0
#> 38 75 60 10  0
#> 48 80 60 10  0
#> 58 75 35 72  0
#> 68 80 35 72  0
#> 77 75 60 72  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 6   4 46 67  1
#> 10  5 34 42  1
#> 5   7 55 28  1
#> 8   9 28 35  1
#> 3  16 50 15  1
#> 1  40  2 51  1
#> 4  66  6 36  1
#> 2  75 35 10  1
#> 21 80 46 67  0
#> 31  4 -3 67  0
#> 41 80 -3 67  0
#> 51  4 46 72  0
#> 61 80 46 72  0
#> 7   4 -3 72  0
#> 81 80 -3 72  0
#> 22 80 34 42  0
#> 32  5 -3 42  0
#> 42 80 -3 42  0
#> 52  5 34 72  0
#> 62 80 34 72  0
#> 71  5 -3 72  0
#> 23 80 55 28  0
#> 33  7 -3 28  0
#> 43 80 -3 28  0
#> 53  7 55 72  0
#> 63 80 55 72  0
#> 72  7 -3 72  0
#> 24 80 28 35  0
#> 34  9 -3 35  0
#> 44 80 -3 35  0
#> 54  9 28 72  0
#> 64 80 28 72  0
#> 73  9 -3 72  0
#> 25 80 50 15  0
#> 35 16 -3 15  0
#> 45 80 -3 15  0
#> 55 16 50 72  0
#> 65 80 50 72  0
#> 74 16 -3 72  0
#> 26 80  2 51  0
#> 36 40 -3 51  0
#> 46 80 -3 51  0
#> 56 40  2 72  0
#> 66 80  2 72  0
#> 75 40 -3 72  0
#> 27 80  6 36  0
#> 37 66 -3 36  0
#> 47 80 -3 36  0
#> 57 66  6 72  0
#> 67 80  6 72  0
#> 76 66 -3 72  0
#> 28 80 35 10  0
#> 38 75 -3 10  0
#> 48 80 -3 10  0
#> 58 75 35 72  0
#> 68 80 35 72  0
#> 77 75 -3 72  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 6   4 46  67  1
#> 10  5 34  42  1
#> 5   7 55  28  1
#> 8   9 28  35  1
#> 3  16 50  15  1
#> 1  40  2  51  1
#> 4  66  6  36  1
#> 2  75 35  10  1
#> 21  0 46  67  0
#> 31  4  0  67  0
#> 41  0  0  67  0
#> 51  4 46 100  0
#> 61  0 46 100  0
#> 7   4  0 100  0
#> 81  0  0 100  0
#> 22  0 34  42  0
#> 32  5  0  42  0
#> 42  0  0  42  0
#> 52  5 34 100  0
#> 62  0 34 100  0
#> 71  5  0 100  0
#> 23  0 55  28  0
#> 33  7  0  28  0
#> 43  0  0  28  0
#> 53  7 55 100  0
#> 63  0 55 100  0
#> 72  7  0 100  0
#> 24  0 28  35  0
#> 34  9  0  35  0
#> 44  0  0  35  0
#> 54  9 28 100  0
#> 64  0 28 100  0
#> 73  9  0 100  0
#> 25  0 50  15  0
#> 35 16  0  15  0
#> 45  0  0  15  0
#> 55 16 50 100  0
#> 65  0 50 100  0
#> 74 16  0 100  0
#> 26  0  2  51  0
#> 36 40  0  51  0
#> 46  0  0  51  0
#> 56 40  2 100  0
#> 66  0  2 100  0
#> 75 40  0 100  0
#> 27  0  6  36  0
#> 37 66  0  36  0
#> 47  0  0  36  0
#> 57 66  6 100  0
#> 67  0  6 100  0
#> 76 66  0 100  0
#> 28  0 35  10  0
#> 38 75  0  10  0
#> 48  0  0  10  0
#> 58 75 35 100  0
#> 68  0 35 100  0
#> 77 75  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     V1 V2 V3 V4 V5 pt
#> 1   56 13 48 20 60  1
#> 2   35 91 44 49 28  1
#> 3   48 14 54 44 83  1
#> 4   20 17 32 50 45  1
#> 5   21 67 51 61 16  1
#> 6   61 13 38 60 24  1
#> 7   19 65 17 59 60  1
#> 8   76 52 20 21 54  1
#> 9   34 30 34 16 33  1
#> 10  52 30 11 59 31  1
#> 11  54 29 15 32 31  1
#> 12  20 35 37 30 25  1
#> 13  20 41 25 26 64  1
#> 14  12 41 23 50 59  1
#> 15   6 39 48 55 34  1
#> 16  44 14 56 22 66  1
#> 17  33 12 33 33 61  1
#> 18  63 30 69 17 71  1
#> 19  41 65 11 28 43  1
#> 20   5 62 66 49 57  1
#> 22  81 13 48 20 60  0
#> 23  56 96 48 20 60  0
#> 24  81 96 48 20 60  0
#> 25  56 13 74 20 60  0
#> 26  81 13 74 20 60  0
#> 27  56 96 74 20 60  0
#> 28  81 96 74 20 60  0
#> 29  56 13 48 66 60  0
#> 30  81 13 48 66 60  0
#> 31  56 96 48 66 60  0
#> 32  81 96 48 66 60  0
#> 33  56 13 74 66 60  0
#> 34  81 13 74 66 60  0
#> 35  56 96 74 66 60  0
#> 36  81 96 74 66 60  0
#> 37  56 13 48 20 88  0
#> 38  81 13 48 20 88  0
#> 39  56 96 48 20 88  0
#> 40  81 96 48 20 88  0
#> 41  56 13 74 20 88  0
#> 42  81 13 74 20 88  0
#> 43  56 96 74 20 88  0
#> 44  81 96 74 20 88  0
#> 45  56 13 48 66 88  0
#> 46  81 13 48 66 88  0
#> 47  56 96 48 66 88  0
#> 48  81 96 48 66 88  0
#> 49  56 13 74 66 88  0
#> 50  81 13 74 66 88  0
#> 51  56 96 74 66 88  0
#> 52  81 96 74 66 88  0
#> 54  81 91 44 49 28  0
#> 55  35 96 44 49 28  0
#> 56  81 96 44 49 28  0
#> 57  35 91 74 49 28  0
#> 58  81 91 74 49 28  0
#> 59  35 96 74 49 28  0
#> 60  81 96 74 49 28  0
#> 61  35 91 44 66 28  0
#> 62  81 91 44 66 28  0
#> 63  35 96 44 66 28  0
#> 64  81 96 44 66 28  0
#> 65  35 91 74 66 28  0
#> 66  81 91 74 66 28  0
#> 67  35 96 74 66 28  0
#> 68  81 96 74 66 28  0
#> 69  35 91 44 49 88  0
#> 70  81 91 44 49 88  0
#> 71  35 96 44 49 88  0
#> 72  81 96 44 49 88  0
#> 73  35 91 74 49 88  0
#> 74  81 91 74 49 88  0
#> 75  35 96 74 49 88  0
#> 76  81 96 74 49 88  0
#> 77  35 91 44 66 88  0
#> 78  81 91 44 66 88  0
#> 79  35 96 44 66 88  0
#> 80  81 96 44 66 88  0
#> 81  35 91 74 66 88  0
#> 82  81 91 74 66 88  0
#> 83  35 96 74 66 88  0
#> 86  81 14 54 44 83  0
#> 87  48 96 54 44 83  0
#> 88  81 96 54 44 83  0
#> 89  48 14 74 44 83  0
#> 90  81 14 74 44 83  0
#> 91  48 96 74 44 83  0
#> 92  81 96 74 44 83  0
#> 93  48 14 54 66 83  0
#> 94  81 14 54 66 83  0
#> 95  48 96 54 66 83  0
#> 96  81 96 54 66 83  0
#> 97  48 14 74 66 83  0
#> 98  81 14 74 66 83  0
#> 99  48 96 74 66 83  0
#> 100 81 96 74 66 83  0
#> 101 48 14 54 44 88  0
#> 102 81 14 54 44 88  0
#> 103 48 96 54 44 88  0
#> 104 81 96 54 44 88  0
#> 105 48 14 74 44 88  0
#> 106 81 14 74 44 88  0
#> 107 48 96 74 44 88  0
#> 108 81 96 74 44 88  0
#> 109 48 14 54 66 88  0
#> 110 81 14 54 66 88  0
#> 111 48 96 54 66 88  0
#> 112 81 96 54 66 88  0
#> 113 48 14 74 66 88  0
#> 114 81 14 74 66 88  0
#> 115 48 96 74 66 88  0
#> 118 81 17 32 50 45  0
#> 119 20 96 32 50 45  0
#> 120 81 96 32 50 45  0
#> 121 20 17 74 50 45  0
#> 122 81 17 74 50 45  0
#> 123 20 96 74 50 45  0
#> 124 81 96 74 50 45  0
#> 125 20 17 32 66 45  0
#> 126 81 17 32 66 45  0
#> 127 20 96 32 66 45  0
#> 128 81 96 32 66 45  0
#> 129 20 17 74 66 45  0
#> 130 81 17 74 66 45  0
#> 131 20 96 74 66 45  0
#> 132 81 96 74 66 45  0
#> 133 20 17 32 50 88  0
#> 134 81 17 32 50 88  0
#> 135 20 96 32 50 88  0
#> 136 81 96 32 50 88  0
#> 137 20 17 74 50 88  0
#> 138 81 17 74 50 88  0
#> 139 20 96 74 50 88  0
#> 140 81 96 74 50 88  0
#> 141 20 17 32 66 88  0
#> 142 81 17 32 66 88  0
#> 143 20 96 32 66 88  0
#> 144 81 96 32 66 88  0
#> 145 20 17 74 66 88  0
#> 146 81 17 74 66 88  0
#> 147 20 96 74 66 88  0
#> 150 81 67 51 61 16  0
#> 151 21 96 51 61 16  0
#> 152 81 96 51 61 16  0
#> 153 21 67 74 61 16  0
#> 154 81 67 74 61 16  0
#> 155 21 96 74 61 16  0
#> 156 81 96 74 61 16  0
#> 157 21 67 51 66 16  0
#> 158 81 67 51 66 16  0
#> 159 21 96 51 66 16  0
#> 160 81 96 51 66 16  0
#> 161 21 67 74 66 16  0
#> 162 81 67 74 66 16  0
#> 163 21 96 74 66 16  0
#> 164 81 96 74 66 16  0
#> 165 21 67 51 61 88  0
#> 166 81 67 51 61 88  0
#> 167 21 96 51 61 88  0
#> 168 81 96 51 61 88  0
#> 169 21 67 74 61 88  0
#> 170 81 67 74 61 88  0
#> 171 21 96 74 61 88  0
#> 172 81 96 74 61 88  0
#> 173 21 67 51 66 88  0
#> 174 81 67 51 66 88  0
#> 175 21 96 51 66 88  0
#> 176 81 96 51 66 88  0
#> 177 21 67 74 66 88  0
#> 178 81 67 74 66 88  0
#> 179 21 96 74 66 88  0
#> 182 81 13 38 60 24  0
#> 183 61 96 38 60 24  0
#> 184 81 96 38 60 24  0
#> 185 61 13 74 60 24  0
#> 186 81 13 74 60 24  0
#> 187 61 96 74 60 24  0
#> 188 81 96 74 60 24  0
#> 189 61 13 38 66 24  0
#> 190 81 13 38 66 24  0
#> 191 61 96 38 66 24  0
#> 192 81 96 38 66 24  0
#> 193 61 13 74 66 24  0
#> 194 81 13 74 66 24  0
#> 195 61 96 74 66 24  0
#> 196 81 96 74 66 24  0
#> 197 61 13 38 60 88  0
#> 198 81 13 38 60 88  0
#> 199 61 96 38 60 88  0
#> 200 81 96 38 60 88  0
#> 201 61 13 74 60 88  0
#> 202 81 13 74 60 88  0
#> 203 61 96 74 60 88  0
#> 204 81 96 74 60 88  0
#> 205 61 13 38 66 88  0
#> 206 81 13 38 66 88  0
#> 207 61 96 38 66 88  0
#> 208 81 96 38 66 88  0
#> 209 61 13 74 66 88  0
#> 211 61 96 74 66 88  0
#> 214 81 65 17 59 60  0
#> 215 19 96 17 59 60  0
#> 216 81 96 17 59 60  0
#> 217 19 65 74 59 60  0
#> 218 81 65 74 59 60  0
#> 219 19 96 74 59 60  0
#> 220 81 96 74 59 60  0
#> 221 19 65 17 66 60  0
#> 222 81 65 17 66 60  0
#> 223 19 96 17 66 60  0
#> 224 81 96 17 66 60  0
#> 225 19 65 74 66 60  0
#> 226 81 65 74 66 60  0
#> 227 19 96 74 66 60  0
#> 229 19 65 17 59 88  0
#> 230 81 65 17 59 88  0
#> 231 19 96 17 59 88  0
#> 232 81 96 17 59 88  0
#> 233 19 65 74 59 88  0
#> 234 81 65 74 59 88  0
#> 235 19 96 74 59 88  0
#> 236 81 96 74 59 88  0
#> 237 19 65 17 66 88  0
#> 238 81 65 17 66 88  0
#> 239 19 96 17 66 88  0
#> 240 81 96 17 66 88  0
#> 241 19 65 74 66 88  0
#> 242 81 65 74 66 88  0
#> 243 19 96 74 66 88  0
#> 246 81 52 20 21 54  0
#> 247 76 96 20 21 54  0
#> 248 81 96 20 21 54  0
#> 249 76 52 74 21 54  0
#> 250 81 52 74 21 54  0
#> 251 76 96 74 21 54  0
#> 252 81 96 74 21 54  0
#> 253 76 52 20 66 54  0
#> 254 81 52 20 66 54  0
#> 255 76 96 20 66 54  0
#> 256 81 96 20 66 54  0
#> 257 76 52 74 66 54  0
#> 258 81 52 74 66 54  0
#> 259 76 96 74 66 54  0
#> 260 81 96 74 66 54  0
#> 261 76 52 20 21 88  0
#> 262 81 52 20 21 88  0
#> 263 76 96 20 21 88  0
#> 264 81 96 20 21 88  0
#> 265 76 52 74 21 88  0
#> 266 81 52 74 21 88  0
#> 267 76 96 74 21 88  0
#> 268 81 96 74 21 88  0
#> 269 76 52 20 66 88  0
#> 270 81 52 20 66 88  0
#> 271 76 96 20 66 88  0
#> 272 81 96 20 66 88  0
#> 273 76 52 74 66 88  0
#> 274 81 52 74 66 88  0
#> 275 76 96 74 66 88  0
#> 278 81 30 34 16 33  0
#> 279 34 96 34 16 33  0
#> 280 81 96 34 16 33  0
#> 281 34 30 74 16 33  0
#> 282 81 30 74 16 33  0
#> 283 34 96 74 16 33  0
#> 284 81 96 74 16 33  0
#> 285 34 30 34 66 33  0
#> 286 81 30 34 66 33  0
#> 287 34 96 34 66 33  0
#> 288 81 96 34 66 33  0
#> 289 34 30 74 66 33  0
#> 290 81 30 74 66 33  0
#> 291 34 96 74 66 33  0
#> 292 81 96 74 66 33  0
#> 293 34 30 34 16 88  0
#> 294 81 30 34 16 88  0
#> 295 34 96 34 16 88  0
#> 296 81 96 34 16 88  0
#> 297 34 30 74 16 88  0
#> 298 81 30 74 16 88  0
#> 299 34 96 74 16 88  0
#> 300 81 96 74 16 88  0
#> 301 34 30 34 66 88  0
#> 302 81 30 34 66 88  0
#> 303 34 96 34 66 88  0
#> 304 81 96 34 66 88  0
#> 305 34 30 74 66 88  0
#> 306 81 30 74 66 88  0
#> 307 34 96 74 66 88  0
#> 310 81 30 11 59 31  0
#> 311 52 96 11 59 31  0
#> 312 81 96 11 59 31  0
#> 313 52 30 74 59 31  0
#> 314 81 30 74 59 31  0
#> 315 52 96 74 59 31  0
#> 316 81 96 74 59 31  0
#> 317 52 30 11 66 31  0
#> 318 81 30 11 66 31  0
#> 319 52 96 11 66 31  0
#> 320 81 96 11 66 31  0
#> 321 52 30 74 66 31  0
#> 322 81 30 74 66 31  0
#> 323 52 96 74 66 31  0
#> 324 81 96 74 66 31  0
#> 325 52 30 11 59 88  0
#> 326 81 30 11 59 88  0
#> 327 52 96 11 59 88  0
#> 328 81 96 11 59 88  0
#> 329 52 30 74 59 88  0
#> 330 81 30 74 59 88  0
#> 331 52 96 74 59 88  0
#> 333 52 30 11 66 88  0
#> 334 81 30 11 66 88  0
#> 335 52 96 11 66 88  0
#> 336 81 96 11 66 88  0
#> 337 52 30 74 66 88  0
#> 339 52 96 74 66 88  0
#> 342 81 29 15 32 31  0
#> 343 54 96 15 32 31  0
#> 344 81 96 15 32 31  0
#> 345 54 29 74 32 31  0
#> 346 81 29 74 32 31  0
#> 347 54 96 74 32 31  0
#> 348 81 96 74 32 31  0
#> 349 54 29 15 66 31  0
#> 350 81 29 15 66 31  0
#> 351 54 96 15 66 31  0
#> 352 81 96 15 66 31  0
#> 353 54 29 74 66 31  0
#> 354 81 29 74 66 31  0
#> 355 54 96 74 66 31  0
#> 357 54 29 15 32 88  0
#> 358 81 29 15 32 88  0
#> 359 54 96 15 32 88  0
#> 360 81 96 15 32 88  0
#> 361 54 29 74 32 88  0
#> 362 81 29 74 32 88  0
#> 363 54 96 74 32 88  0
#> 364 81 96 74 32 88  0
#> 365 54 29 15 66 88  0
#> 366 81 29 15 66 88  0
#> 367 54 96 15 66 88  0
#> 368 81 96 15 66 88  0
#> 369 54 29 74 66 88  0
#> 370 81 29 74 66 88  0
#> 371 54 96 74 66 88  0
#> 374 81 35 37 30 25  0
#> 375 20 96 37 30 25  0
#> 376 81 96 37 30 25  0
#> 377 20 35 74 30 25  0
#> 378 81 35 74 30 25  0
#> 379 20 96 74 30 25  0
#> 380 81 96 74 30 25  0
#> 381 20 35 37 66 25  0
#> 382 81 35 37 66 25  0
#> 383 20 96 37 66 25  0
#> 384 81 96 37 66 25  0
#> 385 20 35 74 66 25  0
#> 386 81 35 74 66 25  0
#> 387 20 96 74 66 25  0
#> 388 81 96 74 66 25  0
#> 389 20 35 37 30 88  0
#> 390 81 35 37 30 88  0
#> 391 20 96 37 30 88  0
#> 392 81 96 37 30 88  0
#> 393 20 35 74 30 88  0
#> 394 81 35 74 30 88  0
#> 395 20 96 74 30 88  0
#> 396 81 96 74 30 88  0
#> 397 20 35 37 66 88  0
#> 398 81 35 37 66 88  0
#> 399 20 96 37 66 88  0
#> 400 81 96 37 66 88  0
#> 401 20 35 74 66 88  0
#> 402 81 35 74 66 88  0
#> 406 81 41 25 26 64  0
#> 407 20 96 25 26 64  0
#> 408 81 96 25 26 64  0
#> 409 20 41 74 26 64  0
#> 410 81 41 74 26 64  0
#> 411 20 96 74 26 64  0
#> 412 81 96 74 26 64  0
#> 413 20 41 25 66 64  0
#> 414 81 41 25 66 64  0
#> 415 20 96 25 66 64  0
#> 416 81 96 25 66 64  0
#> 417 20 41 74 66 64  0
#> 418 81 41 74 66 64  0
#> 419 20 96 74 66 64  0
#> 420 81 96 74 66 64  0
#> 421 20 41 25 26 88  0
#> 422 81 41 25 26 88  0
#> 423 20 96 25 26 88  0
#> 424 81 96 25 26 88  0
#> 425 20 41 74 26 88  0
#> 426 81 41 74 26 88  0
#> 427 20 96 74 26 88  0
#> 428 81 96 74 26 88  0
#> 429 20 41 25 66 88  0
#> 430 81 41 25 66 88  0
#> 431 20 96 25 66 88  0
#> 432 81 96 25 66 88  0
#> 433 20 41 74 66 88  0
#> 434 81 41 74 66 88  0
#> 438 81 41 23 50 59  0
#> 439 12 96 23 50 59  0
#> 440 81 96 23 50 59  0
#> 441 12 41 74 50 59  0
#> 442 81 41 74 50 59  0
#> 443 12 96 74 50 59  0
#> 444 81 96 74 50 59  0
#> 445 12 41 23 66 59  0
#> 446 81 41 23 66 59  0
#> 447 12 96 23 66 59  0
#> 448 81 96 23 66 59  0
#> 449 12 41 74 66 59  0
#> 450 81 41 74 66 59  0
#> 451 12 96 74 66 59  0
#> 452 81 96 74 66 59  0
#> 453 12 41 23 50 88  0
#> 454 81 41 23 50 88  0
#> 455 12 96 23 50 88  0
#> 456 81 96 23 50 88  0
#> 457 12 41 74 50 88  0
#> 458 81 41 74 50 88  0
#> 459 12 96 74 50 88  0
#> 461 12 41 23 66 88  0
#> 462 81 41 23 66 88  0
#> 463 12 96 23 66 88  0
#> 464 81 96 23 66 88  0
#> 465 12 41 74 66 88  0
#> 467 12 96 74 66 88  0
#> 470 81 39 48 55 34  0
#> 471  6 96 48 55 34  0
#> 472 81 96 48 55 34  0
#> 473  6 39 74 55 34  0
#> 474 81 39 74 55 34  0
#> 475  6 96 74 55 34  0
#> 476 81 96 74 55 34  0
#> 477  6 39 48 66 34  0
#> 478 81 39 48 66 34  0
#> 479  6 96 48 66 34  0
#> 480 81 96 48 66 34  0
#> 481  6 39 74 66 34  0
#> 482 81 39 74 66 34  0
#> 483  6 96 74 66 34  0
#> 484 81 96 74 66 34  0
#> 485  6 39 48 55 88  0
#> 486 81 39 48 55 88  0
#> 487  6 96 48 55 88  0
#> 488 81 96 48 55 88  0
#> 489  6 39 74 55 88  0
#> 490 81 39 74 55 88  0
#> 491  6 96 74 55 88  0
#> 492 81 96 74 55 88  0
#> 493  6 39 48 66 88  0
#> 494 81 39 48 66 88  0
#> 495  6 96 48 66 88  0
#> 497  6 39 74 66 88  0
#> 498 81 39 74 66 88  0
#> 499  6 96 74 66 88  0
#> 502 81 14 56 22 66  0
#> 503 44 96 56 22 66  0
#> 504 81 96 56 22 66  0
#> 505 44 14 74 22 66  0
#> 506 81 14 74 22 66  0
#> 507 44 96 74 22 66  0
#> 508 81 96 74 22 66  0
#> 509 44 14 56 66 66  0
#> 510 81 14 56 66 66  0
#> 511 44 96 56 66 66  0
#> 512 81 96 56 66 66  0
#> 513 44 14 74 66 66  0
#> 514 81 14 74 66 66  0
#> 515 44 96 74 66 66  0
#> 516 81 96 74 66 66  0
#> 517 44 14 56 22 88  0
#> 518 81 14 56 22 88  0
#> 519 44 96 56 22 88  0
#> 520 81 96 56 22 88  0
#> 521 44 14 74 22 88  0
#> 522 81 14 74 22 88  0
#> 523 44 96 74 22 88  0
#> 524 81 96 74 22 88  0
#> 525 44 14 56 66 88  0
#> 526 81 14 56 66 88  0
#> 527 44 96 56 66 88  0
#> 528 81 96 56 66 88  0
#> 529 44 14 74 66 88  0
#> 531 44 96 74 66 88  0
#> 534 81 12 33 33 61  0
#> 535 33 96 33 33 61  0
#> 536 81 96 33 33 61  0
#> 537 33 12 74 33 61  0
#> 538 81 12 74 33 61  0
#> 539 33 96 74 33 61  0
#> 540 81 96 74 33 61  0
#> 541 33 12 33 66 61  0
#> 542 81 12 33 66 61  0
#> 543 33 96 33 66 61  0
#> 544 81 96 33 66 61  0
#> 545 33 12 74 66 61  0
#> 546 81 12 74 66 61  0
#> 547 33 96 74 66 61  0
#> 548 81 96 74 66 61  0
#> 549 33 12 33 33 88  0
#> 550 81 12 33 33 88  0
#> 551 33 96 33 33 88  0
#> 552 81 96 33 33 88  0
#> 553 33 12 74 33 88  0
#> 554 81 12 74 33 88  0
#> 555 33 96 74 33 88  0
#> 556 81 96 74 33 88  0
#> 557 33 12 33 66 88  0
#> 558 81 12 33 66 88  0
#> 559 33 96 33 66 88  0
#> 560 81 96 33 66 88  0
#> 561 33 12 74 66 88  0
#> 562 81 12 74 66 88  0
#> 563 33 96 74 66 88  0
#> 566 81 30 69 17 71  0
#> 567 63 96 69 17 71  0
#> 568 81 96 69 17 71  0
#> 569 63 30 74 17 71  0
#> 570 81 30 74 17 71  0
#> 571 63 96 74 17 71  0
#> 572 81 96 74 17 71  0
#> 573 63 30 69 66 71  0
#> 574 81 30 69 66 71  0
#> 575 63 96 69 66 71  0
#> 576 81 96 69 66 71  0
#> 577 63 30 74 66 71  0
#> 578 81 30 74 66 71  0
#> 579 63 96 74 66 71  0
#> 580 81 96 74 66 71  0
#> 581 63 30 69 17 88  0
#> 582 81 30 69 17 88  0
#> 583 63 96 69 17 88  0
#> 584 81 96 69 17 88  0
#> 585 63 30 74 17 88  0
#> 586 81 30 74 17 88  0
#> 587 63 96 74 17 88  0
#> 588 81 96 74 17 88  0
#> 589 63 30 69 66 88  0
#> 590 81 30 69 66 88  0
#> 591 63 96 69 66 88  0
#> 592 81 96 69 66 88  0
#> 593 63 30 74 66 88  0
#> 595 63 96 74 66 88  0
#> 598 81 65 11 28 43  0
#> 599 41 96 11 28 43  0
#> 600 81 96 11 28 43  0
#> 601 41 65 74 28 43  0
#> 602 81 65 74 28 43  0
#> 603 41 96 74 28 43  0
#> 604 81 96 74 28 43  0
#> 605 41 65 11 66 43  0
#> 606 81 65 11 66 43  0
#> 607 41 96 11 66 43  0
#> 608 81 96 11 66 43  0
#> 609 41 65 74 66 43  0
#> 610 81 65 74 66 43  0
#> 611 41 96 74 66 43  0
#> 612 81 96 74 66 43  0
#> 613 41 65 11 28 88  0
#> 614 81 65 11 28 88  0
#> 615 41 96 11 28 88  0
#> 616 81 96 11 28 88  0
#> 617 41 65 74 28 88  0
#> 618 81 65 74 28 88  0
#> 619 41 96 74 28 88  0
#> 620 81 96 74 28 88  0
#> 621 41 65 11 66 88  0
#> 622 81 65 11 66 88  0
#> 623 41 96 11 66 88  0
#> 625 41 65 74 66 88  0
#> 627 41 96 74 66 88  0
#> 630 81 62 66 49 57  0
#> 631  5 96 66 49 57  0
#> 632 81 96 66 49 57  0
#> 633  5 62 74 49 57  0
#> 634 81 62 74 49 57  0
#> 635  5 96 74 49 57  0
#> 636 81 96 74 49 57  0
#> 637  5 62 66 66 57  0
#> 638 81 62 66 66 57  0
#> 639  5 96 66 66 57  0
#> 640 81 96 66 66 57  0
#> 641  5 62 74 66 57  0
#> 642 81 62 74 66 57  0
#> 643  5 96 74 66 57  0
#> 644 81 96 74 66 57  0
#> 645  5 62 66 49 88  0
#> 646 81 62 66 49 88  0
#> 647  5 96 66 49 88  0
#> 648 81 96 66 49 88  0
#> 649  5 62 74 49 88  0
#> 650 81 62 74 49 88  0
#> 651  5 96 74 49 88  0
#> 653  5 62 66 66 88  0
#> 654 81 62 66 66 88  0
#> 655  5 96 66 66 88  0
#> 656 81 96 66 66 88  0
#> 657  5 62 74 66 88  0
#> 658 81 62 74 66 88  0
#> 659  5 96 74 66 88  0