Add all points on the bounding box hit by the rays.

Add all points on the bounding box hit by the rays.

Usage

addRays(
  pts,
  m = apply(pts, 2, min) - 5,
  M = apply(pts, 2, max) + 5,
  direction = 1
)

Arguments

pts

A data frame with all points

m

Minimum values of the bounding box.

M

Maximum values of the bounding box.

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero. If negative, consider the i'th column of the pts minus a value greater than on equal zero.

Value

The points merged with the points on the bounding box. The column pt equals 1 if points from pts and zero otherwise.

Note

Assume that pts has been checked using .checkPts().

Examples

pts <- genNDSet(3,10)[,1:3]
addRays(pts)
#>    z1 z2 z3 pt
#> 1  18 24 25  1
#> 2  50 27  7  1
#> 3  35 30  8  1
#> 4   4 59 36  1
#> 5  27 33 11  1
#> 6  13 19 53  1
#> 7   2 43 43  1
#> 8  53  5 71  1
#> 9  41 70  6  1
#> 10  3 37 49  1
#> 12 58 24 25  0
#> 13 18 75 25  0
#> 14 58 75 25  0
#> 15 18 24 76  0
#> 16 58 24 76  0
#> 17 18 75 76  0
#> 18 58 75 76  0
#> 20 58 27  7  0
#> 21 50 75  7  0
#> 22 58 75  7  0
#> 23 50 27 76  0
#> 24 58 27 76  0
#> 25 50 75 76  0
#> 28 58 30  8  0
#> 29 35 75  8  0
#> 30 58 75  8  0
#> 31 35 30 76  0
#> 32 58 30 76  0
#> 33 35 75 76  0
#> 36 58 59 36  0
#> 37  4 75 36  0
#> 38 58 75 36  0
#> 39  4 59 76  0
#> 40 58 59 76  0
#> 41  4 75 76  0
#> 44 58 33 11  0
#> 45 27 75 11  0
#> 46 58 75 11  0
#> 47 27 33 76  0
#> 48 58 33 76  0
#> 49 27 75 76  0
#> 52 58 19 53  0
#> 53 13 75 53  0
#> 54 58 75 53  0
#> 55 13 19 76  0
#> 56 58 19 76  0
#> 57 13 75 76  0
#> 60 58 43 43  0
#> 61  2 75 43  0
#> 62 58 75 43  0
#> 63  2 43 76  0
#> 64 58 43 76  0
#> 65  2 75 76  0
#> 68 58  5 71  0
#> 69 53 75 71  0
#> 70 58 75 71  0
#> 71 53  5 76  0
#> 72 58  5 76  0
#> 73 53 75 76  0
#> 76 58 70  6  0
#> 77 41 75  6  0
#> 78 58 75  6  0
#> 79 41 70 76  0
#> 80 58 70 76  0
#> 81 41 75 76  0
#> 84 58 37 49  0
#> 85  3 75 49  0
#> 86 58 75 49  0
#> 87  3 37 76  0
#> 88 58 37 76  0
#> 89  3 75 76  0
addRays(pts, dir = c(1,-1,1))
#>    z1 z2 z3 pt
#> 1  18 24 25  1
#> 2  50 27  7  1
#> 3  35 30  8  1
#> 4   4 59 36  1
#> 5  27 33 11  1
#> 6  13 19 53  1
#> 7   2 43 43  1
#> 8  53  5 71  1
#> 9  41 70  6  1
#> 10  3 37 49  1
#> 12 58 24 25  0
#> 13 18  0 25  0
#> 14 58  0 25  0
#> 15 18 24 76  0
#> 16 58 24 76  0
#> 17 18  0 76  0
#> 18 58  0 76  0
#> 20 58 27  7  0
#> 21 50  0  7  0
#> 22 58  0  7  0
#> 23 50 27 76  0
#> 24 58 27 76  0
#> 25 50  0 76  0
#> 28 58 30  8  0
#> 29 35  0  8  0
#> 30 58  0  8  0
#> 31 35 30 76  0
#> 32 58 30 76  0
#> 33 35  0 76  0
#> 36 58 59 36  0
#> 37  4  0 36  0
#> 38 58  0 36  0
#> 39  4 59 76  0
#> 40 58 59 76  0
#> 41  4  0 76  0
#> 44 58 33 11  0
#> 45 27  0 11  0
#> 46 58  0 11  0
#> 47 27 33 76  0
#> 48 58 33 76  0
#> 49 27  0 76  0
#> 52 58 19 53  0
#> 53 13  0 53  0
#> 54 58  0 53  0
#> 55 13 19 76  0
#> 56 58 19 76  0
#> 57 13  0 76  0
#> 60 58 43 43  0
#> 61  2  0 43  0
#> 62 58  0 43  0
#> 63  2 43 76  0
#> 64 58 43 76  0
#> 65  2  0 76  0
#> 68 58  5 71  0
#> 69 53  0 71  0
#> 70 58  0 71  0
#> 71 53  5 76  0
#> 72 58  5 76  0
#> 73 53  0 76  0
#> 76 58 70  6  0
#> 77 41  0  6  0
#> 78 58  0  6  0
#> 79 41 70 76  0
#> 80 58 70 76  0
#> 81 41  0 76  0
#> 84 58 37 49  0
#> 85  3  0 49  0
#> 86 58  0 49  0
#> 87  3 37 76  0
#> 88 58 37 76  0
#> 89  3  0 76  0
addRays(pts, dir = c(-1,-1,1), m = c(0,0,0), M = c(100,100,100))
#>    z1 z2  z3 pt
#> 1  18 24  25  1
#> 2  50 27   7  1
#> 3  35 30   8  1
#> 4   4 59  36  1
#> 5  27 33  11  1
#> 6  13 19  53  1
#> 7   2 43  43  1
#> 8  53  5  71  1
#> 9  41 70   6  1
#> 10  3 37  49  1
#> 12  0 24  25  0
#> 13 18  0  25  0
#> 14  0  0  25  0
#> 15 18 24 100  0
#> 16  0 24 100  0
#> 17 18  0 100  0
#> 18  0  0 100  0
#> 20  0 27   7  0
#> 21 50  0   7  0
#> 22  0  0   7  0
#> 23 50 27 100  0
#> 24  0 27 100  0
#> 25 50  0 100  0
#> 28  0 30   8  0
#> 29 35  0   8  0
#> 30  0  0   8  0
#> 31 35 30 100  0
#> 32  0 30 100  0
#> 33 35  0 100  0
#> 36  0 59  36  0
#> 37  4  0  36  0
#> 38  0  0  36  0
#> 39  4 59 100  0
#> 40  0 59 100  0
#> 41  4  0 100  0
#> 44  0 33  11  0
#> 45 27  0  11  0
#> 46  0  0  11  0
#> 47 27 33 100  0
#> 48  0 33 100  0
#> 49 27  0 100  0
#> 52  0 19  53  0
#> 53 13  0  53  0
#> 54  0  0  53  0
#> 55 13 19 100  0
#> 56  0 19 100  0
#> 57 13  0 100  0
#> 60  0 43  43  0
#> 61  2  0  43  0
#> 62  0  0  43  0
#> 63  2 43 100  0
#> 64  0 43 100  0
#> 65  2  0 100  0
#> 68  0  5  71  0
#> 69 53  0  71  0
#> 70  0  0  71  0
#> 71 53  5 100  0
#> 72  0  5 100  0
#> 73 53  0 100  0
#> 76  0 70   6  0
#> 77 41  0   6  0
#> 78  0  0   6  0
#> 79 41 70 100  0
#> 80  0 70 100  0
#> 81 41  0 100  0
#> 84  0 37  49  0
#> 85  3  0  49  0
#> 86  0  0  49  0
#> 87  3 37 100  0
#> 88  0 37 100  0
#> 89  3  0 100  0
pts <- genSample(5,20)[,1:5]
addRays(pts)
#>     z1 z2 z3 z4 z5 pt
#> 1   36 35 68 39 11  1
#> 2   47 60 84 34 19  1
#> 3   63  6 61 41 62  1
#> 4   45 64 24 69 16  1
#> 5   77 11 59 40 53  1
#> 6    7 38 60 46 67  1
#> 7   11 65 32 44 33  1
#> 8   41 79 42 50 12  1
#> 9   71 52  7 46 63  1
#> 10  40 55 38 42  5  1
#> 11  27 30 77 41 76  1
#> 12  45 76 35 11 46  1
#> 13  40 28 88 59 33  1
#> 14  82 34 47 18 61  1
#> 15  11 72 34 38 47  1
#> 16  19 64 52 16 57  1
#> 17   4 63 53 56 62  1
#> 18  11 60 53 22 51  1
#> 19  15 54 24 60 32  1
#> 20  21 69 35 26 31  1
#> 22  87 35 68 39 11  0
#> 23  36 84 68 39 11  0
#> 24  87 84 68 39 11  0
#> 25  36 35 93 39 11  0
#> 26  87 35 93 39 11  0
#> 27  36 84 93 39 11  0
#> 28  87 84 93 39 11  0
#> 29  36 35 68 74 11  0
#> 30  87 35 68 74 11  0
#> 31  36 84 68 74 11  0
#> 32  87 84 68 74 11  0
#> 33  36 35 93 74 11  0
#> 34  87 35 93 74 11  0
#> 35  36 84 93 74 11  0
#> 36  87 84 93 74 11  0
#> 37  36 35 68 39 81  0
#> 38  87 35 68 39 81  0
#> 39  36 84 68 39 81  0
#> 40  87 84 68 39 81  0
#> 41  36 35 93 39 81  0
#> 42  87 35 93 39 81  0
#> 43  36 84 93 39 81  0
#> 44  87 84 93 39 81  0
#> 45  36 35 68 74 81  0
#> 46  87 35 68 74 81  0
#> 47  36 84 68 74 81  0
#> 48  87 84 68 74 81  0
#> 49  36 35 93 74 81  0
#> 50  87 35 93 74 81  0
#> 51  36 84 93 74 81  0
#> 52  87 84 93 74 81  0
#> 54  87 60 84 34 19  0
#> 55  47 84 84 34 19  0
#> 56  87 84 84 34 19  0
#> 57  47 60 93 34 19  0
#> 58  87 60 93 34 19  0
#> 59  47 84 93 34 19  0
#> 60  87 84 93 34 19  0
#> 61  47 60 84 74 19  0
#> 62  87 60 84 74 19  0
#> 63  47 84 84 74 19  0
#> 64  87 84 84 74 19  0
#> 65  47 60 93 74 19  0
#> 66  87 60 93 74 19  0
#> 67  47 84 93 74 19  0
#> 68  87 84 93 74 19  0
#> 69  47 60 84 34 81  0
#> 70  87 60 84 34 81  0
#> 71  47 84 84 34 81  0
#> 72  87 84 84 34 81  0
#> 73  47 60 93 34 81  0
#> 74  87 60 93 34 81  0
#> 75  47 84 93 34 81  0
#> 76  87 84 93 34 81  0
#> 77  47 60 84 74 81  0
#> 78  87 60 84 74 81  0
#> 79  47 84 84 74 81  0
#> 80  87 84 84 74 81  0
#> 81  47 60 93 74 81  0
#> 82  87 60 93 74 81  0
#> 83  47 84 93 74 81  0
#> 86  87  6 61 41 62  0
#> 87  63 84 61 41 62  0
#> 88  87 84 61 41 62  0
#> 89  63  6 93 41 62  0
#> 90  87  6 93 41 62  0
#> 91  63 84 93 41 62  0
#> 92  87 84 93 41 62  0
#> 93  63  6 61 74 62  0
#> 94  87  6 61 74 62  0
#> 95  63 84 61 74 62  0
#> 96  87 84 61 74 62  0
#> 97  63  6 93 74 62  0
#> 98  87  6 93 74 62  0
#> 99  63 84 93 74 62  0
#> 100 87 84 93 74 62  0
#> 101 63  6 61 41 81  0
#> 102 87  6 61 41 81  0
#> 103 63 84 61 41 81  0
#> 104 87 84 61 41 81  0
#> 105 63  6 93 41 81  0
#> 106 87  6 93 41 81  0
#> 107 63 84 93 41 81  0
#> 108 87 84 93 41 81  0
#> 109 63  6 61 74 81  0
#> 110 87  6 61 74 81  0
#> 111 63 84 61 74 81  0
#> 112 87 84 61 74 81  0
#> 113 63  6 93 74 81  0
#> 114 87  6 93 74 81  0
#> 115 63 84 93 74 81  0
#> 118 87 64 24 69 16  0
#> 119 45 84 24 69 16  0
#> 120 87 84 24 69 16  0
#> 121 45 64 93 69 16  0
#> 122 87 64 93 69 16  0
#> 123 45 84 93 69 16  0
#> 124 87 84 93 69 16  0
#> 125 45 64 24 74 16  0
#> 126 87 64 24 74 16  0
#> 127 45 84 24 74 16  0
#> 128 87 84 24 74 16  0
#> 129 45 64 93 74 16  0
#> 130 87 64 93 74 16  0
#> 131 45 84 93 74 16  0
#> 132 87 84 93 74 16  0
#> 133 45 64 24 69 81  0
#> 134 87 64 24 69 81  0
#> 135 45 84 24 69 81  0
#> 136 87 84 24 69 81  0
#> 137 45 64 93 69 81  0
#> 138 87 64 93 69 81  0
#> 139 45 84 93 69 81  0
#> 140 87 84 93 69 81  0
#> 141 45 64 24 74 81  0
#> 142 87 64 24 74 81  0
#> 143 45 84 24 74 81  0
#> 144 87 84 24 74 81  0
#> 145 45 64 93 74 81  0
#> 146 87 64 93 74 81  0
#> 147 45 84 93 74 81  0
#> 150 87 11 59 40 53  0
#> 151 77 84 59 40 53  0
#> 152 87 84 59 40 53  0
#> 153 77 11 93 40 53  0
#> 154 87 11 93 40 53  0
#> 155 77 84 93 40 53  0
#> 156 87 84 93 40 53  0
#> 157 77 11 59 74 53  0
#> 158 87 11 59 74 53  0
#> 159 77 84 59 74 53  0
#> 160 87 84 59 74 53  0
#> 161 77 11 93 74 53  0
#> 162 87 11 93 74 53  0
#> 163 77 84 93 74 53  0
#> 164 87 84 93 74 53  0
#> 165 77 11 59 40 81  0
#> 166 87 11 59 40 81  0
#> 167 77 84 59 40 81  0
#> 168 87 84 59 40 81  0
#> 169 77 11 93 40 81  0
#> 170 87 11 93 40 81  0
#> 171 77 84 93 40 81  0
#> 172 87 84 93 40 81  0
#> 173 77 11 59 74 81  0
#> 174 87 11 59 74 81  0
#> 175 77 84 59 74 81  0
#> 176 87 84 59 74 81  0
#> 177 77 11 93 74 81  0
#> 178 87 11 93 74 81  0
#> 179 77 84 93 74 81  0
#> 182 87 38 60 46 67  0
#> 183  7 84 60 46 67  0
#> 184 87 84 60 46 67  0
#> 185  7 38 93 46 67  0
#> 186 87 38 93 46 67  0
#> 187  7 84 93 46 67  0
#> 188 87 84 93 46 67  0
#> 189  7 38 60 74 67  0
#> 190 87 38 60 74 67  0
#> 191  7 84 60 74 67  0
#> 192 87 84 60 74 67  0
#> 193  7 38 93 74 67  0
#> 194 87 38 93 74 67  0
#> 195  7 84 93 74 67  0
#> 196 87 84 93 74 67  0
#> 197  7 38 60 46 81  0
#> 198 87 38 60 46 81  0
#> 199  7 84 60 46 81  0
#> 200 87 84 60 46 81  0
#> 201  7 38 93 46 81  0
#> 202 87 38 93 46 81  0
#> 203  7 84 93 46 81  0
#> 204 87 84 93 46 81  0
#> 205  7 38 60 74 81  0
#> 206 87 38 60 74 81  0
#> 207  7 84 60 74 81  0
#> 208 87 84 60 74 81  0
#> 209  7 38 93 74 81  0
#> 210 87 38 93 74 81  0
#> 211  7 84 93 74 81  0
#> 214 87 65 32 44 33  0
#> 215 11 84 32 44 33  0
#> 216 87 84 32 44 33  0
#> 217 11 65 93 44 33  0
#> 218 87 65 93 44 33  0
#> 219 11 84 93 44 33  0
#> 220 87 84 93 44 33  0
#> 221 11 65 32 74 33  0
#> 222 87 65 32 74 33  0
#> 223 11 84 32 74 33  0
#> 224 87 84 32 74 33  0
#> 225 11 65 93 74 33  0
#> 226 87 65 93 74 33  0
#> 227 11 84 93 74 33  0
#> 228 87 84 93 74 33  0
#> 229 11 65 32 44 81  0
#> 230 87 65 32 44 81  0
#> 231 11 84 32 44 81  0
#> 232 87 84 32 44 81  0
#> 233 11 65 93 44 81  0
#> 234 87 65 93 44 81  0
#> 235 11 84 93 44 81  0
#> 236 87 84 93 44 81  0
#> 237 11 65 32 74 81  0
#> 238 87 65 32 74 81  0
#> 239 11 84 32 74 81  0
#> 240 87 84 32 74 81  0
#> 241 11 65 93 74 81  0
#> 242 87 65 93 74 81  0
#> 243 11 84 93 74 81  0
#> 246 87 79 42 50 12  0
#> 247 41 84 42 50 12  0
#> 248 87 84 42 50 12  0
#> 249 41 79 93 50 12  0
#> 250 87 79 93 50 12  0
#> 251 41 84 93 50 12  0
#> 252 87 84 93 50 12  0
#> 253 41 79 42 74 12  0
#> 254 87 79 42 74 12  0
#> 255 41 84 42 74 12  0
#> 256 87 84 42 74 12  0
#> 257 41 79 93 74 12  0
#> 258 87 79 93 74 12  0
#> 259 41 84 93 74 12  0
#> 260 87 84 93 74 12  0
#> 261 41 79 42 50 81  0
#> 262 87 79 42 50 81  0
#> 263 41 84 42 50 81  0
#> 264 87 84 42 50 81  0
#> 265 41 79 93 50 81  0
#> 266 87 79 93 50 81  0
#> 267 41 84 93 50 81  0
#> 268 87 84 93 50 81  0
#> 269 41 79 42 74 81  0
#> 270 87 79 42 74 81  0
#> 271 41 84 42 74 81  0
#> 272 87 84 42 74 81  0
#> 273 41 79 93 74 81  0
#> 274 87 79 93 74 81  0
#> 275 41 84 93 74 81  0
#> 278 87 52  7 46 63  0
#> 279 71 84  7 46 63  0
#> 280 87 84  7 46 63  0
#> 281 71 52 93 46 63  0
#> 282 87 52 93 46 63  0
#> 283 71 84 93 46 63  0
#> 284 87 84 93 46 63  0
#> 285 71 52  7 74 63  0
#> 286 87 52  7 74 63  0
#> 287 71 84  7 74 63  0
#> 288 87 84  7 74 63  0
#> 289 71 52 93 74 63  0
#> 290 87 52 93 74 63  0
#> 291 71 84 93 74 63  0
#> 292 87 84 93 74 63  0
#> 293 71 52  7 46 81  0
#> 294 87 52  7 46 81  0
#> 295 71 84  7 46 81  0
#> 296 87 84  7 46 81  0
#> 297 71 52 93 46 81  0
#> 298 87 52 93 46 81  0
#> 299 71 84 93 46 81  0
#> 301 71 52  7 74 81  0
#> 302 87 52  7 74 81  0
#> 303 71 84  7 74 81  0
#> 304 87 84  7 74 81  0
#> 305 71 52 93 74 81  0
#> 306 87 52 93 74 81  0
#> 307 71 84 93 74 81  0
#> 310 87 55 38 42  5  0
#> 311 40 84 38 42  5  0
#> 312 87 84 38 42  5  0
#> 313 40 55 93 42  5  0
#> 314 87 55 93 42  5  0
#> 315 40 84 93 42  5  0
#> 316 87 84 93 42  5  0
#> 317 40 55 38 74  5  0
#> 318 87 55 38 74  5  0
#> 319 40 84 38 74  5  0
#> 320 87 84 38 74  5  0
#> 321 40 55 93 74  5  0
#> 322 87 55 93 74  5  0
#> 323 40 84 93 74  5  0
#> 324 87 84 93 74  5  0
#> 325 40 55 38 42 81  0
#> 326 87 55 38 42 81  0
#> 327 40 84 38 42 81  0
#> 328 87 84 38 42 81  0
#> 329 40 55 93 42 81  0
#> 330 87 55 93 42 81  0
#> 331 40 84 93 42 81  0
#> 332 87 84 93 42 81  0
#> 333 40 55 38 74 81  0
#> 334 87 55 38 74 81  0
#> 335 40 84 38 74 81  0
#> 336 87 84 38 74 81  0
#> 337 40 55 93 74 81  0
#> 338 87 55 93 74 81  0
#> 339 40 84 93 74 81  0
#> 342 87 30 77 41 76  0
#> 343 27 84 77 41 76  0
#> 344 87 84 77 41 76  0
#> 345 27 30 93 41 76  0
#> 346 87 30 93 41 76  0
#> 347 27 84 93 41 76  0
#> 348 87 84 93 41 76  0
#> 349 27 30 77 74 76  0
#> 350 87 30 77 74 76  0
#> 351 27 84 77 74 76  0
#> 352 87 84 77 74 76  0
#> 353 27 30 93 74 76  0
#> 354 87 30 93 74 76  0
#> 355 27 84 93 74 76  0
#> 356 87 84 93 74 76  0
#> 357 27 30 77 41 81  0
#> 358 87 30 77 41 81  0
#> 359 27 84 77 41 81  0
#> 360 87 84 77 41 81  0
#> 361 27 30 93 41 81  0
#> 362 87 30 93 41 81  0
#> 363 27 84 93 41 81  0
#> 365 27 30 77 74 81  0
#> 366 87 30 77 74 81  0
#> 367 27 84 77 74 81  0
#> 368 87 84 77 74 81  0
#> 369 27 30 93 74 81  0
#> 370 87 30 93 74 81  0
#> 371 27 84 93 74 81  0
#> 374 87 76 35 11 46  0
#> 375 45 84 35 11 46  0
#> 376 87 84 35 11 46  0
#> 377 45 76 93 11 46  0
#> 378 87 76 93 11 46  0
#> 379 45 84 93 11 46  0
#> 380 87 84 93 11 46  0
#> 381 45 76 35 74 46  0
#> 382 87 76 35 74 46  0
#> 383 45 84 35 74 46  0
#> 384 87 84 35 74 46  0
#> 385 45 76 93 74 46  0
#> 386 87 76 93 74 46  0
#> 387 45 84 93 74 46  0
#> 388 87 84 93 74 46  0
#> 389 45 76 35 11 81  0
#> 390 87 76 35 11 81  0
#> 391 45 84 35 11 81  0
#> 392 87 84 35 11 81  0
#> 393 45 76 93 11 81  0
#> 394 87 76 93 11 81  0
#> 395 45 84 93 11 81  0
#> 396 87 84 93 11 81  0
#> 397 45 76 35 74 81  0
#> 398 87 76 35 74 81  0
#> 399 45 84 35 74 81  0
#> 400 87 84 35 74 81  0
#> 401 45 76 93 74 81  0
#> 402 87 76 93 74 81  0
#> 406 87 28 88 59 33  0
#> 407 40 84 88 59 33  0
#> 408 87 84 88 59 33  0
#> 409 40 28 93 59 33  0
#> 410 87 28 93 59 33  0
#> 411 40 84 93 59 33  0
#> 412 87 84 93 59 33  0
#> 413 40 28 88 74 33  0
#> 414 87 28 88 74 33  0
#> 415 40 84 88 74 33  0
#> 416 87 84 88 74 33  0
#> 417 40 28 93 74 33  0
#> 418 87 28 93 74 33  0
#> 419 40 84 93 74 33  0
#> 421 40 28 88 59 81  0
#> 422 87 28 88 59 81  0
#> 423 40 84 88 59 81  0
#> 424 87 84 88 59 81  0
#> 425 40 28 93 59 81  0
#> 426 87 28 93 59 81  0
#> 427 40 84 93 59 81  0
#> 428 87 84 93 59 81  0
#> 429 40 28 88 74 81  0
#> 430 87 28 88 74 81  0
#> 431 40 84 88 74 81  0
#> 432 87 84 88 74 81  0
#> 433 40 28 93 74 81  0
#> 434 87 28 93 74 81  0
#> 438 87 34 47 18 61  0
#> 439 82 84 47 18 61  0
#> 440 87 84 47 18 61  0
#> 441 82 34 93 18 61  0
#> 442 87 34 93 18 61  0
#> 443 82 84 93 18 61  0
#> 444 87 84 93 18 61  0
#> 445 82 34 47 74 61  0
#> 446 87 34 47 74 61  0
#> 447 82 84 47 74 61  0
#> 448 87 84 47 74 61  0
#> 449 82 34 93 74 61  0
#> 450 87 34 93 74 61  0
#> 451 82 84 93 74 61  0
#> 452 87 84 93 74 61  0
#> 453 82 34 47 18 81  0
#> 454 87 34 47 18 81  0
#> 455 82 84 47 18 81  0
#> 456 87 84 47 18 81  0
#> 457 82 34 93 18 81  0
#> 458 87 34 93 18 81  0
#> 459 82 84 93 18 81  0
#> 460 87 84 93 18 81  0
#> 461 82 34 47 74 81  0
#> 462 87 34 47 74 81  0
#> 463 82 84 47 74 81  0
#> 464 87 84 47 74 81  0
#> 465 82 34 93 74 81  0
#> 466 87 34 93 74 81  0
#> 467 82 84 93 74 81  0
#> 470 87 72 34 38 47  0
#> 471 11 84 34 38 47  0
#> 472 87 84 34 38 47  0
#> 473 11 72 93 38 47  0
#> 474 87 72 93 38 47  0
#> 475 11 84 93 38 47  0
#> 476 87 84 93 38 47  0
#> 477 11 72 34 74 47  0
#> 478 87 72 34 74 47  0
#> 479 11 84 34 74 47  0
#> 480 87 84 34 74 47  0
#> 481 11 72 93 74 47  0
#> 482 87 72 93 74 47  0
#> 483 11 84 93 74 47  0
#> 484 87 84 93 74 47  0
#> 485 11 72 34 38 81  0
#> 486 87 72 34 38 81  0
#> 487 11 84 34 38 81  0
#> 488 87 84 34 38 81  0
#> 489 11 72 93 38 81  0
#> 490 87 72 93 38 81  0
#> 491 11 84 93 38 81  0
#> 492 87 84 93 38 81  0
#> 493 11 72 34 74 81  0
#> 494 87 72 34 74 81  0
#> 495 11 84 34 74 81  0
#> 496 87 84 34 74 81  0
#> 497 11 72 93 74 81  0
#> 498 87 72 93 74 81  0
#> 502 87 64 52 16 57  0
#> 503 19 84 52 16 57  0
#> 504 87 84 52 16 57  0
#> 505 19 64 93 16 57  0
#> 506 87 64 93 16 57  0
#> 507 19 84 93 16 57  0
#> 508 87 84 93 16 57  0
#> 509 19 64 52 74 57  0
#> 510 87 64 52 74 57  0
#> 511 19 84 52 74 57  0
#> 512 87 84 52 74 57  0
#> 513 19 64 93 74 57  0
#> 514 87 64 93 74 57  0
#> 515 19 84 93 74 57  0
#> 516 87 84 93 74 57  0
#> 517 19 64 52 16 81  0
#> 518 87 64 52 16 81  0
#> 519 19 84 52 16 81  0
#> 520 87 84 52 16 81  0
#> 521 19 64 93 16 81  0
#> 522 87 64 93 16 81  0
#> 523 19 84 93 16 81  0
#> 524 87 84 93 16 81  0
#> 525 19 64 52 74 81  0
#> 526 87 64 52 74 81  0
#> 527 19 84 52 74 81  0
#> 528 87 84 52 74 81  0
#> 529 19 64 93 74 81  0
#> 531 19 84 93 74 81  0
#> 534 87 63 53 56 62  0
#> 535  4 84 53 56 62  0
#> 536 87 84 53 56 62  0
#> 537  4 63 93 56 62  0
#> 538 87 63 93 56 62  0
#> 539  4 84 93 56 62  0
#> 540 87 84 93 56 62  0
#> 541  4 63 53 74 62  0
#> 542 87 63 53 74 62  0
#> 543  4 84 53 74 62  0
#> 544 87 84 53 74 62  0
#> 545  4 63 93 74 62  0
#> 546 87 63 93 74 62  0
#> 547  4 84 93 74 62  0
#> 549  4 63 53 56 81  0
#> 550 87 63 53 56 81  0
#> 551  4 84 53 56 81  0
#> 552 87 84 53 56 81  0
#> 553  4 63 93 56 81  0
#> 554 87 63 93 56 81  0
#> 555  4 84 93 56 81  0
#> 556 87 84 93 56 81  0
#> 557  4 63 53 74 81  0
#> 558 87 63 53 74 81  0
#> 559  4 84 53 74 81  0
#> 560 87 84 53 74 81  0
#> 561  4 63 93 74 81  0
#> 562 87 63 93 74 81  0
#> 563  4 84 93 74 81  0
#> 566 87 60 53 22 51  0
#> 567 11 84 53 22 51  0
#> 568 87 84 53 22 51  0
#> 569 11 60 93 22 51  0
#> 570 87 60 93 22 51  0
#> 571 11 84 93 22 51  0
#> 572 87 84 93 22 51  0
#> 573 11 60 53 74 51  0
#> 574 87 60 53 74 51  0
#> 575 11 84 53 74 51  0
#> 576 87 84 53 74 51  0
#> 577 11 60 93 74 51  0
#> 578 87 60 93 74 51  0
#> 579 11 84 93 74 51  0
#> 580 87 84 93 74 51  0
#> 581 11 60 53 22 81  0
#> 582 87 60 53 22 81  0
#> 583 11 84 53 22 81  0
#> 584 87 84 53 22 81  0
#> 585 11 60 93 22 81  0
#> 586 87 60 93 22 81  0
#> 587 11 84 93 22 81  0
#> 588 87 84 93 22 81  0
#> 589 11 60 53 74 81  0
#> 590 87 60 53 74 81  0
#> 591 11 84 53 74 81  0
#> 593 11 60 93 74 81  0
#> 598 87 54 24 60 32  0
#> 599 15 84 24 60 32  0
#> 600 87 84 24 60 32  0
#> 601 15 54 93 60 32  0
#> 602 87 54 93 60 32  0
#> 603 15 84 93 60 32  0
#> 604 87 84 93 60 32  0
#> 605 15 54 24 74 32  0
#> 606 87 54 24 74 32  0
#> 607 15 84 24 74 32  0
#> 608 87 84 24 74 32  0
#> 609 15 54 93 74 32  0
#> 610 87 54 93 74 32  0
#> 611 15 84 93 74 32  0
#> 612 87 84 93 74 32  0
#> 613 15 54 24 60 81  0
#> 614 87 54 24 60 81  0
#> 615 15 84 24 60 81  0
#> 616 87 84 24 60 81  0
#> 617 15 54 93 60 81  0
#> 618 87 54 93 60 81  0
#> 619 15 84 93 60 81  0
#> 620 87 84 93 60 81  0
#> 621 15 54 24 74 81  0
#> 622 87 54 24 74 81  0
#> 623 15 84 24 74 81  0
#> 625 15 54 93 74 81  0
#> 626 87 54 93 74 81  0
#> 627 15 84 93 74 81  0
#> 630 87 69 35 26 31  0
#> 631 21 84 35 26 31  0
#> 632 87 84 35 26 31  0
#> 633 21 69 93 26 31  0
#> 634 87 69 93 26 31  0
#> 635 21 84 93 26 31  0
#> 636 87 84 93 26 31  0
#> 637 21 69 35 74 31  0
#> 638 87 69 35 74 31  0
#> 639 21 84 35 74 31  0
#> 640 87 84 35 74 31  0
#> 641 21 69 93 74 31  0
#> 642 87 69 93 74 31  0
#> 643 21 84 93 74 31  0
#> 644 87 84 93 74 31  0
#> 645 21 69 35 26 81  0
#> 646 87 69 35 26 81  0
#> 647 21 84 35 26 81  0
#> 648 87 84 35 26 81  0
#> 649 21 69 93 26 81  0
#> 650 87 69 93 26 81  0
#> 651 21 84 93 26 81  0
#> 652 87 84 93 26 81  0
#> 653 21 69 35 74 81  0
#> 654 87 69 35 74 81  0
#> 655 21 84 35 74 81  0
#> 657 21 69 93 74 81  0
#> 658 87 69 93 74 81  0
#> 659 21 84 93 74 81  0