Classify a set of nondominated points

The classification is supported (true/false), extreme (true/false), supported non-extreme (true/false)

Usage

classifyNDSet(pts, direction = 1)

Arguments

pts

A set of non-dominated points. It is assumed that ncol(pts) equals the number of objectives ($p$).

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero (minimize objective $i$). If negative, consider the i'th column of the pts minus a value greater than on equal zero (maximize objective $i$).

Value

The ND set with classification columns.

Note

It is assumed that pts are nondominated.

Examples

# \donttest{
pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3, byrow = TRUE)
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-2,max(pts[,1])+2),
  ylim = c(min(pts[,2])-2,max(pts[,2])+2),
  zlim = c(min(pts[,3])-2,max(pts[,3])+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), useRGLBBox = TRUE)
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
plotCones3D(pts[,1:3], rectangle = TRUE, argsPolygon3d = list(alpha = 1))
finalize3D()

pts
#>     z1  z2   z3    se   sne    us cls
#> 1 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 2 0.00 1.0 0.00  TRUE FALSE FALSE  se
#> 3 1.00 0.0 0.00  TRUE FALSE FALSE  se
#> 4 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 5 0.25 0.5 0.25 FALSE  TRUE FALSE sne

pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.2,0.1,0.1, 0.1,0.45,0.45), ncol = 3, byrow = TRUE)
di <- -1 # maximize
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-1,max(pts[,1])+1),
  ylim = c(min(pts[,2])-1,max(pts[,2])+1),
  zlim = c(min(pts[,3])-1,max(pts[,3])+1)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), direction = di,
           addText = "coord")
pts <- classifyNDSet(pts[,1:3], direction = di)
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
plotCones3D(pts[,1:3], rectangle = TRUE, argsPolygon3d = list(alpha = 1), direction = di)
finalize3D()

pts
#>    z1   z2   z3    se   sne    us cls
#> 1 0.0 0.00 1.00  TRUE FALSE FALSE  se
#> 2 0.0 1.00 0.00  TRUE FALSE FALSE  se
#> 3 1.0 0.00 0.00  TRUE FALSE FALSE  se
#> 4 0.2 0.10 0.10 FALSE FALSE  TRUE  us
#> 5 0.1 0.45 0.45 FALSE  TRUE FALSE sne

pts <- matrix(c(0,0,1, 0,0,1, 0,1,0, 0.5,0.2,0.5, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3,
              byrow = TRUE)
classifyNDSet(pts)
#>        z1  z2   z3    se   sne    us cls
#> 1    0.00 0.0 1.00  TRUE FALSE FALSE  se
#> ...2 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 2    0.00 1.0 0.00  TRUE FALSE FALSE  se
#> 3    0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 4    1.00 0.0 0.00  TRUE FALSE FALSE  se
#> ...6 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 5    0.25 0.5 0.25 FALSE  TRUE FALSE sne

pts <- genNDSet(3,50)[,1:3]
ini3D(argsPlot3d = list(xlim = c(0,max(pts$z1)+2),
  ylim = c(0,max(pts$z2)+2),
  zlim = c(0,max(pts$z3)+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5))
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
finalize3D()

pts
#>    z1 z2 z3    se   sne    us cls
#> 1   2 58 53  TRUE FALSE FALSE  se
#> 2   4 59 35  TRUE FALSE FALSE  se
#> 3   5 36 63  TRUE FALSE FALSE  se
#> 4  15 24 73 FALSE FALSE  TRUE  us
#> 5  18 31 18  TRUE FALSE FALSE  se
#> 6  22 39 11  TRUE FALSE FALSE  se
#> 7  22 53 10  TRUE FALSE FALSE  se
#> 8  23 41 10  TRUE FALSE FALSE  se
#> 9  24 26 17  TRUE FALSE FALSE  se
#> 10 26  8 57  TRUE FALSE FALSE  se
#> 11 26 19 22  TRUE FALSE FALSE  se
#> 12 29 10 33  TRUE FALSE FALSE  se
#> 13 30 51  5  TRUE FALSE FALSE  se
#> 14 34  4 48  TRUE FALSE FALSE  se
#> 15 34 35  6  TRUE FALSE FALSE  se
#> 16 35  4 41  TRUE FALSE FALSE  se
#> 17 38  3 45  TRUE FALSE FALSE  se
#> 18 39  4 39  TRUE FALSE FALSE  se
#> 19 40 31  6  TRUE FALSE FALSE  se
#> 20 41  3 41  TRUE FALSE FALSE  se
#> 21 42  2 52  TRUE FALSE FALSE  se
#> 22 46 31  5  TRUE FALSE FALSE  se
#> 23 46 68  4 FALSE FALSE  TRUE  us
#> 24 47 30  6 FALSE FALSE  TRUE  us
#> 25 48 66  3  TRUE FALSE FALSE  se
#> 26 49  1 44  TRUE FALSE FALSE  se
#> 27 55 25  8  TRUE FALSE FALSE  se
#> 28 71 12 26  TRUE FALSE FALSE  se
#> 29 75 17 24 FALSE FALSE  TRUE  us
# }