Classify a set of nondominated points

The classification is supported (true/false), extreme (true/false), supported non-extreme (true/false)

Usage

classifyNDSet(pts, direction = 1)

Arguments

pts

A set of non-dominated points. It is assumed that ncol(pts) equals the number of objectives ($p$).

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero (minimize objective $i$). If negative, consider the i'th column of the pts minus a value greater than on equal zero (maximize objective $i$).

Value

The ND set with classification columns.

Note

It is assumed that pts are nondominated.

Examples

# \donttest{
pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3, byrow = TRUE)
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-2,max(pts[,1])+2),
  ylim = c(min(pts[,2])-2,max(pts[,2])+2),
  zlim = c(min(pts[,3])-2,max(pts[,3])+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), useRGLBBox = TRUE)
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
plotCones3D(pts[,1:3], rectangle = TRUE, argsPolygon3d = list(alpha = 1))
finalize3D()

pts
#>     z1  z2   z3    se   sne    us cls
#> 1 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 2 0.00 1.0 0.00  TRUE FALSE FALSE  se
#> 3 1.00 0.0 0.00  TRUE FALSE FALSE  se
#> 4 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 5 0.25 0.5 0.25 FALSE  TRUE FALSE sne

pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.2,0.1,0.1, 0.1,0.45,0.45), ncol = 3, byrow = TRUE)
di <- -1 # maximize
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-1,max(pts[,1])+1),
  ylim = c(min(pts[,2])-1,max(pts[,2])+1),
  zlim = c(min(pts[,3])-1,max(pts[,3])+1)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), direction = di,
           addText = "coord")
pts <- classifyNDSet(pts[,1:3], direction = di)
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
plotCones3D(pts[,1:3], rectangle = TRUE, argsPolygon3d = list(alpha = 1), direction = di)
finalize3D()

pts
#>    z1   z2   z3    se   sne    us cls
#> 1 0.0 0.00 1.00  TRUE FALSE FALSE  se
#> 2 0.0 1.00 0.00  TRUE FALSE FALSE  se
#> 3 1.0 0.00 0.00  TRUE FALSE FALSE  se
#> 4 0.2 0.10 0.10 FALSE FALSE  TRUE  us
#> 5 0.1 0.45 0.45 FALSE  TRUE FALSE sne

pts <- matrix(c(0,0,1, 0,0,1, 0,1,0, 0.5,0.2,0.5, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3,
              byrow = TRUE)
classifyNDSet(pts)
#>        z1  z2   z3    se   sne    us cls
#> 1    0.00 0.0 1.00  TRUE FALSE FALSE  se
#> ...2 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 2    0.00 1.0 0.00  TRUE FALSE FALSE  se
#> 3    0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 4    1.00 0.0 0.00  TRUE FALSE FALSE  se
#> ...6 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 5    0.25 0.5 0.25 FALSE  TRUE FALSE sne

pts <- genNDSet(3,50)[,1:3]
ini3D(argsPlot3d = list(xlim = c(0,max(pts$z1)+2),
  ylim = c(0,max(pts$z2)+2),
  zlim = c(0,max(pts$z3)+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5))
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
finalize3D()

pts
#>    z1 z2 z3    se   sne    us cls
#> 1   1 50 56  TRUE FALSE FALSE  se
#> 2   3 52 35  TRUE FALSE FALSE  se
#> 3   7 35 69 FALSE FALSE  TRUE  us
#> 4   8 32 34  TRUE FALSE FALSE  se
#> 5   8 63 29 FALSE FALSE  TRUE  us
#> 6  10 29 32  TRUE FALSE FALSE  se
#> 7  11 30 29  TRUE FALSE FALSE  se
#> 8  12 48 20  TRUE FALSE FALSE  se
#> 9  14 31 23  TRUE FALSE FALSE  se
#> 10 15 25 73 FALSE FALSE  TRUE  us
#> 11 18 32 18  TRUE FALSE FALSE  se
#> 12 21 11 49  TRUE FALSE FALSE  se
#> 13 22 25 19  TRUE FALSE FALSE  se
#> 14 22 54 10  TRUE FALSE FALSE  se
#> 15 27  7 52  TRUE FALSE FALSE  se
#> 16 33 26 11  TRUE FALSE FALSE  se
#> 17 34  6 37  TRUE FALSE FALSE  se
#> 18 36  3 51  TRUE FALSE FALSE  se
#> 19 43 67  4  TRUE FALSE FALSE  se
#> 20 44 23 10  TRUE FALSE FALSE  se
#> 21 45  6 30  TRUE FALSE FALSE  se
#> 22 48 18 14  TRUE FALSE FALSE  se
#> 23 48 63  3  TRUE FALSE FALSE  se
#> 24 51  8 25  TRUE FALSE FALSE  se
#> 25 54 20 12  TRUE FALSE FALSE  se
#> 26 54 62  3  TRUE FALSE FALSE  se
#> 27 70 15 22 FALSE FALSE  TRUE  us
#> 28 70 60  6 FALSE FALSE  TRUE  us
# }