Classify a set of nondominated points

The classification is supported (true/false), extreme (true/false), supported non-extreme (true/false)

classifyNDSet(pts, direction = 1)

Arguments

pts
A set of non-dominated points. It is assumed that ncol(pts) equals the number of objectives ($p$).

direction
Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero (minimize objective $i$). If negative, consider the i'th column of the pts minus a value greater than on equal zero (maximize objective $i$).

Value

The ND set with classification columns.

Note

It is assumed that pts are nondominated.

Examples

pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3, byrow = TRUE)
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-2,max(pts[,1])+2),
  ylim = c(min(pts[,2])-2,max(pts[,2])+2),
  zlim = c(min(pts[,3])-2,max(pts[,3])+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), useRGLBBox = TRUE)
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
plotCones3D(pts[,1:3], rectangle = TRUE, argsPolygon3d = list(alpha = 1))
finalize3D()
pts

#>     z1  z2   z3    se   sne    us cls
#> 1 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 2 0.00 1.0 0.00  TRUE FALSE FALSE  se
#> 3 1.00 0.0 0.00  TRUE FALSE FALSE  se
#> 4 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 5 0.25 0.5 0.25 FALSE  TRUE FALSE sne

pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.2,0.1,0.1, 0.1,0.45,0.45), ncol = 3, byrow = TRUE)
di <- -1 # maximize
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-1,max(pts[,1])+1),
  ylim = c(min(pts[,2])-1,max(pts[,2])+1),
  zlim = c(min(pts[,3])-1,max(pts[,3])+1)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), direction = di,
           addText = "coord")
pts <- classifyNDSet(pts[,1:3], direction = di)
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
plotCones3D(pts[,1:3], rectangle = TRUE, argsPolygon3d = list(alpha = 1), direction = di)
finalize3D()
pts

#>    z1   z2   z3    se   sne    us cls
#> 1 0.0 0.00 1.00  TRUE FALSE FALSE  se
#> 2 0.0 1.00 0.00  TRUE FALSE FALSE  se
#> 3 1.0 0.00 0.00  TRUE FALSE FALSE  se
#> 4 0.2 0.10 0.10 FALSE FALSE  TRUE  us
#> 5 0.1 0.45 0.45 FALSE  TRUE FALSE sne

pts <- matrix(c(0,0,1, 0,0,1, 0,1,0, 0.5,0.2,0.5, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3,
              byrow = TRUE)
classifyNDSet(pts)

#>     z1  z2   z3    se   sne    us cls
#> 1 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 2 0.00 0.0 1.00  TRUE FALSE FALSE  se
#> 3 0.00 1.0 0.00  TRUE FALSE FALSE  se
#> 4 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 5 1.00 0.0 0.00  TRUE FALSE FALSE  se
#> 6 0.50 0.2 0.50 FALSE FALSE  TRUE  us
#> 7 0.25 0.5 0.25 FALSE  TRUE FALSE sne

pts <- genNDSet(3,50)[,1:3]
ini3D(argsPlot3d = list(xlim = c(0,max(pts$z1)+2),
  ylim = c(0,max(pts$z2)+2),
  zlim = c(0,max(pts$z3)+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5))
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
finalize3D()
pts

#>    z1 z2 z3    se   sne    us cls
#> 1   1 52 44  TRUE FALSE FALSE  se
#> 2   4 33 47  TRUE FALSE FALSE  se
#> 3   5 56 32  TRUE FALSE FALSE  se
#> 4   6 33 39  TRUE FALSE FALSE  se
#> 5   7 35 32  TRUE FALSE FALSE  se
#> 6  10 39 25  TRUE FALSE FALSE  se
#> 7  11 45 21  TRUE FALSE FALSE  se
#> 8  18 14 44  TRUE FALSE FALSE  se
#> 9  18 22 27  TRUE FALSE FALSE  se
#> 10 18 39 15  TRUE FALSE FALSE  se
#> 11 20 12 49  TRUE FALSE FALSE  se
#> 12 20 19 28  TRUE FALSE FALSE  se
#> 13 21 13 38  TRUE FALSE FALSE  se
#> 14 21 49 11  TRUE FALSE FALSE  se
#> 15 26 23 17  TRUE FALSE FALSE  se
#> 16 26 46  8  TRUE FALSE FALSE  se
#> 17 27  7 47  TRUE FALSE FALSE  se
#> 18 27 13 28  TRUE FALSE FALSE  se
#> 19 30  6 41  TRUE FALSE FALSE  se
#> 20 30 41  7  TRUE FALSE FALSE  se
#> 21 33  4 50  TRUE FALSE FALSE  se
#> 22 33  8 33 FALSE FALSE  TRUE  us
#> 23 34 33  7  TRUE FALSE FALSE  se
#> 24 36  6 34  TRUE FALSE FALSE  se
#> 25 39 51  2  TRUE FALSE FALSE  se
#> 26 45 13 19 FALSE FALSE  TRUE  us
#> 27 45 14 17  TRUE FALSE FALSE  se
#> 28 53 12 20  TRUE FALSE FALSE  se
#> 29 55 29  6  TRUE FALSE FALSE  se
#> 30 67 39  5 FALSE FALSE  TRUE  us