Find extreme points of a nondominated set of points

Find extreme points of a nondominated set of points

Usage

classifyNDSetExtreme(pts, direction = 1)

Arguments

pts

A set of non-dominated points. It is assumed that ncol(pts) equals the number of objectives ($p$).

direction

Ray direction. If i'th entry is positive, consider the i'th column of the pts plus a value greater than on equal zero (minimize objective $i$). If negative, consider the i'th column of the pts minus a value greater than on equal zero (maximize objective $i$).

Value

The classification is extreme (se), supported non-extreme (sne) and unsupported us nondominated points. Return the ND set with classification columns se (true/false), sne (true/false), us (true/false) and cls (se, sne or us).

Note

It is assumed that pts are nondominated. This algorithm is faster than classifyNDSet(), since only check for extreme points.

See also

Examples

# \donttest{
pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3, byrow = TRUE)
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-2,max(pts[,1])+2),
  ylim = c(min(pts[,2])-2,max(pts[,2])+2),
  zlim = c(min(pts[,3])-2,max(pts[,3])+2)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), useRGLBBox = TRUE)
pts <- classifyNDSetExtreme(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))  # extreme
plotPoints3D(pts[is.na(pts$cls),1:3], argsPlot3d = list(col = "yellow"))  # unclassified
finalize3D()

3D plot

pts
#>     z1  z2   z3    se   sne    us  cls
#> 1 0.00 0.0 1.00  TRUE FALSE FALSE   se
#> 2 0.00 1.0 0.00  TRUE FALSE FALSE   se
#> 3 1.00 0.0 0.00  TRUE FALSE FALSE   se
#> 4 0.50 0.2 0.50 FALSE    NA    NA <NA>
#> 5 0.25 0.5 0.25 FALSE    NA    NA <NA>

pts <- matrix(c(0,0,1, 0,1,0, 1,0,0, 0.2,0.1,0.1, 0.1,0.45,0.45), ncol = 3, byrow = TRUE)
di <- -1 # maximize
ini3D(argsPlot3d = list(xlim = c(min(pts[,1])-1,max(pts[,1])+1),
  ylim = c(min(pts[,2])-1,max(pts[,2])+1),
  zlim = c(min(pts[,3])-1,max(pts[,3])+1)))
plotHull3D(pts, addRays = TRUE, argsPolygon3d = list(alpha = 0.5), direction = di,
           addText = "coord")
pts <- classifyNDSetExtreme(pts[,1:3], direction = di)
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[is.na(pts$cls),1:3], argsPlot3d = list(col = "yellow"))  # unclassified
finalize3D()

3D plot

pts
#>    z1   z2   z3    se   sne    us  cls
#> 1 0.0 0.00 1.00  TRUE FALSE FALSE   se
#> 2 0.0 1.00 0.00  TRUE FALSE FALSE   se
#> 3 1.0 0.00 0.00  TRUE FALSE FALSE   se
#> 4 0.2 0.10 0.10 FALSE    NA    NA <NA>
#> 5 0.1 0.45 0.45 FALSE    NA    NA <NA>

pts <- matrix(c(0,0,1, 0,0,1, 0,1,0, 0.5,0.2,0.5, 1,0,0, 0.5,0.2,0.5, 0.25,0.5,0.25), ncol = 3,
              byrow = TRUE)
classifyNDSetExtreme(pts)
#>     z1  z2   z3    se   sne    us  cls
#> 1 0.00 0.0 1.00  TRUE FALSE FALSE   se
#> 2 0.00 1.0 0.00  TRUE FALSE FALSE   se
#> 3 0.50 0.2 0.50 FALSE    NA    NA <NA>
#> 4 1.00 0.0 0.00  TRUE FALSE FALSE   se
#> 5 0.25 0.5 0.25 FALSE    NA    NA <NA>
#> 6 6.00 0.0 1.00 FALSE    NA    NA <NA>
#> 7 6.00 1.0 0.00 FALSE    NA    NA <NA>

pts <- genNDSet(3,15)[,1:3]
ini3D(argsPlot3d = list(xlim = c(0,max(pts$z1)+2),
  ylim = c(0,max(pts$z2)+2),
  zlim = c(0,max(pts$z3)+2)))
plotHull3D(pts[, 1:3], addRays = TRUE, argsPolygon3d = list(alpha = 0.5))
pts <- classifyNDSetExtreme(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[is.na(pts$cls),1:3], argsPlot3d = list(col = "yellow"))  # unclassified
finalize3D()

3D plot

pts
#>    z1 z2 z3    se   sne    us  cls
#> 1  12 58 21  TRUE FALSE FALSE   se
#> 2  12 28 73 FALSE    NA    NA <NA>
#> 3  19 35 16  TRUE FALSE FALSE   se
#> 4  28  7 59  TRUE FALSE FALSE   se
#> 5  20 11 47  TRUE FALSE FALSE   se
#> 6   5 33 44  TRUE FALSE FALSE   se
#> 7  80 48 11  TRUE FALSE FALSE   se
#> 8  72 19 19  TRUE FALSE FALSE   se
#> 9  53  3 36  TRUE FALSE FALSE   se
#> 10 46  1 52  TRUE FALSE FALSE   se
#> 11 51  2 40  TRUE FALSE FALSE   se
#> 12 10 74 34 FALSE    NA    NA <NA>
#> 13 72 29 12  TRUE FALSE FALSE   se
#> 14 26  8 58 FALSE    NA    NA <NA>
#> 15 68  9 30 FALSE    NA    NA <NA>

pts <- genNDSet(3, 15, keepDom = FALSE, argsSphere = list(below = FALSE, factor = 10))[,1:3]
ini3D(argsPlot3d = list(xlim = c(0,max(pts$z1)+2),
  ylim = c(0,max(pts$z2)+2),
  zlim = c(0,max(pts$z3)+2)))
plotHull3D(pts[, 1:3], addRays = TRUE, argsPolygon3d = list(alpha = 0.5))
pts <- classifyNDSetExtreme(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[is.na(pts$cls),1:3], argsPlot3d = list(col = "yellow"))  # unclassified
finalize3D()

3D plot

pts
#>    z1 z2 z3    se   sne    us  cls
#> 1  78 70 14  TRUE FALSE FALSE   se
#> 2  90 23 40  TRUE FALSE FALSE   se
#> 3  61 80 12  TRUE FALSE FALSE   se
#> 4  62 17 85 FALSE    NA    NA <NA>
#> 5  90 56 21  TRUE FALSE FALSE   se
#> 6  93 55 26 FALSE    NA    NA <NA>
#> 7  64 79 12  TRUE FALSE FALSE   se
#> 8  34 92 30  TRUE FALSE FALSE   se
#> 9  28 39 93  TRUE FALSE FALSE   se
#> 10 15 65 82 FALSE    NA    NA <NA>
#> 11 76 11 67  TRUE FALSE FALSE   se
#> 12 11 69 74  TRUE FALSE FALSE   se
#> 13 41 24 91  TRUE FALSE FALSE   se
#> 14 56 88 18 FALSE    NA    NA <NA>
#> 15 22 90 43  TRUE FALSE FALSE   se
# }